Calculate combined probabilities of similarity or intersection metrics of tracks

combined_prob() calculates the combined probabilities of similarity and intersection metrics derived from different models. The function uses simulation data to extract p-values, providing insight into the significance of combined metrics across various similarity assessments.

Usage

combined_prob(data, metrics = NULL)

Arguments

data

A track R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

metrics

A list of track similarity and/or track intersection R objects derived from different tests. All tests must be based on the same number of simulations.

Value

A list containing:

P_values (model names)

A matrix of p-values for the combined metrics across all trajectories. Each entry represents the probability of observing the combined metrics between the corresponding pair of trajectories.

P_values_combined (model names)

A numeric value representing the overall probability of observing the combined metrics, across all pairs of trajectories.

Details

The combined_prob() function combines p-values derived from multiple similarity metric tests and intersection tests. It calculates the combined p-values by assessing the probability of observing the combined metrics across simulated datasets. This function is particularly useful for comparing multiple models and evaluating their collective performance in terms of p-values.

Logo

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: "Directed" model and similarity metrics.
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
DTW1 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-13 21:08:28.985399 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 32.99155
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.98716 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 25.31786
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.988801 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 19.69963
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.990442 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 21.08054
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.992088 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 18.59471
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.993677 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 15.31543
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.995325 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 25.80199
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.996929 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 15.89334
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:28.998547 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 18.18086
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:29.000218 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 16.66715
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.7
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.7
#> 
Frechet1 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-13 21:08:29.370807 Iteration 1
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.8675854
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:29.55603 Iteration 2
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.8080506
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:29.736302 Iteration 3
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7740225
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:29.908892 Iteration 4
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7738793
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:30.081001 Iteration 5
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7731691
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:30.268328 Iteration 6
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.8211556
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:30.444504 Iteration 7
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.9410678
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:30.620863 Iteration 8
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7772298
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:30.797079 Iteration 9
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821592
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:31.162616 Iteration 10
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.9474063
#> Track_2      NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Frechet_distance_metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821257
#> Track_2      NA        NA
#> 
#> $Frechet_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.4
#> Track_2      NA      NA
#> 
#> $Frechet_metric_p_values_combined
#> [1] 0.4
#> 
int1 <- track_intersection(PaluxyRiver, test = TRUE, H1 = "Lower", sim = s1,
  origin.permutation = "None")
#> 2025-05-13 21:08:31.186136 Iteration 1
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.197029 Iteration 2
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       3
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.207947 Iteration 3
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       2
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.218803 Iteration 4
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.229941 Iteration 5
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.240905 Iteration 6
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.251953 Iteration 7
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       2
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.263207 Iteration 8
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       4
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.274646 Iteration 9
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.290914 Iteration 10
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       4
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Intersection_metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> 
combined_prob(PaluxyRiver, metrics = list(DTW1, Frechet1, int1))
#> $`P_values (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#>         Track_1 Track_2
#> Track_1      NA     0.2
#> Track_2      NA      NA
#> 
#> $`P_values_combined (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#> [1] 0.2
#> 

# Example 2: "Constrained" model and similarity metrics.
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
DTW2 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2,
  superposition = "None")
#> 2025-05-13 21:08:31.324329 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 150.7881
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.325961 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 149.7137
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.327581 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 27.53148
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.329332 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 69.17355
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.331626 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 190.9703
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.333704 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 64.56412
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.335399 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 108.2882
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.337066 Iteration 8
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 90.9885
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:31.338672 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 39.00249
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.34031 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 144.1645
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
Frechet2 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s2,
  superposition = "None")
#> 2025-05-13 21:08:31.665532 Iteration 1
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 7.213888
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:31.813751 Iteration 2
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 8.390197
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:32.007175 Iteration 3
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 1.242323
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:32.176649 Iteration 4
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 3.659994
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:32.308454 Iteration 5
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 11.22782
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:32.490846 Iteration 6
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 2.725772
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:32.658037 Iteration 7
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 6.028844
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:32.822852 Iteration 8
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 4.558586
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:33.003608 Iteration 9
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 2.132696
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:33.149601 Iteration 10
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA 6.99222
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Frechet_distance_metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821257
#> Track_2      NA        NA
#> 
#> $Frechet_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $Frechet_metric_p_values_combined
#> [1] 0
#> 
int2 <- track_intersection(PaluxyRiver, test = TRUE, H1 = "Lower", sim = s2,
  origin.permutation = "Min.Box")
#> 2025-05-13 21:08:33.534974 Permutation 1
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:34.082072 Permutation 2
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:34.197201 Permutation 3
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:34.684203 Permutation 4
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:35.016777 Permutation 5
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:35.130222 Permutation 6
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:35.288979 Permutation 7
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:35.381777 Permutation 8
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:35.759566 Permutation 9
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:08:35.937176 Permutation 10
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> PERMUTATION COMPLETED
#> ------------------------------------
#>  
#> 2025-05-13 21:08:35.948311 Iteration 1
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:35.959775 Iteration 2
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:35.970602 Iteration 3
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       2
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:35.981508 Iteration 4
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:35.992503 Iteration 5
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.003604 Iteration 6
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.014324 Iteration 7
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.025273 Iteration 8
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.03595 Iteration 9
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.046661 Iteration 10
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Intersection_metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> 
combined_prob(PaluxyRiver, metrics = list(DTW2, Frechet2, int2))
#> $`P_values (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $`P_values_combined (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#> [1] 0
#> 

# Example 3: "Unconstrained" model and similarity metrics.
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
DTW3 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3,
  superposition = "None")
#> 2025-05-13 21:08:36.081931 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 590.3857
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.083759 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 481.8072
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.085505 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 798.1471
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.087194 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 401.0092
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.088871 Iteration 5
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 58.7754
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.090537 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 624.5319
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.092308 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 741.7511
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.094072 Iteration 8
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 775.137
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.095785 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 428.8708
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.097502 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 148.9373
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
Frechet3 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s3,
  superposition = "None")
#> 2025-05-13 21:08:36.320842 Iteration 1
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 24.83123
#> Track_2      NA       NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:08:36.353014 Iteration 2
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:08:36.382874 Iteration 3
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:08:36.419648 Iteration 4
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.593784 Iteration 5
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 3.069257
#> Track_2      NA       NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:08:36.623833 Iteration 6
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:08:36.658314 Iteration 7
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:08:36.687716 Iteration 8
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:36.763319 Iteration 9
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 16.14453
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:36.944252 Iteration 10
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 4.459178
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Frechet_distance_metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821257
#> Track_2      NA        NA
#> 
#> $Frechet_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.6
#> Track_2      NA      NA
#> 
#> $Frechet_metric_p_values_combined
#> [1] 0.6
#> 
int3 <- track_intersection(PaluxyRiver, test = TRUE, H1 = "Lower", sim = s3,
  origin.permutation = "Conv.Hull")
#> 2025-05-13 21:08:37.296089 Permutation 1
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:37.456706 Permutation 2
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:37.859434 Permutation 3
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:38.220125 Permutation 4
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:38.387578 Permutation 5
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:38.587353 Permutation 6
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:38.707876 Permutation 7
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:39.03793 Permutation 8
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:39.171832 Permutation 9
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:08:39.314862 Permutation 10
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> PERMUTATION COMPLETED
#> ------------------------------------
#>  
#> 2025-05-13 21:08:39.325681 Iteration 1
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.336468 Iteration 2
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.34761 Iteration 3
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.358957 Iteration 4
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.3698 Iteration 5
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.380409 Iteration 6
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.390934 Iteration 7
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.401539 Iteration 8
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.41246 Iteration 9
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:39.422984 Iteration 10
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Intersection_metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> 
combined_prob(PaluxyRiver, metrics = list(DTW3, Frechet3, int3))
#> $`P_values (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $`P_values_combined (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#> [1] 0
#>