Calculate combined probabilities of similarity or intersection metrics of tracks

combined_prob() calculates the combined probabilities of similarity and intersection metrics derived from different models. The function uses simulation data to extract p-values, providing insight into the significance of combined metrics across various similarity assessments.

Usage

combined_prob(data, metrics = NULL)

Arguments

data

A track R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

metrics

A list of track similarity and/or track intersection R objects derived from different tests. All tests must be based on the same number of simulations.

Value

A list containing:

P_values (model names)

A matrix of p-values for the combined metrics across all trajectories. Each entry represents the probability of observing the combined metrics between the corresponding pair of trajectories.

P_values_combined (model names)

A numeric value representing the overall probability of observing the combined metrics, across all pairs of trajectories.

Details

The combined_prob() function combines p-values derived from multiple similarity metric tests and intersection tests. It calculates the combined p-values by assessing the probability of observing the combined metrics across simulated datasets. This function is particularly useful for comparing multiple models and evaluating their collective performance in terms of p-values.

Logo

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: "Directed" model and similarity metrics.
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
DTW1 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-13 21:00:57.80172 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 32.99155
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.803426 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 25.31786
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.805024 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 19.69963
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.80663 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 21.08054
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.808231 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 18.59471
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.809808 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 15.31543
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.811401 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 25.80199
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.812989 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 15.89334
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.814583 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 18.18086
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:00:57.816163 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 16.66715
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.7
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.7
#> 
Frechet1 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-13 21:00:58.173688 Iteration 1
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.8675854
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:58.349506 Iteration 2
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.8080506
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:58.51516 Iteration 3
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7740225
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:58.68437 Iteration 4
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7738793
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:58.850097 Iteration 5
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7731691
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:59.027744 Iteration 6
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.8211556
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:59.195788 Iteration 7
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.9410678
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:59.36187 Iteration 8
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7772298
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:59.527449 Iteration 9
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821592
#> Track_2      NA        NA
#> ------------------------------------
#> 2025-05-13 21:00:59.885608 Iteration 10
#>  
#> Frechet metric
#>         Track_1   Track_2
#> Track_1      NA 0.9474063
#> Track_2      NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Frechet_distance_metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821257
#> Track_2      NA        NA
#> 
#> $Frechet_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.4
#> Track_2      NA      NA
#> 
#> $Frechet_metric_p_values_combined
#> [1] 0.4
#> 
int1 <- track_intersection(PaluxyRiver, test = TRUE, H1 = "Lower", sim = s1,
  origin.permutation = "None")
#> 2025-05-13 21:00:59.908376 Iteration 1
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.922648 Iteration 2
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       3
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.932876 Iteration 3
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       2
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.943104 Iteration 4
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.95335 Iteration 5
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.963548 Iteration 6
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.973702 Iteration 7
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       2
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.98397 Iteration 8
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       4
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:00:59.994259 Iteration 9
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:00.00456 Iteration 10
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       4
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Intersection_metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> 
combined_prob(PaluxyRiver, metrics = list(DTW1, Frechet1, int1))
#> $`P_values (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#>         Track_1 Track_2
#> Track_1      NA     0.2
#> Track_2      NA      NA
#> 
#> $`P_values_combined (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#> [1] 0.2
#> 

# Example 2: "Constrained" model and similarity metrics.
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
DTW2 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2,
  superposition = "None")
#> 2025-05-13 21:01:00.03867 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 150.7881
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.04036 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 149.7137
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.042001 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 27.53148
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.043627 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 69.17355
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.045261 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 190.9703
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.046867 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 64.56412
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.048498 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 108.2882
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.050128 Iteration 8
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 90.9885
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:00.051725 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 39.00249
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.053345 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 144.1645
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
Frechet2 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s2,
  superposition = "None")
#> 2025-05-13 21:01:00.371661 Iteration 1
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 7.213888
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.512194 Iteration 2
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 8.390197
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.692745 Iteration 3
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 1.242323
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.856074 Iteration 4
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 3.659994
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:00.982497 Iteration 5
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 11.22782
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:01.158772 Iteration 6
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 2.725772
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:01.318223 Iteration 7
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 6.028844
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:01.476786 Iteration 8
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 4.558586
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:01.653317 Iteration 9
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 2.132696
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:01.795139 Iteration 10
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA 6.99222
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Frechet_distance_metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821257
#> Track_2      NA        NA
#> 
#> $Frechet_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $Frechet_metric_p_values_combined
#> [1] 0
#> 
int2 <- track_intersection(PaluxyRiver, test = TRUE, H1 = "Lower", sim = s2,
  origin.permutation = "Min.Box")
#> 2025-05-13 21:01:02.167505 Permutation 1
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:02.695302 Permutation 2
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:02.817656 Permutation 3
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:03.277195 Permutation 4
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:03.605901 Permutation 5
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:03.718004 Permutation 6
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:03.872266 Permutation 7
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:03.960792 Permutation 8
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:04.316201 Permutation 9
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> 2025-05-13 21:01:04.489459 Permutation 10
#>  
#> Permutation of coordinates at origin using Min.Box
#> ------------------------------------
#> PERMUTATION COMPLETED
#> ------------------------------------
#>  
#> 2025-05-13 21:01:04.499703 Iteration 1
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.510176 Iteration 2
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.520494 Iteration 3
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       2
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.530813 Iteration 4
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.541258 Iteration 5
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.551676 Iteration 6
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.562076 Iteration 7
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.572342 Iteration 8
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.582526 Iteration 9
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.59282 Iteration 10
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Intersection_metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> 
combined_prob(PaluxyRiver, metrics = list(DTW2, Frechet2, int2))
#> $`P_values (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $`P_values_combined (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#> [1] 0
#> 

# Example 3: "Unconstrained" model and similarity metrics.
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
DTW3 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3,
  superposition = "None")
#> 2025-05-13 21:01:04.627159 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 590.3857
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.628816 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 481.8072
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.63046 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 798.1471
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.632091 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 401.0092
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.633708 Iteration 5
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 58.7754
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.635353 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 624.5319
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.636974 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 741.7511
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.638599 Iteration 8
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 775.137
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:04.640233 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 428.8708
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:04.641866 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 148.9373
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
Frechet3 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s3,
  superposition = "None")
#> 2025-05-13 21:01:04.851167 Iteration 1
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 24.83123
#> Track_2      NA       NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:01:04.881726 Iteration 2
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:01:04.910657 Iteration 3
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:01:04.947125 Iteration 4
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:05.111671 Iteration 5
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 3.069257
#> Track_2      NA       NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:01:05.140939 Iteration 6
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:01:05.169792 Iteration 7
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> Warning: The Frechet distance was unable to be found
#> 2025-05-13 21:01:05.201827 Iteration 8
#>  
#> Frechet metric
#>         Track_1 Track_2
#> Track_1      NA      -1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:05.27146 Iteration 9
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 16.14453
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:01:05.443026 Iteration 10
#>  
#> Frechet metric
#>         Track_1  Track_2
#> Track_1      NA 4.459178
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Frechet_distance_metric
#>         Track_1   Track_2
#> Track_1      NA 0.7821257
#> Track_2      NA        NA
#> 
#> $Frechet_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.6
#> Track_2      NA      NA
#> 
#> $Frechet_metric_p_values_combined
#> [1] 0.6
#> 
int3 <- track_intersection(PaluxyRiver, test = TRUE, H1 = "Lower", sim = s3,
  origin.permutation = "Conv.Hull")
#> 2025-05-13 21:01:05.771465 Permutation 1
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:05.926077 Permutation 2
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:06.119426 Permutation 3
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:06.271799 Permutation 4
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:06.606777 Permutation 5
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:06.986808 Permutation 6
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:07.101679 Permutation 7
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:07.417281 Permutation 8
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:07.5525 Permutation 9
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> 2025-05-13 21:01:07.686704 Permutation 10
#>  
#> Permutation of coordinates at origin using Conv.Hull
#> ------------------------------------
#> PERMUTATION COMPLETED
#> ------------------------------------
#>  
#> 2025-05-13 21:01:07.696863 Iteration 1
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.707245 Iteration 2
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.717405 Iteration 3
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.727643 Iteration 4
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.737875 Iteration 5
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.748188 Iteration 6
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.758471 Iteration 7
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.768599 Iteration 8
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.778996 Iteration 9
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:01:07.789371 Iteration 10
#>  
#> Intersect metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $Intersection_metric
#>         Track_1 Track_2
#> Track_1      NA       1
#> Track_2      NA      NA
#> 
combined_prob(PaluxyRiver, metrics = list(DTW3, Frechet3, int3))
#> $`P_values (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $`P_values_combined (DTW_distance_metric, Frechet_distance_metric, Intersection_metric)`
#> [1] 0
#>