Similarity metric using Dynamic Time Warping (DTW)

simil_DTW_metric() computes similarity metrics between two or more trajectories using Dynamic Time Warping (DTW). It allows for different superposition methods to align trajectories before calculating the DTW metric. The function also supports testing with simulations to calculate p-values for the DTW distance metrics.

Usage

simil_DTW_metric(data, test = NULL, sim = NULL, superposition = NULL)

Arguments

data

A track R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

test

Logical; if TRUE, the function compares the observed DTW distances against simulated trajectories and calculates p-values. Default is FALSE.

sim

A track simulation R object consisting of a list of simulated trajectories to use for comparison when test = TRUE.

superposition

A character string indicating the method used to align trajectories. Options are "None", "Centroid", or "Origin". Default is "None".

Value

A track similarity R object consisting ofa list containing the following elements:

DTW_distance_metric

A matrix containing the pairwise DTW distances between trajectories.

DTW_distance_metric_p_values

(If test is TRUE) A matrix containing the p-values for the pairwise DTW distances.

DTW_metric_p_values_combined

(If test is TRUE) The overall p-value for the combined DTW distances.

DTW_distance_metric_simulations

(If test is TRUE) A list of DTW distance matrices from each simulated dataset.

Details

The simil_DTW_metric() function calculates the similarity between trajectories using the Dynamic Time Warping (DTW) algorithm from the dtw package. The dtw() function is used with the dist.method argument set to "Euclidean" for computing the local distances between points in the trajectories.

DTW aligns two time series by minimizing the cumulative distance between their points, creating an optimal alignment despite variations in length or temporal distortions. The algorithm constructs a distance matrix where each element represents the cost of aligning points between the two series and finds a warping path through this matrix that minimizes the total distance. The warping path is contiguous and monotonic, starting from the bottom-left corner and ending at the top-right corner (Cleasby et al., 2019).

DTW measures are non-negative and unbounded, with larger values indicating greater dissimilarity between the time series. This method has been used in various contexts, including ecological studies to analyze and cluster trajectory data (Cleasby et al., 2019).

Potential limitations and biases of DTW include sensitivity to noise and outliers, computational complexity, and the need for appropriate distance metrics. Additionally, DTW may not always account for all structural differences between trajectories and can be biased by the chosen alignment constraints. While DTW can handle trajectories of different lengths due to its elastic nature, having trajectories of similar lengths can improve the accuracy and interpretability of the similarity measure. Similar lengths result in a more meaningful alignment and can make the computation more efficient. When trajectories differ significantly in length, preprocessing or normalization might be necessary, and careful analysis is required to understand the alignment path. The function’s flexibility in handling different lengths allows it to be applied in various contexts. However, large differences in trajectory lengths might introduce potential biases that should be considered when interpreting the results.

The function offers three different superposition methods to align the trajectories before DTW() is applied:

"None": No superposition is applied.

"Centroid": Trajectories are shifted to align based on their centroids.

"Origin": Trajectories are shifted to align based on their starting point.

If test = TRUE, the function can compute p-values by comparing the observed DTW distances with those generated from a set of simulated trajectories. The p-values are calculated for both individual trajectory pairs and for the entire set of trajectories.

Logo

References

Cleasby, I. R., Wakefield, E. D., Morrissey, B. J., Bodey, T. W., Votier, S. C., Bearhop, S., & Hamer, K. C. (2019). Using time-series similarity measures to compare animal movement trajectories in ecology. Behavioral Ecology and Sociobiology, 73, 1-19.

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Simulating tracks using the "Directed" model and comparing DTW distance
# in the PaluxyRiver dataset
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-14 06:09:04.163677 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 31.77845
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.165532 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 30.54245
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.167291 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 20.30751
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.169023 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 24.93207
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.170729 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 17.38193
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.172406 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 22.66588
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.174109 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 33.02965
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.175802 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 14.41302
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.177448 Iteration 9
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 20.1916
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-14 06:09:04.179187 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 31.72831
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.5
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.5
#> 

# Example 2: Simulating tracks using the "Constrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
#> 2025-05-14 06:09:04.226248 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 134.0087
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.22815 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 77.49626
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.229931 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 340.5661
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.231692 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 163.8732
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.233362 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 43.73261
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.235128 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 179.4593
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.236875 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 160.7052
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.238545 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 115.4045
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.240239 Iteration 9
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 32.1273
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-14 06:09:04.241966 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 70.31156
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 3: Simulating tracks using the "Unconstrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
#> 2025-05-14 06:09:04.274892 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 617.6054
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.276725 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 298.1122
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.278454 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 533.2894
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.280178 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 704.5754
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.281861 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 527.8494
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.283523 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 786.4886
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.285278 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 679.9316
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.286991 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 235.2818
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.288736 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 674.6949
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.290427 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 837.7564
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 4: Simulating and comparing DTW distance in the MountTom dataset using the
# "Centroid" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
#> 2025-05-14 06:09:04.553872 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 39.85817 40.35185  8.899142 40.52671 38.934309 22.77367
#> Track_02       NA       NA 32.07870 37.083893 35.27363 22.320176 41.04979
#> Track_03       NA       NA       NA 43.866541  3.98054  9.791687 23.67253
#> Track_04       NA       NA       NA        NA 44.66590 40.743732 29.26088
#> Track_07       NA       NA       NA        NA       NA 12.694808 22.18156
#> Track_08       NA       NA       NA        NA       NA        NA 26.81167
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 19.051397 38.07647 26.514386 32.414854
#> Track_02 39.000317 39.79095 25.865614 10.207865
#> Track_03 24.875734 13.50160  7.438192 15.895046
#> Track_04 25.363759 43.47309 29.771278 31.932952
#> Track_07 23.971035 10.95658  6.823488 18.581557
#> Track_08 27.471587 20.67593 10.614219  9.415399
#> Track_09  5.817647 16.22618 12.499916 26.311890
#> Track_13        NA 18.94607 14.509294 25.958271
#> Track_15        NA       NA  8.436358 24.079640
#> Track_16        NA       NA        NA 12.366009
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:04.602498 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 44.06543 38.59194 27.78086 22.165417 23.12129 23.841954
#> Track_02       NA       NA 13.73134 57.61047 35.345589 11.93815 41.842039
#> Track_03       NA       NA       NA 41.47215 22.081036 12.20595 27.632480
#> Track_04       NA       NA       NA       NA  7.851895 35.23261  6.616643
#> Track_07       NA       NA       NA       NA        NA 18.64645  4.265497
#> Track_08       NA       NA       NA       NA        NA       NA 23.659744
#> Track_09       NA       NA       NA       NA        NA       NA        NA
#> Track_13       NA       NA       NA       NA        NA       NA        NA
#> Track_15       NA       NA       NA       NA        NA       NA        NA
#> Track_16       NA       NA       NA       NA        NA       NA        NA
#> Track_18       NA       NA       NA       NA        NA       NA        NA
#>           Track_13  Track_15  Track_16 Track_18
#> Track_01 14.205906 17.141786 21.514109 36.90862
#> Track_02 38.088489 19.773341 11.618505 34.18179
#> Track_03 27.593236 17.810516  8.588206 17.74593
#> Track_04 11.389691 32.260749 27.150419 28.29245
#> Track_07  9.790185 18.186900 12.441784 12.59135
#> Track_08 20.139295  6.530998  4.994458 20.43399
#> Track_09  9.385689 22.427470 17.005756 17.47615
#> Track_13        NA 17.865722 15.934743 20.89706
#> Track_15        NA        NA  8.572651 22.87181
#> Track_16        NA        NA        NA 12.78381
#> Track_18        NA        NA        NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.651353 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 58.83738 33.99368 34.58840 40.48177 19.739178 38.501363
#> Track_02       NA       NA 24.07640 53.27569 10.33400 32.817555 30.122407
#> Track_03       NA       NA       NA 19.35985 17.92618 10.737915  3.889582
#> Track_04       NA       NA       NA       NA 42.06467  9.283309 19.337475
#> Track_07       NA       NA       NA       NA       NA 21.650826 24.329864
#> Track_08       NA       NA       NA       NA       NA        NA 13.248702
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16 Track_18
#> Track_01 45.783742 39.366419 20.964158 22.38566
#> Track_02 15.352616 15.002930 21.974116 29.54424
#> Track_03 23.938662 19.453473  6.662378 20.84593
#> Track_04 48.412173 42.234162 15.805243 37.79614
#> Track_07  6.576818  3.706977 12.899483 15.50086
#> Track_08 26.999399 21.746160  5.688009 17.58923
#> Track_09 30.817287 25.829737  8.596377 26.80650
#> Track_13        NA  5.403799 17.592377 17.36645
#> Track_15        NA        NA 13.279995 13.61773
#> Track_16        NA        NA        NA 12.02297
#> Track_18        NA        NA        NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.709101 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 38.69518 39.93715 47.69796 35.03534 24.772334 50.66342
#> Track_02       NA       NA 18.90405 59.82315 11.65038  7.630597 41.01641
#> Track_03       NA       NA       NA 39.93462  6.14401 11.866535 19.13493
#> Track_04       NA       NA       NA       NA 41.13863 39.952885 29.30068
#> Track_07       NA       NA       NA       NA       NA  7.420919 23.65397
#> Track_08       NA       NA       NA       NA       NA        NA 26.45277
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 47.657522  9.184192 28.584393 25.827911
#> Track_02 36.553834 27.946368 23.750201  6.284334
#> Track_03 15.321548 22.296765 10.314554 12.701369
#> Track_04 31.742383 28.800349 19.931393 41.468568
#> Track_07 19.996313 20.695435 11.780653  7.836044
#> Track_08 23.414651 15.167365 11.984405  1.878083
#> Track_09  6.432533 28.949733  8.041936 27.904173
#> Track_13        NA 27.452116  8.708225 24.592208
#> Track_15        NA        NA 13.604797 16.358711
#> Track_16        NA        NA        NA 12.965901
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:04.759496 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 11.42897  7.766967 66.25953 39.13201 36.95542 44.151032
#> Track_02       NA       NA 11.026109 62.99472 33.39920 34.46015 37.365999
#> Track_03       NA       NA        NA 41.35638 20.51154 18.09271 25.090332
#> Track_04       NA       NA        NA       NA 27.21903 11.54199 33.624752
#> Track_07       NA       NA        NA       NA       NA 15.51530  4.339441
#> Track_08       NA       NA        NA       NA       NA       NA 20.803638
#> Track_09       NA       NA        NA       NA       NA       NA        NA
#> Track_13       NA       NA        NA       NA       NA       NA        NA
#> Track_15       NA       NA        NA       NA       NA       NA        NA
#> Track_16       NA       NA        NA       NA       NA       NA        NA
#> Track_18       NA       NA        NA       NA       NA       NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 35.84824 13.733382 23.171325 29.371860
#> Track_02 27.76790 17.068785 22.071798 28.679014
#> Track_03 21.20645  5.267348  9.796302 13.505673
#> Track_04 46.47374 40.626154 20.512722 24.134202
#> Track_07 14.00083 22.509074 11.335516 19.241065
#> Track_08 25.70951 17.069837  6.248246  8.129488
#> Track_09 13.98567 27.065105 15.834259 24.616000
#> Track_13       NA 25.361756 17.308676 26.555083
#> Track_15       NA        NA  9.703801 11.489774
#> Track_16       NA        NA        NA  5.635970
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:04.807623 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 52.98635 39.76537 28.65333 41.68817 15.156634 42.051883
#> Track_02       NA       NA 29.13428 38.17780 22.52273 24.110655 41.978682
#> Track_03       NA       NA       NA 42.73168  7.51475 20.527450  9.414592
#> Track_04       NA       NA       NA       NA 41.90870  8.561833 51.494159
#> Track_07       NA       NA       NA       NA       NA 20.262872 18.002673
#> Track_08       NA       NA       NA       NA       NA        NA 26.501998
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 39.328306 42.061666 29.061127 24.759128
#> Track_02  9.786452 27.538321 18.363733 17.836922
#> Track_03 24.712765  4.444202  6.527563 20.508148
#> Track_04 24.637547 43.977525 29.411795 10.955325
#> Track_07 20.908844  4.510252  6.202228 18.950349
#> Track_08 15.000687 21.766914 12.055124  6.057863
#> Track_09 34.851903 14.793056 12.290781 28.188578
#> Track_13        NA 24.467048 15.685923  9.860997
#> Track_15        NA        NA  7.187465 21.350909
#> Track_16        NA        NA        NA 11.512595
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:04.855509 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 14.94689 16.850528 59.69429 13.35986 32.89962  4.618884
#> Track_02       NA       NA  9.313509 60.12425 18.92543 32.85852 11.793902
#> Track_03       NA       NA        NA 42.21609 13.61763 19.80106 11.375467
#> Track_04       NA       NA        NA       NA 33.55558 10.23134 44.671496
#> Track_07       NA       NA        NA       NA       NA 14.17736  8.358740
#> Track_08       NA       NA        NA       NA       NA       NA 21.847294
#> Track_09       NA       NA        NA       NA       NA       NA        NA
#> Track_13       NA       NA        NA       NA       NA       NA        NA
#> Track_15       NA       NA        NA       NA       NA       NA        NA
#> Track_16       NA       NA        NA       NA       NA       NA        NA
#> Track_18       NA       NA        NA       NA       NA       NA        NA
#>           Track_13  Track_15  Track_16 Track_18
#> Track_01 44.922643  6.142617 17.926027 39.41363
#> Track_02 42.555663 11.614092 18.675312 34.88032
#> Track_03 27.745971 10.582805 10.432614 19.47778
#> Track_04  6.203218 40.837055 22.788781 23.05244
#> Track_07 22.742108  6.779830  6.218078 20.56171
#> Track_08  7.999661 18.824142  8.084479 12.96300
#> Track_09 31.472277  4.132562 10.231892 26.53259
#> Track_13        NA 27.381280 14.444593 14.01010
#> Track_15        NA        NA  9.275047 22.18297
#> Track_16        NA        NA        NA 12.08386
#> Track_18        NA        NA        NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:04.915072 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 58.83388 40.760730 50.67296 32.881838 39.13687 43.288357
#> Track_02       NA       NA  5.846582 35.36129 20.326086 22.94644 32.763582
#> Track_03       NA       NA        NA 27.76271 12.948066 12.74661 20.594433
#> Track_04       NA       NA        NA       NA  8.851782 42.58323 52.787506
#> Track_07       NA       NA        NA       NA        NA 20.79463 27.730919
#> Track_08       NA       NA        NA       NA        NA       NA  6.326259
#> Track_09       NA       NA        NA       NA        NA       NA        NA
#> Track_13       NA       NA        NA       NA        NA       NA        NA
#> Track_15       NA       NA        NA       NA        NA       NA        NA
#> Track_16       NA       NA        NA       NA        NA       NA        NA
#> Track_18       NA       NA        NA       NA        NA       NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 35.69652 42.464700 27.385961 39.806356
#> Track_02 38.90391 23.268763 12.049680  6.825146
#> Track_03 24.57827 13.400209  6.489276  2.876731
#> Track_04 53.54986 45.653838 21.223540 23.714083
#> Track_07 28.94301 22.768663  7.969480 10.617259
#> Track_08 13.57199  4.035441 10.449707 14.548152
#> Track_09 10.87301  9.344801 16.158320 22.623015
#> Track_13       NA 16.590710 17.816776 25.766004
#> Track_15       NA        NA 11.893525 15.403065
#> Track_16       NA        NA        NA  5.488689
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:04.963065 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 46.49362 35.97620  6.839742 40.00027 35.51940 17.83566
#> Track_02       NA       NA 10.16011 50.357297 24.84484 33.28324 40.18093
#> Track_03       NA       NA       NA 39.113733 11.20159 17.25245 25.43694
#> Track_04       NA       NA       NA        NA 42.49233 36.49087 16.63340
#> Track_07       NA       NA       NA        NA       NA 11.34558 25.40074
#> Track_08       NA       NA       NA        NA       NA       NA 17.82031
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 45.299238 40.060325 26.990043  6.983464
#> Track_02 31.875033 29.277731 12.978403 30.289900
#> Track_03 16.398207 14.308148  4.984731 19.094783
#> Track_04 47.763734 42.400911 28.907825  7.635871
#> Track_07  6.455369  4.397960  8.404362 21.949409
#> Track_08 11.638855  9.153832 11.031879 17.306543
#> Track_09 28.990313 24.779318 17.914370  8.798761
#> Track_13        NA  3.741605 11.847443 26.090715
#> Track_15        NA        NA  9.860525 21.899018
#> Track_16        NA        NA        NA 12.612806
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:05.011329 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 57.27626 38.33683 56.95589 27.65643 40.103142 26.12336
#> Track_02       NA       NA 24.66749 50.50474 21.38754 18.917271 43.13247
#> Track_03       NA       NA       NA 11.78510 20.15022  5.237084 18.72254
#> Track_04       NA       NA       NA       NA 42.22916 20.490605 24.91217
#> Track_07       NA       NA       NA       NA       NA 19.091881 26.58144
#> Track_08       NA       NA       NA       NA       NA        NA 22.53477
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 35.948053  6.320355 28.879758  7.076236
#> Track_02 20.625983 35.496747 15.322096 34.047085
#> Track_03 24.905360 20.873190  6.520752 18.388202
#> Track_04 48.614568 39.239639 20.500640 34.908008
#> Track_07  6.812578 14.853256 11.751054 14.069744
#> Track_08 22.366706 22.380957  5.791449 20.268010
#> Track_09 33.088664 17.510609 16.622735 14.448896
#> Track_13        NA 20.988935 15.552988 20.215023
#> Track_15        NA        NA 14.188982  3.565237
#> Track_16        NA        NA        NA 12.382900
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 7.155827 6.889059 7.146449 8.550492 9.358030 41.27125
#> Track_02       NA       NA 7.367287 7.095034 8.511176 8.404712 44.46323
#> Track_03       NA       NA       NA 2.752378 5.017897 5.558007 23.61988
#> Track_04       NA       NA       NA       NA 3.748965 5.898481 21.55969
#> Track_07       NA       NA       NA       NA       NA 3.362680 27.87452
#> Track_08       NA       NA       NA       NA       NA       NA 27.11038
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 23.69080 10.006241  8.956456 11.561263
#> Track_02 26.76916  8.652232  8.004512  9.939214
#> Track_03 11.38055  4.011654  6.187040  4.835886
#> Track_04 10.75445  2.107453  7.121107  3.128265
#> Track_07 16.24588  4.076379  5.378815  5.173189
#> Track_08 16.01165  5.869619  3.814577  6.169448
#> Track_09 12.44577 21.459210 36.643358 21.460416
#> Track_13       NA 11.406049 22.371414 11.843307
#> Track_15       NA        NA  7.547234  1.280341
#> Track_16       NA        NA        NA  8.182026
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0      0.0      0.1      0.0      0.0      0.6
#> Track_02       NA       NA      0.1      0.0      0.0      0.1      1.0
#> Track_03       NA       NA       NA      0.0      0.1      0.1      0.6
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.0      1.0
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.2      0.3      0.0      0.2
#> Track_02      0.3      0.0      0.0      0.2
#> Track_03      0.0      0.0      0.1      0.1
#> Track_04      0.1      0.0      0.0      0.0
#> Track_07      0.5      0.1      0.0      0.0
#> Track_08      0.4      0.1      0.0      0.2
#> Track_09      0.4      0.5      1.0      0.3
#> Track_13       NA      0.2      1.0      0.1
#> Track_15       NA       NA      0.1      0.0
#> Track_16       NA       NA       NA      0.2
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 5: Simulating and comparing DTW distance in the MountTom dataset using the
# "Origin" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
#> 2025-05-14 06:09:05.301971 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 114.9419 27.61222  15.19719 78.242440 76.410618 47.20332
#> Track_02       NA       NA 68.29994 119.16984  7.066379  8.077849 65.45061
#> Track_03       NA       NA       NA  20.98662 43.590491 41.668877 41.19465
#> Track_04       NA       NA       NA        NA 81.067041 79.679975 56.22650
#> Track_07       NA       NA       NA        NA        NA  3.394761 43.28059
#> Track_08       NA       NA       NA        NA        NA        NA 39.69709
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15 Track_16 Track_18
#> Track_01 92.33101 42.605363 45.81118 61.81022
#> Track_02 28.37741 59.025928 26.17131 56.43568
#> Track_03 49.84236 34.193632 22.49202 22.71821
#> Track_04 94.13492 51.338571 48.79552 57.76961
#> Track_07 18.13536 37.214182 13.61901 34.12374
#> Track_08 20.12824 34.096370 11.41660 33.93014
#> Track_09 61.55986  4.280689 18.18983 51.40889
#> Track_13       NA 52.904450 23.79197 30.81986
#> Track_15       NA        NA 15.85254 43.08428
#> Track_16       NA        NA       NA 21.91301
#> Track_18       NA        NA       NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:05.358974 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 106.9958 59.51294  35.02641 66.511787 55.45289 77.83471
#> Track_02       NA       NA 26.74464 114.96187 14.909491 64.72675 75.11784
#> Track_03       NA       NA       NA  70.90295  8.727898 41.07661 52.14028
#> Track_04       NA       NA       NA        NA 74.050292 39.24158 60.13534
#> Track_07       NA       NA       NA        NA        NA 39.38521 49.88928
#> Track_08       NA       NA       NA        NA        NA       NA 10.66167
#> Track_09       NA       NA       NA        NA        NA       NA       NA
#> Track_13       NA       NA       NA        NA        NA       NA       NA
#> Track_15       NA       NA       NA        NA        NA       NA       NA
#> Track_16       NA       NA       NA        NA        NA       NA       NA
#> Track_18       NA       NA       NA        NA        NA       NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 56.790908 77.142247 46.832568 53.922156
#> Track_02 79.844629  6.647412 42.575500 29.041901
#> Track_03 52.608716 19.568460 24.522782  4.102476
#> Track_04 35.090336 78.931188 39.196348 65.630143
#> Track_07 51.095504 11.352300 22.883910 10.220759
#> Track_08  8.266075 40.233504  6.895479 38.591261
#> Track_09 21.331765 48.281949  9.916854 50.740156
#> Track_13        NA 53.424901 14.657031 48.839575
#> Track_15        NA        NA 23.386369 20.243503
#> Track_16        NA        NA        NA 23.037902
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:05.414089 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 61.27835 74.60001 103.85635 25.50259 28.19245 92.99342
#> Track_02       NA       NA 70.22435  49.02892 57.88358 18.72927 70.40276
#> Track_03       NA       NA       NA  68.12676 37.70066 41.53848 23.62746
#> Track_04       NA       NA       NA        NA 79.35284 49.55198 50.17552
#> Track_07       NA       NA       NA        NA       NA 25.69345 52.76757
#> Track_08       NA       NA       NA        NA       NA       NA 47.55108
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>          Track_13  Track_15 Track_16 Track_18
#> Track_01 58.23308 76.441447 51.48413 44.52798
#> Track_02 80.63730 71.298464 40.31906 66.36912
#> Track_03 31.81236  3.848501 12.96093 27.78640
#> Track_04 93.50904 68.422137 36.55312 79.59084
#> Track_07 21.10220 41.151037 24.08375 13.88949
#> Track_08 43.03033 43.095919 21.11539 32.66261
#> Track_09 56.03263 22.574589 10.64465 47.55897
#> Track_13       NA 35.243426 27.43441  4.83576
#> Track_15       NA        NA 13.65743 30.72697
#> Track_16       NA        NA       NA 21.01989
#> Track_18       NA        NA       NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:05.462974 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 82.46812 77.51597 136.5904 25.29040 76.59155 91.59829
#> Track_02       NA       NA 62.35406 106.2899 62.73109 38.61964 60.62170
#> Track_03       NA       NA       NA  15.6012 38.76344 21.55043 13.65032
#> Track_04       NA       NA       NA       NA 83.10795 39.24681 20.63777
#> Track_07       NA       NA       NA       NA       NA 42.44922 48.31352
#> Track_08       NA       NA       NA       NA       NA       NA 15.40527
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 35.031353 50.82020 54.115597 75.027660
#> Track_02 75.376644 68.72300 31.712774 32.062952
#> Track_03 45.421968 29.14467 13.935966 25.634507
#> Track_04 92.369254 66.38923 35.977614 47.632849
#> Track_07  5.300677 17.07603 25.001497 43.149673
#> Track_08 52.376019 39.96305  6.548631  5.297463
#> Track_09 57.532860 42.73605 12.885493 20.989976
#> Track_13        NA 16.70806 32.267327 53.503373
#> Track_15        NA       NA 23.422616 41.901050
#> Track_16        NA       NA        NA  8.441816
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:05.512545 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03   Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 120.9467 78.48570 136.219980 70.417740 61.32926 97.82329
#> Track_02       NA       NA 18.40515   2.965189 31.628675 53.80254 29.57720
#> Track_03       NA       NA       NA  21.501977  9.723601 36.26665 30.13534
#> Track_04       NA       NA       NA         NA 36.659897 64.05659 35.92769
#> Track_07       NA       NA       NA         NA        NA 38.01811 37.82505
#> Track_08       NA       NA       NA         NA        NA       NA 28.39558
#> Track_09       NA       NA       NA         NA        NA       NA       NA
#> Track_13       NA       NA       NA         NA        NA       NA       NA
#> Track_15       NA       NA       NA         NA        NA       NA       NA
#> Track_16       NA       NA       NA         NA        NA       NA       NA
#> Track_18       NA       NA       NA         NA        NA       NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 23.09985 43.30047 35.16588 23.679274
#> Track_02 81.54340 64.00528 44.51584 69.621089
#> Track_03 46.93249 32.28134 23.61209 41.553788
#> Track_04 94.13468 74.60776 54.15657 81.633464
#> Track_07 38.60080 23.80142 22.70254 38.738208
#> Track_08 45.26071 41.38659 11.39804 23.715404
#> Track_09 65.27773 54.22342 29.29551 49.404717
#> Track_13       NA 17.01220 23.37619 24.332411
#> Track_15       NA       NA 22.19391 30.417722
#> Track_16       NA       NA       NA  9.232054
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:05.574592 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 93.18309 42.14161 137.40040 79.22886 71.49545 87.47432
#> Track_02       NA       NA 30.59065  87.74585 58.81261 63.60379 28.64604
#> Track_03       NA       NA       NA  83.31230 44.93936 42.14863 37.26424
#> Track_04       NA       NA       NA        NA 19.33071 38.29402 43.51667
#> Track_07       NA       NA       NA        NA       NA 12.16438 30.42843
#> Track_08       NA       NA       NA        NA       NA       NA 37.51664
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 89.955962 63.07213 40.136331 76.802375
#> Track_02 70.920416 67.85648 21.874904 28.525353
#> Track_03 52.959517 42.98612  8.992861 31.558168
#> Track_04 26.543332 55.22997 54.349545 35.405859
#> Track_07  5.968543 22.71999 24.964480 22.531275
#> Track_08  9.884529 11.39745 24.187025 28.494368
#> Track_09 39.766338 45.14072 21.218228  5.012777
#> Track_13        NA 22.19081 31.593757 30.369166
#> Track_15        NA       NA 25.242545 35.709426
#> Track_16        NA       NA        NA 17.411763
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:05.62368 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA  114.636 11.12555 73.46063 78.39042 16.153662 32.84386
#> Track_02       NA       NA 66.43579 87.01535 32.84035 67.255173 81.84437
#> Track_03       NA       NA       NA 42.82704 42.36006  7.843127 20.88472
#> Track_04       NA       NA       NA       NA 81.74412 54.252545 77.07378
#> Track_07       NA       NA       NA       NA       NA 39.020723 46.37701
#> Track_08       NA       NA       NA       NA       NA        NA 11.66552
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>          Track_13 Track_15  Track_16 Track_18
#> Track_01 32.11386 75.79065 53.463118 74.98701
#> Track_02 68.55352 16.09951 26.522582 12.63322
#> Track_03 15.41233 39.10506 23.254536 37.72959
#> Track_04 23.89604 51.93955 54.420712 53.18963
#> Track_07 51.47187 27.85541  7.458226 24.21372
#> Track_08 23.00608 39.89830 22.020006 38.66925
#> Track_09 39.32825 51.47940 30.398190 49.92521
#> Track_13       NA 39.23997 29.793574 38.57059
#> Track_15       NA       NA 17.194568  4.23111
#> Track_16       NA       NA        NA 14.79304
#> Track_18       NA       NA        NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:05.673577 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 108.5092 78.47841 94.11726 80.946449 10.97670 33.25192
#> Track_02       NA       NA 14.64402 51.87306 18.372664 61.79305 60.31717
#> Track_03       NA       NA       NA 50.63828  4.756774 39.62581 43.31253
#> Track_04       NA       NA       NA       NA 56.848298 56.49025 41.74736
#> Track_07       NA       NA       NA       NA        NA 43.08265 46.89236
#> Track_08       NA       NA       NA       NA        NA       NA 16.03387
#> Track_09       NA       NA       NA       NA        NA       NA       NA
#> Track_13       NA       NA       NA       NA        NA       NA       NA
#> Track_15       NA       NA       NA       NA        NA       NA       NA
#> Track_16       NA       NA       NA       NA        NA       NA       NA
#> Track_18       NA       NA       NA       NA        NA       NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 84.98165 56.97390 34.99100 79.415889
#> Track_02 67.24907 66.83004 32.93033 20.140690
#> Track_03 37.29293 40.81728 19.43878  4.099402
#> Track_04 98.33663 86.04472 30.95792 56.633246
#> Track_07 38.47242 44.31813 21.87842  4.011966
#> Track_08 46.85649 30.21927 13.04613 40.981389
#> Track_09 59.74359 45.88990 10.69795 45.450650
#> Track_13       NA 22.13308 31.95259 34.530608
#> Track_15       NA       NA 24.46041 39.570805
#> Track_16       NA       NA       NA 20.520207
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-05-14 06:09:05.735453 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 96.37777 24.32841 71.99472 20.63707 74.66871  93.08738
#> Track_02       NA       NA 64.07417 53.58912 47.95073 17.31372  73.37925
#> Track_03       NA       NA       NA 61.84562 20.47955 40.09152  45.01094
#> Track_04       NA       NA       NA       NA 29.92240 54.08574 103.02297
#> Track_07       NA       NA       NA       NA       NA 34.07448  53.68247
#> Track_08       NA       NA       NA       NA       NA       NA  39.51899
#> Track_09       NA       NA       NA       NA       NA       NA        NA
#> Track_13       NA       NA       NA       NA       NA       NA        NA
#> Track_15       NA       NA       NA       NA       NA       NA        NA
#> Track_16       NA       NA       NA       NA       NA       NA        NA
#> Track_18       NA       NA       NA       NA       NA       NA        NA
#>           Track_13  Track_15  Track_16 Track_18
#> Track_01 85.318873 72.571128 35.157145 53.08566
#> Track_02 72.275893  5.957859 44.248630 19.17104
#> Track_03 40.491988 41.862576  9.204589 33.29441
#> Track_04 99.222945 40.702763 52.002770 21.78652
#> Track_07 52.096061 31.786535 17.844821 19.99105
#> Track_08 39.283141 12.585835 22.808547 19.21655
#> Track_09  6.763623 51.892079 25.714578 51.53802
#> Track_13        NA 52.165586 22.327876 49.29969
#> Track_15        NA        NA 26.087395 12.08625
#> Track_16        NA        NA        NA 21.48215
#> Track_18        NA        NA        NA       NA
#> ------------------------------------
#> 2025-05-14 06:09:05.784027 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 102.7012 11.99830  70.61498 15.032315 65.49136 55.50118
#> Track_02       NA       NA 61.69689 118.18384 65.262123 13.71579 41.20147
#> Track_03       NA       NA       NA  38.60831  4.185147 34.32260 33.93688
#> Track_04       NA       NA       NA        NA 35.974914 80.40040 88.67016
#> Track_07       NA       NA       NA        NA        NA 37.61286 37.90455
#> Track_08       NA       NA       NA        NA        NA       NA 19.30177
#> Track_09       NA       NA       NA        NA        NA       NA       NA
#> Track_13       NA       NA       NA        NA        NA       NA       NA
#> Track_15       NA       NA       NA        NA        NA       NA       NA
#> Track_16       NA       NA       NA        NA        NA       NA       NA
#> Track_18       NA       NA       NA        NA        NA       NA       NA
#>          Track_13 Track_15  Track_16  Track_18
#> Track_01 92.64006 81.15583 28.852153 71.160715
#> Track_02 55.14249 29.66951 43.298863  7.013416
#> Track_03 49.43132 43.45165  9.158358 36.870169
#> Track_04 72.25278 79.13756 27.804887 80.647924
#> Track_07 51.45218 47.34783  9.617857 39.968430
#> Track_08 40.82779 24.12643 21.868857  6.505135
#> Track_09 60.65969 45.15392 28.155304 26.335398
#> Track_13       NA 20.16201 27.977926 35.928299
#> Track_15       NA       NA 24.953538 18.638614
#> Track_16       NA       NA        NA 22.465872
#> Track_18       NA       NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 8.365556 8.140001 13.288601 17.963246 14.043056 80.66455
#> Track_02       NA       NA 9.928547 13.870318 15.997694 11.986057 83.72110
#> Track_03       NA       NA       NA  4.242441  9.250629  6.131319 46.49768
#> Track_04       NA       NA       NA        NA  5.401950  3.172849 42.61717
#> Track_07       NA       NA       NA        NA        NA  5.261381 51.73148
#> Track_08       NA       NA       NA        NA        NA        NA 54.42241
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.00788 18.730370 14.932046 22.179224
#> Track_02 47.19796 17.799527  9.836672 20.569492
#> Track_03 20.29566  7.910351 10.158387  9.928217
#> Track_04 19.90114  4.165947  8.969975  6.396727
#> Track_07 29.44950  2.944420  6.952380  5.167036
#> Track_08 27.32867  4.915326  5.349900  6.863688
#> Track_09 22.41087 41.825031 68.973412 42.969072
#> Track_13       NA 20.579060 39.239001 21.761749
#> Track_15       NA        NA  8.572483  2.784293
#> Track_16       NA        NA        NA 10.795457
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0        0      0.0      0.1      0.1      0.5
#> Track_02       NA       NA        0      0.1      0.2      0.1      1.0
#> Track_03       NA       NA       NA      0.0      0.3      0.0      0.9
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.1      0.8
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.3      0.0      0.0      0.0
#> Track_02      0.1      0.3      0.0      0.4
#> Track_03      0.1      0.1      0.3      0.2
#> Track_04      0.0      0.0      0.0      0.0
#> Track_07      0.4      0.0      0.0      0.1
#> Track_08      0.4      0.0      0.0      0.2
#> Track_09      0.2      0.2      1.0      0.3
#> Track_13       NA      0.3      1.0      0.1
#> Track_15       NA       NA      0.0      0.0
#> Track_16       NA       NA       NA      0.2
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#>