Similarity metric using Dynamic Time Warping (DTW)

simil_DTW_metric() computes similarity metrics between two or more trajectories using Dynamic Time Warping (DTW). It allows for different superposition methods to align trajectories before calculating the DTW metric. The function also supports testing with simulations to calculate p-values for the DTW distance metrics.

Usage

simil_DTW_metric(data, test = NULL, sim = NULL, superposition = NULL)

Arguments

data

A track R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

test

Logical; if TRUE, the function compares the observed DTW distances against simulated trajectories and calculates p-values. Default is FALSE.

sim

A track simulation R object consisting of a list of simulated trajectories to use for comparison when test = TRUE.

superposition

A character string indicating the method used to align trajectories. Options are "None", "Centroid", or "Origin". Default is "None".

Value

A track similarity R object consisting ofa list containing the following elements:

DTW_distance_metric

A matrix containing the pairwise DTW distances between trajectories.

DTW_distance_metric_p_values

(If test is TRUE) A matrix containing the p-values for the pairwise DTW distances.

DTW_metric_p_values_combined

(If test is TRUE) The overall p-value for the combined DTW distances.

DTW_distance_metric_simulations

(If test is TRUE) A list of DTW distance matrices from each simulated dataset.

Details

The simil_DTW_metric() function calculates the similarity between trajectories using the Dynamic Time Warping (DTW) algorithm from the dtw package. The dtw() function is used with the dist.method argument set to "Euclidean" for computing the local distances between points in the trajectories.

DTW aligns two time series by minimizing the cumulative distance between their points, creating an optimal alignment despite variations in length or temporal distortions. The algorithm constructs a distance matrix where each element represents the cost of aligning points between the two series and finds a warping path through this matrix that minimizes the total distance. The warping path is contiguous and monotonic, starting from the bottom-left corner and ending at the top-right corner (Cleasby et al., 2019).

DTW measures are non-negative and unbounded, with larger values indicating greater dissimilarity between the time series. This method has been used in various contexts, including ecological studies to analyze and cluster trajectory data (Cleasby et al., 2019).

Potential limitations and biases of DTW include sensitivity to noise and outliers, computational complexity, and the need for appropriate distance metrics. Additionally, DTW may not always account for all structural differences between trajectories and can be biased by the chosen alignment constraints. While DTW can handle trajectories of different lengths due to its elastic nature, having trajectories of similar lengths can improve the accuracy and interpretability of the similarity measure. Similar lengths result in a more meaningful alignment and can make the computation more efficient. When trajectories differ significantly in length, preprocessing or normalization might be necessary, and careful analysis is required to understand the alignment path. The function’s flexibility in handling different lengths allows it to be applied in various contexts. However, large differences in trajectory lengths might introduce potential biases that should be considered when interpreting the results.

The function offers three different superposition methods to align the trajectories before DTW() is applied:

"None": No superposition is applied.

"Centroid": Trajectories are shifted to align based on their centroids.

"Origin": Trajectories are shifted to align based on their starting point.

If test = TRUE, the function can compute p-values by comparing the observed DTW distances with those generated from a set of simulated trajectories. The p-values are calculated for both individual trajectory pairs and for the entire set of trajectories.

Logo

References

Cleasby, I. R., Wakefield, E. D., Morrissey, B. J., Bodey, T. W., Votier, S. C., Bearhop, S., & Hamer, K. C. (2019). Using time-series similarity measures to compare animal movement trajectories in ecology. Behavioral Ecology and Sociobiology, 73, 1-19.

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Simulating tracks using the "Directed" model and comparing DTW distance
# in the PaluxyRiver dataset
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-04-16 07:49:42.063148 Iteration 1
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 12.8887
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-04-16 07:49:42.064945 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 22.57879
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.066734 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 27.97542
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.068409 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 21.78067
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.070085 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 15.62574
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.071757 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 14.81205
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.073377 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 16.12791
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.075019 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 19.94657
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.084407 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 28.88161
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.086524 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 33.21714
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.7
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.7
#> 

# Example 2: Simulating tracks using the "Constrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
#> 2025-04-16 07:49:42.120611 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 99.58085
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.122423 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 345.4952
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.124177 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 268.4955
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.125914 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 101.9855
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.127653 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 114.5313
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.129371 Iteration 6
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 174.109
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-04-16 07:49:42.131092 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 38.55042
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.13272 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 59.56638
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.134318 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 88.63144
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.135939 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 38.28449
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 3: Simulating tracks using the "Unconstrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
#> 2025-04-16 07:49:42.166354 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 676.6944
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.168185 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 630.3554
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.169974 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 660.8249
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.171652 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 563.0772
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.173295 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 304.8335
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.174979 Iteration 6
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 671.057
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-04-16 07:49:42.177109 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 622.3867
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.178778 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 587.9958
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.180402 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 783.5547
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-04-16 07:49:42.182071 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 203.4594
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 4: Simulating and comparing DTW distance in the MountTom dataset using the
# "Centroid" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
#> 2025-04-16 07:49:42.432085 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 49.48839 37.72422  6.499613 39.891361 19.76850 46.260151
#> Track_02       NA       NA 31.06919 51.929135 22.467777 20.53299 33.375851
#> Track_03       NA       NA       NA 40.662627  9.328148 20.88033  5.611117
#> Track_04       NA       NA       NA        NA 43.236580 20.54884 49.482108
#> Track_07       NA       NA       NA        NA        NA 20.08584  8.387010
#> Track_08       NA       NA       NA        NA        NA       NA 26.228865
#> Track_09       NA       NA       NA        NA        NA       NA        NA
#> Track_13       NA       NA       NA        NA        NA       NA        NA
#> Track_15       NA       NA       NA        NA        NA       NA        NA
#> Track_16       NA       NA       NA        NA        NA       NA        NA
#> Track_18       NA       NA       NA        NA        NA       NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 46.458903 38.915949 12.896604 37.674661
#> Track_02 27.088404 11.307420 22.853667 13.634140
#> Track_03 11.561243 16.232329 12.345964 13.348203
#> Track_04 50.054226 42.513122 15.452667 40.514773
#> Track_07  5.099155  9.307728 13.572504  7.033696
#> Track_08 25.324050 16.811545  8.155777 15.580030
#> Track_09  8.784503 17.508859 17.741318 14.602142
#> Track_13        NA 13.115524 17.982035 10.905696
#> Track_15        NA        NA 13.731172  3.603609
#> Track_16        NA        NA        NA 12.248871
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.478266 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 6.496744 38.44905 27.60842 39.61492  7.754836 41.28644
#> Track_02       NA       NA 34.33788 29.60120 36.29327  9.714476 39.26616
#> Track_03       NA       NA       NA 42.08161 10.68869 19.999599 20.89808
#> Track_04       NA       NA       NA       NA 37.24443 16.006888 29.71245
#> Track_07       NA       NA       NA       NA       NA 20.247520 12.09664
#> Track_08       NA       NA       NA       NA       NA        NA 21.96640
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 10.622288 28.30571 21.514450 39.079278
#> Track_02 11.176914 27.27619 18.147633 35.508651
#> Track_03 26.239643 20.84914  8.897552  5.297214
#> Track_04 14.845670 14.85471 28.035637 40.085945
#> Track_07 25.894541 16.27588 12.292096  5.931462
#> Track_08  5.958309 12.91756  9.993296 20.220322
#> Track_09 25.867680 11.48428 18.108545 16.738753
#> Track_13        NA 15.50938 14.867228 26.151796
#> Track_15        NA       NA 13.879913 18.868867
#> Track_16        NA       NA        NA 10.389073
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.533464 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 32.60506 35.22810 64.28502 39.44302 39.20680  9.293318
#> Track_02       NA       NA 35.02558 38.14640 34.86112 25.99102 19.731432
#> Track_03       NA       NA       NA 36.58276  6.08327 15.27986 25.499172
#> Track_04       NA       NA       NA       NA 29.74903 10.93935 45.659887
#> Track_07       NA       NA       NA       NA       NA 11.54970 27.289022
#> Track_08       NA       NA       NA       NA       NA       NA 25.940240
#> Track_09       NA       NA       NA       NA       NA       NA        NA
#> Track_13       NA       NA       NA       NA       NA       NA        NA
#> Track_15       NA       NA       NA       NA       NA       NA        NA
#> Track_16       NA       NA       NA       NA       NA       NA        NA
#> Track_18       NA       NA       NA       NA       NA       NA        NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 34.56306 39.902238 18.08819 19.543236
#> Track_02 41.06756 22.354015 11.82155  8.611295
#> Track_03  9.84077 19.950487 12.93445 20.337174
#> Track_04 48.68773  7.580148 25.54926 31.401490
#> Track_07 15.35222 16.618225 13.35391 20.910585
#> Track_08 24.24450  6.541708 11.33778 16.713083
#> Track_09 27.55308 25.634256  8.90140  9.256943
#> Track_13       NA 28.706963 17.71077 25.342310
#> Track_15       NA        NA 11.62186 15.525032
#> Track_16       NA        NA       NA  4.746869
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.582193 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 54.05736 12.81918 68.18528 39.72505  6.194643 10.203785
#> Track_02       NA       NA 35.05053 28.16934 29.69493 33.027697 43.345266
#> Track_03       NA       NA       NA 42.33315 18.05821  5.839445  4.232463
#> Track_04       NA       NA       NA       NA 21.14661 43.031756 52.625734
#> Track_07       NA       NA       NA       NA       NA 20.795972 25.776636
#> Track_08       NA       NA       NA       NA       NA        NA  6.211979
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 45.193483 40.212744 25.244560 34.129537
#> Track_02 31.284594  8.839024 11.832105  9.408638
#> Track_03 22.564860 22.210060 11.977166 19.523877
#> Track_04 20.360929 15.219600 21.992419 26.543879
#> Track_07  4.881588 15.858105 11.940707 19.156541
#> Track_08 25.129402 21.447940 11.079230 17.218361
#> Track_09 30.845863 29.219962 17.504424 25.906949
#> Track_13        NA 17.316058 14.993282 21.874984
#> Track_15        NA        NA  7.445933  8.119447
#> Track_16        NA        NA        NA  4.966350
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.628308 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 10.61006 38.44387 44.84501 37.16162 31.731786 28.58519
#> Track_02       NA       NA 33.44924 47.16458 34.30621 30.423888 21.86477
#> Track_03       NA       NA       NA 39.71073 13.72439 16.618113 16.31216
#> Track_04       NA       NA       NA       NA 21.09903 11.383072 46.76723
#> Track_07       NA       NA       NA       NA       NA  5.981128 23.92511
#> Track_08       NA       NA       NA       NA       NA        NA 23.78463
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01  5.217078 22.119530 19.638701 22.841045
#> Track_02  6.879275 15.588919 19.001866 23.678349
#> Track_03 26.013534 17.338894 12.551134 19.754292
#> Track_04 34.854514 42.235620 15.969121 10.690704
#> Track_07 24.800241 22.417526  9.023775 12.608280
#> Track_08 20.782867 21.058122  6.376813  7.499457
#> Track_09 20.143988  6.545967 16.108783 23.080650
#> Track_13        NA 15.088868 12.281960 15.300243
#> Track_15        NA        NA 12.970731 19.648293
#> Track_16        NA        NA        NA  4.772045
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.702154 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04  Track_07  Track_08  Track_09
#> Track_01       NA 41.52605 25.113829 51.02697  7.668833  9.218685  9.833736
#> Track_02       NA       NA  8.995112 63.10255 28.103764 22.040894 26.064125
#> Track_03       NA       NA        NA 42.69529 13.939712 10.060556 13.010351
#> Track_04       NA       NA        NA       NA 30.595770 36.446037 43.775858
#> Track_07       NA       NA        NA       NA        NA  5.615136  8.473484
#> Track_08       NA       NA        NA       NA        NA        NA  4.391815
#> Track_09       NA       NA        NA       NA        NA        NA        NA
#> Track_13       NA       NA        NA       NA        NA        NA        NA
#> Track_15       NA       NA        NA       NA        NA        NA        NA
#> Track_16       NA       NA        NA       NA        NA        NA        NA
#> Track_18       NA       NA        NA       NA        NA        NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.35876 39.809809 15.214314 11.410778
#> Track_02 39.64227 35.536520 21.707243 20.389310
#> Track_03 25.79559 21.698333  9.726492  8.953525
#> Track_04 21.75461 17.755364 23.145157 38.032580
#> Track_07 24.45328 19.834865  5.214698  6.923908
#> Track_08 25.61186 21.108073  6.752757  1.755988
#> Track_09 32.93482 27.902992  9.321203  4.353211
#> Track_13       NA  4.573231 16.358399 26.387796
#> Track_15       NA        NA 12.287026 21.759091
#> Track_16       NA        NA        NA  7.206225
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.752476 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 35.38378 22.34237 67.17550 35.98111 39.385702 50.69397
#> Track_02       NA       NA 33.01023 41.12075 36.11587 23.872541 37.81101
#> Track_03       NA       NA       NA 43.11192 10.50121 20.697066 25.23233
#> Track_04       NA       NA       NA       NA 36.49181  7.410911 17.26381
#> Track_07       NA       NA       NA       NA       NA 17.659378 17.62982
#> Track_08       NA       NA       NA       NA       NA        NA 10.49352
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 20.65355 37.475920 28.714105 15.368418
#> Track_02 11.63919 36.898007 24.331773 13.052992
#> Track_03 22.30405 11.578289 10.670770 15.840371
#> Track_04 43.07703 36.937093 20.962827 38.106528
#> Track_07 28.03959  1.801336  8.204668 20.864953
#> Track_08 22.42256 18.398046  8.428318 17.543587
#> Track_09 32.99090 17.751906  8.558342 27.023651
#> Track_13       NA 28.909334 18.257616  5.853317
#> Track_15       NA        NA  8.908265 21.825645
#> Track_16       NA        NA        NA 12.831407
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.799852 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 16.64702 25.70184 53.74426 34.899512  7.001622 48.42532
#> Track_02       NA       NA 16.81010 39.48817 26.511385 12.858215 42.08496
#> Track_03       NA       NA       NA 13.82454  7.472254 13.240113 21.30478
#> Track_04       NA       NA       NA       NA  9.040556 35.662520 31.39062
#> Track_07       NA       NA       NA       NA        NA 18.000021 16.16132
#> Track_08       NA       NA       NA       NA        NA        NA 26.67017
#> Track_09       NA       NA       NA       NA        NA        NA       NA
#> Track_13       NA       NA       NA       NA        NA        NA       NA
#> Track_15       NA       NA       NA       NA        NA        NA       NA
#> Track_16       NA       NA       NA       NA        NA        NA       NA
#> Track_18       NA       NA       NA       NA        NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 10.970294 21.849218 26.576956 39.579901
#> Track_02 19.115439 11.812732 22.388122 36.050900
#> Track_03 19.525549  5.393333  7.879617 16.803587
#> Track_04 45.146985 21.805255 17.926532 29.461000
#> Track_07 24.681685 12.280824  6.076864 13.053550
#> Track_08  7.375553 11.445043 12.193272 20.876035
#> Track_09 32.430856 25.519529 10.737620  4.412086
#> Track_13        NA 18.363534 17.078373 26.090940
#> Track_15        NA        NA 10.641122 20.561528
#> Track_16        NA        NA        NA  7.949054
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.846548 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 34.72789 40.45212 42.66309 24.714960 39.17279 50.376743
#> Track_02       NA       NA 31.87355 59.17641  5.779139 22.60396 39.625103
#> Track_03       NA       NA       NA 33.60193 18.691362  9.58802  3.537922
#> Track_04       NA       NA       NA       NA 41.709052 41.88800 39.630262
#> Track_07       NA       NA       NA       NA        NA 13.47034 25.573411
#> Track_08       NA       NA       NA       NA        NA       NA 13.534642
#> Track_09       NA       NA       NA       NA        NA       NA        NA
#> Track_13       NA       NA       NA       NA        NA       NA        NA
#> Track_15       NA       NA       NA       NA        NA       NA        NA
#> Track_16       NA       NA       NA       NA        NA       NA        NA
#> Track_18       NA       NA       NA       NA        NA       NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 46.241519 26.640530 28.280742 20.764974
#> Track_02 40.733785 35.877688 19.011826 31.517149
#> Track_03  9.256287 19.040777  7.221922 17.602914
#> Track_04 31.030644  8.404456 29.362663 11.583660
#> Track_07 26.628518 21.965847  9.822451 18.016172
#> Track_08 18.117669 22.982201  5.185380 20.059422
#> Track_09  9.880841 25.366661 10.665485 24.138018
#> Track_13        NA 20.071733 13.113376 19.771872
#> Track_15        NA        NA 14.323895  4.710111
#> Track_16        NA        NA        NA 12.329980
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:42.904648 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 53.20537 40.23948 66.43437 26.18644 40.518389 50.138099
#> Track_02       NA       NA 13.67984 43.17202 35.84426 24.845442 23.816154
#> Track_03       NA       NA       NA 20.71934 19.42603  8.916236  7.503936
#> Track_04       NA       NA       NA       NA 29.93001  7.266559 13.865524
#> Track_07       NA       NA       NA       NA       NA 15.392214 23.371250
#> Track_08       NA       NA       NA       NA       NA        NA  6.836780
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01  7.717392 27.860019 25.324314 37.678864
#> Track_02 39.855316 37.254863 23.869786 30.262099
#> Track_03 27.002325 21.076511 10.292606 13.254425
#> Track_04 47.951063 31.792697 18.982776 12.313564
#> Track_07 15.660772  2.192856  7.053502 11.212806
#> Track_08 26.688859 16.621569  7.805818  6.179761
#> Track_09 33.713472 24.684818 12.915678 13.072357
#> Track_13        NA 16.407948 15.448685 23.590179
#> Track_15        NA        NA  8.298826 11.972671
#> Track_16        NA        NA        NA  5.596648
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 7.155827 6.889059 7.146449 8.550492 9.358030 41.27125
#> Track_02       NA       NA 7.367287 7.095034 8.511176 8.404712 44.46323
#> Track_03       NA       NA       NA 2.752378 5.017897 5.558007 23.61988
#> Track_04       NA       NA       NA       NA 3.748965 5.898481 21.55969
#> Track_07       NA       NA       NA       NA       NA 3.362680 27.87452
#> Track_08       NA       NA       NA       NA       NA       NA 27.11038
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 23.69080 10.006241  8.956456 11.561263
#> Track_02 26.76916  8.652232  8.004512  9.939214
#> Track_03 11.38055  4.011654  6.187040  4.835886
#> Track_04 10.75445  2.107453  7.121107  3.128265
#> Track_07 16.24588  4.076379  5.378815  5.173189
#> Track_08 16.01165  5.869619  3.814577  6.169448
#> Track_09 12.44577 21.459210 36.643358 21.460416
#> Track_13       NA 11.406049 22.371414 11.843307
#> Track_15       NA        NA  7.547234  1.280341
#> Track_16       NA        NA        NA  8.182026
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA      0.1        0      0.1      0.1      0.4      0.4
#> Track_02       NA       NA        0      0.0      0.1      0.0      1.0
#> Track_03       NA       NA       NA      0.0      0.0      0.0      0.8
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.2
#> Track_07       NA       NA       NA       NA       NA      0.0      1.0
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.5      0.0      0.0      0.1
#> Track_02      0.4      0.0      0.0      0.2
#> Track_03      0.2      0.0      0.0      0.0
#> Track_04      0.0      0.0      0.0      0.0
#> Track_07      0.4      0.2      0.1      0.0
#> Track_08      0.2      0.0      0.0      0.1
#> Track_09      0.2      0.4      1.0      0.6
#> Track_13       NA      0.1      1.0      0.2
#> Track_15       NA       NA      0.1      0.0
#> Track_16       NA       NA       NA      0.6
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 5: Simulating and comparing DTW distance in the MountTom dataset using the
# "Origin" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
#> 2025-04-16 07:49:43.180958 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 90.40259 75.90712 114.52521 79.92326 73.99139 75.34066
#> Track_02       NA       NA 21.15161  20.80824 46.52682 62.59534 81.92975
#> Track_03       NA       NA       NA  15.32559 19.58108 33.09085 49.25529
#> Track_04       NA       NA       NA        NA 44.67610 67.78561 91.53357
#> Track_07       NA       NA       NA        NA       NA 19.28353 37.77531
#> Track_08       NA       NA       NA        NA       NA       NA 16.65489
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>          Track_13 Track_15  Track_16  Track_18
#> Track_01 26.03147 56.68455 47.899845 70.377197
#> Track_02 50.01113 14.03918 45.375656 14.991980
#> Track_03 43.64772 20.09758 23.212204  4.985953
#> Track_04 69.72594 28.74177 53.334647 12.474063
#> Track_07 53.14525 34.25233 18.244908 20.912721
#> Track_08 50.72021 40.58592  9.394512 32.534029
#> Track_09 58.99561 52.77343 10.316236 48.476598
#> Track_13       NA 26.05384 30.802960 38.053243
#> Track_15       NA       NA 25.442502 15.725010
#> Track_16       NA       NA        NA 22.147178
#> Track_18       NA       NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.229455 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 72.25973 69.11681 125.91233 78.74509 54.21547 73.05871
#> Track_02       NA       NA 67.78140  55.56065 61.93515 66.52466 19.47827
#> Track_03       NA       NA       NA  78.26432 20.74158 12.95021 50.59618
#> Track_04       NA       NA       NA        NA 52.65177 84.76590 27.96034
#> Track_07       NA       NA       NA        NA       NA 29.52835 41.07528
#> Track_08       NA       NA       NA        NA       NA       NA 51.00344
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 10.40794 76.782502 54.49697 71.916477
#> Track_02 57.48127 69.921930 35.41344 66.398684
#> Track_03 44.28025  7.615042 18.87892  5.893040
#> Track_04 97.18510 78.338816 32.97002 70.925296
#> Track_07 54.54831 17.584457 11.10255 15.171248
#> Track_08 32.76062 19.493019 21.37709 17.459324
#> Track_09 53.85232 51.920629 20.55003 48.167593
#> Track_13       NA 52.020225 32.87013 45.880295
#> Track_15       NA        NA 19.53292  3.889321
#> Track_16       NA        NA       NA 16.041225
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.276591 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 92.44972 79.19355 133.53517 10.30060 27.25575 99.367437
#> Track_02       NA       NA 43.44337  86.55940 54.46174 63.79083 62.883876
#> Track_03       NA       NA       NA  19.71583 44.94207 40.77946 11.649980
#> Track_04       NA       NA       NA        NA 82.60953 75.40074  7.308315
#> Track_07       NA       NA       NA        NA       NA 16.14401 55.126125
#> Track_08       NA       NA       NA        NA       NA       NA 50.763739
#> Track_09       NA       NA       NA        NA       NA       NA        NA
#> Track_13       NA       NA       NA        NA       NA       NA        NA
#> Track_15       NA       NA       NA        NA       NA       NA        NA
#> Track_16       NA       NA       NA        NA       NA       NA        NA
#> Track_18       NA       NA       NA        NA       NA       NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 77.30423 13.465470 25.772071 66.542706
#> Track_02 79.34939 52.519718 32.934325 66.586674
#> Track_03 36.24055 44.481422 23.033620 28.780379
#> Track_04 51.02331 83.205242 54.264162 44.394466
#> Track_07 46.89652  3.890143  9.350172 37.168925
#> Track_08 31.25231 18.897142 12.971107 25.037086
#> Track_09 36.67864 55.736311 34.208609 30.219171
#> Track_13       NA 49.114210 28.259617  2.794258
#> Track_15       NA        NA  9.355554 38.367585
#> Track_16       NA        NA        NA 21.009183
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.322865 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08   Track_09
#> Track_01       NA 5.550484 13.04124 101.81798 76.09181 56.96874  64.396941
#> Track_02       NA       NA 12.08804  88.85917 65.15921 50.27125  59.050564
#> Track_03       NA       NA       NA  69.68828 41.47140 26.29408  33.405277
#> Track_04       NA       NA       NA        NA 34.85453 82.51927 106.013974
#> Track_07       NA       NA       NA        NA       NA 37.72494  55.446497
#> Track_08       NA       NA       NA        NA       NA       NA   7.766668
#> Track_09       NA       NA       NA        NA       NA       NA         NA
#> Track_13       NA       NA       NA        NA       NA       NA         NA
#> Track_15       NA       NA       NA        NA       NA       NA         NA
#> Track_16       NA       NA       NA        NA       NA       NA         NA
#> Track_18       NA       NA       NA        NA       NA       NA         NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 93.14652 78.09344 27.61268 76.070104
#> Track_02 81.33934 67.20656 23.05107 65.043683
#> Track_03 53.23222 40.21428  8.75038 41.046040
#> Track_04 75.09618 81.64188 52.30957 35.174121
#> Track_07 24.61515 32.51964 23.64092  1.106286
#> Track_08 35.14964 18.31166 12.82275 37.150622
#> Track_09 53.48461 30.85768 20.11297 54.990316
#> Track_13       NA 20.39980 30.86467 23.972188
#> Track_15       NA       NA 22.01292 31.798580
#> Track_16       NA       NA       NA 23.443411
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.380036 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 4.363377 76.07916 34.08594 74.358591 74.792504 95.12423
#> Track_02       NA       NA 66.42392 28.40465 64.651796 64.924689 83.24433
#> Track_03       NA       NA       NA 73.14578  9.205186  3.612741 17.06426
#> Track_04       NA       NA       NA       NA 67.293274 71.148627 95.35534
#> Track_07       NA       NA       NA       NA        NA 10.048253 27.26887
#> Track_08       NA       NA       NA       NA        NA        NA 14.84788
#> Track_09       NA       NA       NA       NA        NA        NA       NA
#> Track_13       NA       NA       NA       NA        NA        NA       NA
#> Track_15       NA       NA       NA       NA        NA        NA       NA
#> Track_16       NA       NA       NA       NA        NA        NA       NA
#> Track_18       NA       NA       NA       NA        NA        NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 89.35850 64.32650 21.42133 77.303338
#> Track_02 78.83204 54.99016 16.80235 67.539442
#> Track_03 30.54572 21.11957 23.79606 18.709943
#> Track_04 95.15379 51.35334 27.78191 80.662805
#> Track_07 39.40213 14.35476 23.71095 26.328468
#> Track_08 28.02998 20.59588 22.68434 16.354639
#> Track_09 20.34202 39.77376 34.88762  8.249193
#> Track_13       NA 49.12390 31.37901  9.846911
#> Track_15       NA       NA 21.27493 35.358500
#> Track_16       NA       NA       NA 24.366962
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.427006 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 95.10978 14.77717 138.71476 52.63772 24.78273 90.69693
#> Track_02       NA       NA 58.83774  82.61187 68.63877 43.36454 69.82375
#> Track_03       NA       NA       NA  82.95532 22.91524 16.48567 46.80456
#> Track_04       NA       NA       NA        NA 67.76269 82.96450 27.70365
#> Track_07       NA       NA       NA        NA       NA 34.77956 32.57539
#> Track_08       NA       NA       NA        NA       NA       NA 49.80417
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 92.98843 76.020114 50.62783 75.681715
#> Track_02 31.74766 18.919014 41.55815 61.833016
#> Track_03 51.63716 39.399303 20.00769 35.871009
#> Track_04 38.97117 44.970750 34.13209 33.338638
#> Track_07 51.57682 42.298667 13.85482 21.817867
#> Track_08 46.02738 33.235321 22.86869 41.383123
#> Track_09 38.45780 37.701993 11.52964  7.508252
#> Track_13       NA  8.955358 24.46793 35.911092
#> Track_15       NA        NA 19.92977 32.673570
#> Track_16       NA        NA       NA  7.044453
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.473947 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 104.5423 77.38901 74.40663 32.46911 10.30369 76.08784
#> Track_02       NA       NA 53.47219 64.20173 47.64263 64.02520 75.79721
#> Track_03       NA       NA       NA 87.48810 44.42565 41.86456 20.02698
#> Track_04       NA       NA       NA       NA 21.66254 50.57241 96.51976
#> Track_07       NA       NA       NA       NA       NA 18.31010 48.30507
#> Track_08       NA       NA       NA       NA       NA       NA 41.48235
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 91.67238 23.423172 35.091859 74.973597
#> Track_02 50.63317 54.967838 29.815631 56.507359
#> Track_03 10.79171 45.327867 23.185836  5.384576
#> Track_04 95.08799 31.343061 29.360014 86.776837
#> Track_07 52.65720  6.485488  8.803811 42.648902
#> Track_08 53.02233 13.275668 15.364663 38.915758
#> Track_09 33.79730 47.644308 32.104718 14.460476
#> Track_13       NA 54.669453 28.856381 16.036209
#> Track_15       NA        NA 11.303356 42.963925
#> Track_16       NA        NA        NA 22.504490
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.531878 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 58.19854 70.60093 47.78880  8.37460 47.08483 18.99238
#> Track_02       NA       NA 68.28150 20.05497 37.92019 61.59736 54.41424
#> Track_03       NA       NA       NA 82.41631 37.10689 18.46648 41.89334
#> Track_04       NA       NA       NA       NA 34.16296 68.71280 51.29356
#> Track_07       NA       NA       NA       NA       NA 23.89422 10.12326
#> Track_08       NA       NA       NA       NA       NA       NA 22.89900
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>           Track_13  Track_15 Track_16  Track_18
#> Track_01 63.396827 41.631697 46.16290 59.065825
#> Track_02 77.643673 61.093974 25.20675 66.348552
#> Track_03 17.795817 24.835263 22.72206 10.189205
#> Track_04 86.770362 66.943511 39.10840 76.688071
#> Track_07 35.187608 21.496768 20.71188 30.375461
#> Track_08  7.396134  6.567391 22.96826  9.509602
#> Track_09 35.341794 18.646425 31.22105 32.907442
#> Track_13        NA 14.361851 31.10115  5.944212
#> Track_15        NA        NA 24.89614 15.382132
#> Track_16        NA        NA       NA 22.910910
#> Track_18        NA        NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.579118 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 68.04439 75.59562 110.42685 78.36778 19.55401 51.85580
#> Track_02       NA       NA 40.40994 121.54244 64.92945 50.58624 80.55552
#> Track_03       NA       NA       NA  73.68433 24.14964 40.65117 55.28068
#> Track_04       NA       NA       NA        NA 42.45540 58.72591 44.39820
#> Track_07       NA       NA       NA        NA       NA 40.10151 46.66892
#> Track_08       NA       NA       NA        NA       NA       NA 20.86853
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15 Track_16  Track_18
#> Track_01 62.131333 77.53583 40.27206 75.826121
#> Track_02  8.244845 68.91922 43.27173 43.954978
#> Track_03 28.197936 31.74411 22.40501  4.527874
#> Track_04 96.946052 27.34194 30.31968 68.070655
#> Track_07 49.324963 10.45467 18.31394 19.821015
#> Track_08 40.868482 37.98965 13.63118 39.940597
#> Track_09 64.213193 39.93154 13.04887 53.948075
#> Track_13        NA 54.66377 30.90360 30.803624
#> Track_15        NA       NA 16.35608 27.748796
#> Track_16        NA       NA       NA 21.049879
#> Track_18        NA       NA       NA        NA
#> ------------------------------------
#> 2025-04-16 07:49:43.626211 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 54.79545 25.54294 44.12114 72.02948 73.53451 67.56060
#> Track_02       NA       NA 49.23022 20.03424 44.47405 68.52744 29.10902
#> Track_03       NA       NA       NA 49.29517 41.01074 34.12264 46.34650
#> Track_04       NA       NA       NA       NA 61.83020 81.98838 47.45462
#> Track_07       NA       NA       NA       NA       NA 28.70211 18.25098
#> Track_08       NA       NA       NA       NA       NA       NA 45.62091
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 59.39630 12.917589 52.029902 25.256355
#> Track_02 17.94248 42.898387 43.195193 49.404699
#> Track_03 42.75277  9.543386 21.466898  1.366888
#> Track_04 35.27529 39.525187 54.585278 49.309264
#> Track_07 24.79204 43.156891 15.149349 41.481825
#> Track_08 48.52180 40.495082  8.288692 35.135165
#> Track_09 10.91292 45.073843 27.197890 47.163358
#> Track_13       NA 40.435013 27.749313 43.236163
#> Track_15       NA        NA 24.721410  9.113895
#> Track_16       NA        NA        NA 22.008800
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 8.365556 8.140001 13.288601 17.963246 14.043056 80.66455
#> Track_02       NA       NA 9.928547 13.870318 15.997694 11.986057 83.72110
#> Track_03       NA       NA       NA  4.242441  9.250629  6.131319 46.49768
#> Track_04       NA       NA       NA        NA  5.401950  3.172849 42.61717
#> Track_07       NA       NA       NA        NA        NA  5.261381 51.73148
#> Track_08       NA       NA       NA        NA        NA        NA 54.42241
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.00788 18.730370 14.932046 22.179224
#> Track_02 47.19796 17.799527  9.836672 20.569492
#> Track_03 20.29566  7.910351 10.158387  9.928217
#> Track_04 19.90114  4.165947  8.969975  6.396727
#> Track_07 29.44950  2.944420  6.952380  5.167036
#> Track_08 27.32867  4.915326  5.349900  6.863688
#> Track_09 22.41087 41.825031 68.973412 42.969072
#> Track_13       NA 20.579060 39.239001 21.761749
#> Track_15       NA        NA  8.572483  2.784293
#> Track_16       NA        NA        NA 10.795457
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA      0.2        0        0      0.2      0.1      0.7
#> Track_02       NA       NA        0        0      0.0      0.0      1.0
#> Track_03       NA       NA       NA        0      0.1      0.1      0.6
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.0      0.8
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.2      0.2      0.0      0.0
#> Track_02      0.3      0.1      0.0      0.1
#> Track_03      0.2      0.1      0.1      0.5
#> Track_04      0.0      0.0      0.0      0.0
#> Track_07      0.2      0.0      0.0      0.1
#> Track_08      0.1      0.0      0.0      0.0
#> Track_09      0.2      0.5      1.0      0.5
#> Track_13       NA      0.3      1.0      0.4
#> Track_15       NA       NA      0.0      0.0
#> Track_16       NA       NA       NA      0.1
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#>