Similarity metric using Dynamic Time Warping (DTW)

simil_DTW_metric() computes similarity metrics between two or more trajectories using Dynamic Time Warping (DTW). It allows for different superposition methods to align trajectories before calculating the DTW metric. The function also supports testing with simulations to calculate p-values for the DTW distance metrics.

Usage

simil_DTW_metric(data, test = NULL, sim = NULL, superposition = NULL)

Arguments

data

A track R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

test

Logical; if TRUE, the function compares the observed DTW distances against simulated trajectories and calculates p-values. Default is FALSE.

sim

A track simulation R object consisting of a list of simulated trajectories to use for comparison when test = TRUE.

superposition

A character string indicating the method used to align trajectories. Options are "None", "Centroid", or "Origin". Default is "None".

Value

A track similarity R object consisting ofa list containing the following elements:

DTW_distance_metric

A matrix containing the pairwise DTW distances between trajectories.

DTW_distance_metric_p_values

(If test is TRUE) A matrix containing the p-values for the pairwise DTW distances.

DTW_metric_p_values_combined

(If test is TRUE) The overall p-value for the combined DTW distances.

DTW_distance_metric_simulations

(If test is TRUE) A list of DTW distance matrices from each simulated dataset.

Details

The simil_DTW_metric() function calculates the similarity between trajectories using the Dynamic Time Warping (DTW) algorithm from the dtw package. The dtw() function is used with the dist.method argument set to "Euclidean" for computing the local distances between points in the trajectories.

DTW aligns two time series by minimizing the cumulative distance between their points, creating an optimal alignment despite variations in length or temporal distortions. The algorithm constructs a distance matrix where each element represents the cost of aligning points between the two series and finds a warping path through this matrix that minimizes the total distance. The warping path is contiguous and monotonic, starting from the bottom-left corner and ending at the top-right corner (Cleasby et al., 2019).

DTW measures are non-negative and unbounded, with larger values indicating greater dissimilarity between the time series. This method has been used in various contexts, including ecological studies to analyze and cluster trajectory data (Cleasby et al., 2019).

Potential limitations and biases of DTW include sensitivity to noise and outliers, computational complexity, and the need for appropriate distance metrics. Additionally, DTW may not always account for all structural differences between trajectories and can be biased by the chosen alignment constraints. While DTW can handle trajectories of different lengths due to its elastic nature, having trajectories of similar lengths can improve the accuracy and interpretability of the similarity measure. Similar lengths result in a more meaningful alignment and can make the computation more efficient. When trajectories differ significantly in length, preprocessing or normalization might be necessary, and careful analysis is required to understand the alignment path. The function’s flexibility in handling different lengths allows it to be applied in various contexts. However, large differences in trajectory lengths might introduce potential biases that should be considered when interpreting the results.

The function offers three different superposition methods to align the trajectories before DTW() is applied:

"None": No superposition is applied.

"Centroid": Trajectories are shifted to align based on their centroids.

"Origin": Trajectories are shifted to align based on their starting point.

If test = TRUE, the function can compute p-values by comparing the observed DTW distances with those generated from a set of simulated trajectories. The p-values are calculated for both individual trajectory pairs and for the entire set of trajectories.

Logo

References

Cleasby, I. R., Wakefield, E. D., Morrissey, B. J., Bodey, T. W., Votier, S. C., Bearhop, S., & Hamer, K. C. (2019). Using time-series similarity measures to compare animal movement trajectories in ecology. Behavioral Ecology and Sociobiology, 73, 1-19.

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Simulating tracks using the "Directed" model and comparing DTW distance
# in the PaluxyRiver dataset
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-13 21:08:51.59223 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 19.28826
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.594025 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 16.31147
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.595697 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 19.76826
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.597393 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 21.58083
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.599078 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 14.65521
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.600887 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 17.88554
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.60253 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 48.56188
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.604187 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 23.02089
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.605829 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 23.08311
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.60746 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 25.24207
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.8
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.8
#> 

# Example 2: Simulating tracks using the "Constrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
#> 2025-05-13 21:08:51.643805 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 40.61314
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.645464 Iteration 2
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 122.509
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:51.647106 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 75.72611
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.648699 Iteration 4
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 134.359
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:51.650344 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 66.07477
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.65203 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 115.0806
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.653678 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 122.8675
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.655308 Iteration 8
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA   91.38
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-05-13 21:08:51.656983 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 68.09007
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.658621 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 79.40186
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 3: Simulating tracks using the "Unconstrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
#> 2025-05-13 21:08:51.687928 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 469.6693
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.689581 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 237.1966
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.691534 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 587.3318
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.693271 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 694.4483
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.694931 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 242.5454
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.696679 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 429.9336
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.698405 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 759.0527
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.700135 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 395.7381
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.701794 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 182.4649
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:51.703429 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 784.8165
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 4: Simulating and comparing DTW distance in the MountTom dataset using the
# "Centroid" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
#> 2025-05-13 21:08:51.938228 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 52.29743 34.22329 24.88062 19.377209 40.38798 44.571112
#> Track_02       NA       NA  8.50326 40.21892 23.660082 23.63879 33.197106
#> Track_03       NA       NA       NA 24.43988 11.407553 16.06984 22.894395
#> Track_04       NA       NA       NA       NA  7.369931 41.51480 48.530823
#> Track_07       NA       NA       NA       NA        NA 21.21279 25.889720
#> Track_08       NA       NA       NA       NA        NA       NA  7.506039
#> Track_09       NA       NA       NA       NA        NA       NA        NA
#> Track_13       NA       NA       NA       NA        NA       NA        NA
#> Track_15       NA       NA       NA       NA        NA       NA        NA
#> Track_16       NA       NA       NA       NA        NA       NA        NA
#> Track_18       NA       NA       NA       NA        NA       NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 47.58287 36.551550 26.152500 30.947048
#> Track_02 13.47377  5.843377 24.160035 11.227210
#> Track_03 13.39272  2.234708 12.919406  2.354313
#> Track_04 43.34044 27.698616 30.022955 21.130055
#> Track_07 23.45609 13.154052 13.443442  9.426562
#> Track_08 12.63391 15.072900  8.066920 16.156120
#> Track_09 20.56580 22.176158  7.085655 22.558239
#> Track_13       NA 11.529384 16.085125 14.353252
#> Track_15       NA        NA 12.977623  4.249908
#> Track_16       NA        NA        NA 12.262202
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:51.991417 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 54.42915 20.99616  7.950864 30.94951 12.268992 27.01431
#> Track_02       NA       NA 34.46177 52.187908 14.60502 28.462330 41.72536
#> Track_03       NA       NA       NA 24.438691 19.85870 13.769213  4.08444
#> Track_04       NA       NA       NA        NA 28.54393  9.513043 30.87830
#> Track_07       NA       NA       NA        NA       NA 11.739231 25.87432
#> Track_08       NA       NA       NA        NA       NA        NA 19.63146
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 46.92565 31.409547 17.036107 23.897509
#> Track_02 26.45379 36.085827 24.651072 34.617863
#> Track_03 21.55187  7.214357  5.276073  2.335948
#> Track_04 47.91648 35.061895 19.199775 27.609597
#> Track_07 23.03503 23.766023 12.334319 20.475065
#> Track_08 26.56788 19.358458  8.881064 15.249406
#> Track_09 25.29822  6.661962  7.897354  3.904970
#> Track_13       NA 19.303426 16.293746 20.305763
#> Track_15       NA        NA  8.916419  5.215535
#> Track_16       NA        NA        NA  5.892352
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.038695 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 13.42674 40.02687 58.774204 40.45305 14.06757 22.14156
#> Track_02       NA       NA 30.49280 45.052142 34.80038  7.47002 24.82966
#> Track_03       NA       NA       NA  9.672435 16.32140 16.67800 27.66053
#> Track_04       NA       NA       NA        NA 35.36977 29.71183 49.46793
#> Track_07       NA       NA       NA        NA       NA 20.88123 21.64853
#> Track_08       NA       NA       NA        NA       NA       NA 18.04136
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01  9.871386 41.34971 23.200906 40.447116
#> Track_02  5.124726 35.67117 16.117020 31.424436
#> Track_03 23.297987 17.11178  8.470175  3.492537
#> Track_04 38.945843 36.68892 16.795781 14.373982
#> Track_07 26.173903  1.56601 13.129887 13.593798
#> Track_08  6.507983 21.60486  7.605146 17.340805
#> Track_09 18.409913 22.05436 17.801734 26.609732
#> Track_13        NA 26.97560 12.020625 23.806576
#> Track_15        NA       NA 13.893406 14.403421
#> Track_16        NA       NA        NA  8.762511
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.086192 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 48.57413 40.16938 41.94914 41.266987 34.72663 50.51976
#> Track_02       NA       NA 14.30285 63.57891 25.959882 34.42228 33.67936
#> Track_03       NA       NA       NA 43.13868  9.797489 17.57498 15.12203
#> Track_04       NA       NA       NA       NA 36.321467 17.63013 39.35744
#> Track_07       NA       NA       NA       NA        NA 11.78100  4.84876
#> Track_08       NA       NA       NA       NA        NA       NA 13.48995
#> Track_09       NA       NA       NA       NA        NA       NA       NA
#> Track_13       NA       NA       NA       NA        NA       NA       NA
#> Track_15       NA       NA       NA       NA        NA       NA       NA
#> Track_16       NA       NA       NA       NA        NA       NA       NA
#> Track_18       NA       NA       NA       NA        NA       NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 47.007073 39.834202 29.179748 38.119138
#> Track_02 32.554135 36.170821 16.629877 11.287235
#> Track_03 14.717089 18.862049  6.167541  2.600206
#> Track_04 37.355283 22.602311 29.233126 42.816510
#> Track_07  5.861924 11.570016  6.383158 11.072449
#> Track_08 12.422464  4.201333  9.943569 17.883413
#> Track_09  2.929183 12.478134  9.591047 16.739643
#> Track_13        NA 11.504673  9.237181 16.129235
#> Track_15        NA        NA 10.991478 19.344359
#> Track_16        NA        NA        NA  6.281953
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.138131 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 60.46438 10.85509 39.62488 15.28821 39.81639 22.381064
#> Track_02       NA       NA 37.01023 52.58779 36.16609 10.57009 40.403859
#> Track_03       NA       NA       NA 33.46132 12.02146 20.69773  9.009762
#> Track_04       NA       NA       NA       NA 15.95287 40.42027 44.665643
#> Track_07       NA       NA       NA       NA       NA 21.59176 20.059380
#> Track_08       NA       NA       NA       NA       NA       NA 22.395371
#> Track_09       NA       NA       NA       NA       NA       NA        NA
#> Track_13       NA       NA       NA       NA       NA       NA        NA
#> Track_15       NA       NA       NA       NA       NA       NA        NA
#> Track_16       NA       NA       NA       NA       NA       NA        NA
#> Track_18       NA       NA       NA       NA       NA       NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 45.188425 21.163262 13.689194  7.108107
#> Track_02 15.940310 36.202227 25.875992 35.782216
#> Track_03 24.603056  7.284913  6.392033  4.175312
#> Track_04 48.314875 40.484448 22.874762 27.576783
#> Track_07 27.508281 16.577987  7.848784  8.165043
#> Track_08  7.215607 19.162792 12.886228 19.809673
#> Track_09 24.801958  3.558880 10.064428 11.477482
#> Track_13        NA 21.740315 17.086110 24.371802
#> Track_15        NA        NA  8.988482  9.890770
#> Track_16        NA        NA        NA  4.327760
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.184845 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA  46.8945 27.32468 25.29508 21.12641 39.90592 40.027664
#> Track_02       NA       NA 35.80738 59.84393 16.38389 22.41563  5.813055
#> Track_03       NA       NA       NA 16.11522 19.12623 16.77899 27.185995
#> Track_04       NA       NA       NA       NA 33.64784 41.21160 50.007931
#> Track_07       NA       NA       NA       NA       NA 17.64911 13.239284
#> Track_08       NA       NA       NA       NA       NA       NA 16.182574
#> Track_09       NA       NA       NA       NA       NA       NA        NA
#> Track_13       NA       NA       NA       NA       NA       NA        NA
#> Track_15       NA       NA       NA       NA       NA       NA        NA
#> Track_16       NA       NA       NA       NA       NA       NA        NA
#> Track_18       NA       NA       NA       NA       NA       NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 13.097270 30.025897 14.827264 33.904483
#> Track_02 28.020827  9.233240 23.668136 35.211534
#> Track_03 21.831937 21.806703  8.082925  5.574797
#> Track_04 30.316444 40.639639 16.540052 25.196772
#> Track_07  9.628326  6.164450  9.631851 20.505200
#> Track_08 25.754844 16.416743 12.522193 14.006473
#> Track_09 23.508469  8.325869 17.453897 26.770525
#> Track_13        NA 15.282953 10.866405 25.058713
#> Track_15        NA        NA 12.181361 22.398020
#> Track_16        NA        NA        NA  9.931314
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.231925 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 56.74348 40.26262 43.59504 11.02055  6.219828 40.85591
#> Track_02       NA       NA 23.64601 57.30841 32.39164 32.927289 36.67758
#> Track_03       NA       NA       NA 25.16712 22.46255 19.789424 10.45221
#> Track_04       NA       NA       NA       NA 34.17324 27.797554 14.74830
#> Track_07       NA       NA       NA       NA       NA  4.899418 25.82916
#> Track_08       NA       NA       NA       NA       NA        NA 22.22236
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 47.342301 11.986856 15.169354 11.772405
#> Track_02  9.894926 32.783791 24.937637 34.877106
#> Track_03 13.911131 23.375953 12.298703 18.673268
#> Track_04 43.648970 36.008497 19.657373 20.869271
#> Track_07 26.628472  1.419586  8.299171  9.576154
#> Track_08 25.649533  5.804989  5.371083  5.082530
#> Track_09 25.142628 26.747598 14.019272 18.767304
#> Track_13        NA 27.195145 17.820081 26.571928
#> Track_15        NA        NA  9.027362 10.568255
#> Track_16        NA        NA        NA  4.501208
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.278908 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 58.11606 10.12294 41.72810 26.89893  8.870986 10.003953
#> Track_02       NA       NA 33.63159 55.43841 27.37756 33.789934 44.930769
#> Track_03       NA       NA       NA 34.19011 10.78448  7.240144 11.669663
#> Track_04       NA       NA       NA       NA 44.06099 23.028328 26.150513
#> Track_07       NA       NA       NA       NA       NA 15.258710 23.165811
#> Track_08       NA       NA       NA       NA       NA        NA  3.369376
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 48.171126 24.971505 13.827950 39.062189
#> Track_02  5.155883 30.918710 24.336335  9.613968
#> Track_03 27.311459  9.547052  5.247110 18.972972
#> Track_04 47.789211 44.685965 25.271698 42.774448
#> Track_07 22.160440  3.554969  9.058772 13.372286
#> Track_08 26.823503 14.971466  5.782505 20.135269
#> Track_09 36.591457 22.631449 10.209798 29.379226
#> Track_13        NA 25.108590 18.883775  8.245672
#> Track_15        NA        NA  9.501980 16.016828
#> Track_16        NA        NA        NA 12.439280
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.333748 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA  26.7255 37.21829 64.83891 28.72747 33.616001 47.390136
#> Track_02       NA       NA 35.59734 47.22288 12.35413 18.795602 34.960653
#> Track_03       NA       NA       NA 29.84946 21.12356 18.920992 18.004816
#> Track_04       NA       NA       NA       NA 25.89875 17.670376  7.020958
#> Track_07       NA       NA       NA       NA       NA  5.374643 17.812584
#> Track_08       NA       NA       NA       NA       NA        NA 12.544323
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16   Track_18
#> Track_01 14.825219 13.361595 14.868869 33.5024767
#> Track_02  9.343087  9.788614 12.574669 18.5857610
#> Track_03 27.115956 21.344238 12.603673 19.0267151
#> Track_04 44.473617 38.568215 27.409701 17.9036033
#> Track_07 13.932158 11.343473  8.671995  5.2191705
#> Track_08 18.005592 14.590347  9.702652  0.4060508
#> Track_09 30.420145 24.785139 17.059132 12.7074759
#> Track_13        NA  5.465146  7.808402 17.8783774
#> Track_15        NA        NA  5.831418 14.4510387
#> Track_16        NA        NA        NA  9.7102303
#> Track_18        NA        NA        NA         NA
#> ------------------------------------
#> 2025-05-13 21:08:52.380993 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 19.29643 15.604882  9.811651 32.93203 37.142545 11.35223
#> Track_02       NA       NA  4.879246 14.615872 33.17055 33.772791 20.53222
#> Track_03       NA       NA        NA 12.971749 20.91659 20.887834 14.59974
#> Track_04       NA       NA        NA        NA 37.30989 40.181653 16.17114
#> Track_07       NA       NA        NA        NA       NA  7.599156 18.52740
#> Track_08       NA       NA        NA        NA       NA        NA 22.32826
#> Track_09       NA       NA        NA        NA       NA        NA       NA
#> Track_13       NA       NA        NA        NA       NA        NA       NA
#> Track_15       NA       NA        NA        NA       NA        NA       NA
#> Track_16       NA       NA        NA        NA       NA        NA       NA
#> Track_18       NA       NA        NA        NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 44.710642 25.297472 19.676193 21.378985
#> Track_02 39.543684 29.257527 12.396628 25.793742
#> Track_03 25.846953 18.610053  6.531643 15.838712
#> Track_04 47.527453 30.628560 20.504835 26.248501
#> Track_07 12.445198  6.838247 13.543000  7.892867
#> Track_08  6.441817 12.298440 12.552552 12.553477
#> Track_09 27.843960 13.058691 13.830818 10.312078
#> Track_13        NA 17.869763 16.595107 17.593420
#> Track_15        NA        NA 12.935230  2.955873
#> Track_16        NA        NA        NA 11.298182
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 7.155827 6.889059 7.146449 8.550492 9.358030 41.27125
#> Track_02       NA       NA 7.367287 7.095034 8.511176 8.404712 44.46323
#> Track_03       NA       NA       NA 2.752378 5.017897 5.558007 23.61988
#> Track_04       NA       NA       NA       NA 3.748965 5.898481 21.55969
#> Track_07       NA       NA       NA       NA       NA 3.362680 27.87452
#> Track_08       NA       NA       NA       NA       NA       NA 27.11038
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 23.69080 10.006241  8.956456 11.561263
#> Track_02 26.76916  8.652232  8.004512  9.939214
#> Track_03 11.38055  4.011654  6.187040  4.835886
#> Track_04 10.75445  2.107453  7.121107  3.128265
#> Track_07 16.24588  4.076379  5.378815  5.173189
#> Track_08 16.01165  5.869619  3.814577  6.169448
#> Track_09 12.44577 21.459210 36.643358 21.460416
#> Track_13       NA 11.406049 22.371414 11.843307
#> Track_15       NA        NA  7.547234  1.280341
#> Track_16       NA        NA        NA  8.182026
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0      0.0        0        0      0.2      0.7
#> Track_02       NA       NA      0.1        0        0      0.1      0.9
#> Track_03       NA       NA       NA        0        0      0.0      0.8
#> Track_04       NA       NA       NA       NA        0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.0      1.0
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.3      0.0      0.0      0.1
#> Track_02      0.7      0.1      0.0      0.1
#> Track_03      0.0      0.1      0.3      0.5
#> Track_04      0.0      0.0      0.0      0.0
#> Track_07      0.4      0.3      0.0      0.0
#> Track_08      0.5      0.2      0.0      0.2
#> Track_09      0.1      0.5      1.0      0.6
#> Track_13       NA      0.1      1.0      0.1
#> Track_15       NA       NA      0.1      0.0
#> Track_16       NA       NA       NA      0.4
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 5: Simulating and comparing DTW distance in the MountTom dataset using the
# "Origin" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
#> 2025-05-13 21:08:52.645277 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 114.2689 77.68143 131.65395 33.53321 54.92854  8.75146
#> Track_02       NA       NA 46.62520  10.83419 59.43374 37.32212 81.97472
#> Track_03       NA       NA       NA  48.96827 45.65576 38.97382 50.78574
#> Track_04       NA       NA       NA        NA 72.93743 49.56925 96.57571
#> Track_07       NA       NA       NA        NA       NA 16.10334 21.64385
#> Track_08       NA       NA       NA        NA       NA       NA 35.50542
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15  Track_16 Track_18
#> Track_01 91.780510 75.52442 35.483548 32.18992
#> Track_02 11.590124 11.86530 35.878150 68.91937
#> Track_03 30.001281 33.30282 23.762171 31.57764
#> Track_04  8.278468 22.15607 46.165911 80.47733
#> Track_07 47.962756 33.49308  9.567524 29.65689
#> Track_08 31.997424 18.18482  8.442258 35.30693
#> Track_09 62.868255 49.99391 19.962398 21.11773
#> Track_13        NA 14.70555 26.341892 50.26167
#> Track_15        NA       NA 17.456837 41.47944
#> Track_16        NA       NA        NA 18.80497
#> Track_18        NA       NA        NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:52.691841 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 115.7204 13.23092  29.03928 78.60993 52.25967 90.90574
#> Track_02       NA       NA 66.81801 120.71176 15.99894 58.09380 11.90421
#> Track_03       NA       NA       NA  16.61623 41.93968 22.84293 49.47090
#> Track_04       NA       NA       NA        NA 83.58064 41.94944 96.82534
#> Track_07       NA       NA       NA        NA       NA 31.25564 18.77690
#> Track_08       NA       NA       NA        NA       NA       NA 45.96641
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15 Track_16  Track_18
#> Track_01 91.292372 60.80645 30.73481 28.110600
#> Track_02  8.060873 44.32817 44.41314 66.488390
#> Track_03 52.033596 33.37028  9.73242  9.803602
#> Track_04 97.775589 74.07527 27.63380 15.135020
#> Track_07  8.965184 34.29438 23.37566 40.166844
#> Track_08 42.607408 40.54729  9.92655 15.395482
#> Track_09 12.818422 27.81189 33.04930 49.911473
#> Track_13        NA 36.09942 31.09877 51.092333
#> Track_15        NA       NA 23.00324 38.226058
#> Track_16        NA       NA       NA  6.597684
#> Track_18        NA       NA       NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.740018 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03   Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 113.8404 69.35117   8.138557 23.11485 69.65884  99.64524
#> Track_02       NA       NA 31.76874 114.809127 63.26452 40.43533  25.25371
#> Track_03       NA       NA       NA  69.096315 29.88269 35.97931  39.20581
#> Track_04       NA       NA       NA         NA 21.90387 72.33269 101.48425
#> Track_07       NA       NA       NA         NA       NA 40.01948  55.89072
#> Track_08       NA       NA       NA         NA       NA       NA  20.93814
#> Track_09       NA       NA       NA         NA       NA       NA        NA
#> Track_13       NA       NA       NA         NA       NA       NA        NA
#> Track_15       NA       NA       NA         NA       NA       NA        NA
#> Track_16       NA       NA       NA         NA       NA       NA        NA
#> Track_18       NA       NA       NA         NA       NA       NA        NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 62.29275 30.08910 39.75060 38.400916
#> Track_02 67.09549 64.71639 34.43144 56.607155
#> Track_03 51.23821 42.05538 11.05334 21.832199
#> Track_04 65.70944 32.41274 40.68150 37.620642
#> Track_07 43.78871 25.33581 12.95899 10.496526
#> Track_08 19.04111 28.69217 22.34260 39.781705
#> Track_09 45.07295 51.38189 31.89808 53.733655
#> Track_13       NA 21.03313 30.49216 47.402844
#> Track_15       NA       NA 21.72359 31.259076
#> Track_16       NA       NA       NA  8.123888
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.792948 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 15.65781 37.67941 67.81142 70.53184 63.730688 91.79658
#> Track_02       NA       NA 39.37413 73.26639 56.57239 50.692582 82.66130
#> Track_03       NA       NA       NA 13.09126 44.15417 39.556425 40.38893
#> Track_04       NA       NA       NA       NA 82.19651 79.383646 67.47949
#> Track_07       NA       NA       NA       NA       NA  4.571878 45.85422
#> Track_08       NA       NA       NA       NA       NA        NA 43.86123
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16 Track_18
#> Track_01 14.417224 50.311793 46.826120 43.48258
#> Track_02  5.790812 49.722565 37.332140 31.41345
#> Track_03 29.871939  9.203745 23.881465 34.80321
#> Track_04 60.734656 13.919821 56.251689 72.45042
#> Track_07 42.926492 44.545627  8.718657 18.49541
#> Track_08 37.584478 39.224251  5.541506 15.31560
#> Track_09 62.098836 32.661230 30.283129 49.76268
#> Track_13        NA 37.195265 25.883149 23.06009
#> Track_15        NA        NA 23.969740 37.01426
#> Track_16        NA        NA        NA 10.76269
#> Track_18        NA        NA        NA       NA
#> ------------------------------------
#> 2025-05-13 21:08:52.839638 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07 Track_08 Track_09
#> Track_01       NA   38.514 75.64546 128.77552 76.531880 20.69261 88.55435
#> Track_02       NA       NA 62.04600 101.39480 63.207411 37.26777 83.29304
#> Track_03       NA       NA       NA  17.12432  2.513166 40.41750 29.10496
#> Track_04       NA       NA       NA        NA 15.307569 82.40129 64.51662
#> Track_07       NA       NA       NA        NA        NA 41.77055 31.10453
#> Track_08       NA       NA       NA        NA        NA       NA 43.85334
#> Track_09       NA       NA       NA        NA        NA       NA       NA
#> Track_13       NA       NA       NA        NA        NA       NA       NA
#> Track_15       NA       NA       NA        NA        NA       NA       NA
#> Track_16       NA       NA       NA        NA        NA       NA       NA
#> Track_18       NA       NA       NA        NA        NA       NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 89.47330 14.42035 49.86590 62.403407
#> Track_02 79.73875 16.88877 35.41482 37.876947
#> Track_03 16.78949 43.13912 11.36156 21.812150
#> Track_04 45.33599 81.83315 28.56407 36.521569
#> Track_07 18.82133 44.73233 11.94804 21.870975
#> Track_08 48.14274 16.33351 23.16180 36.300051
#> Track_09 17.22357 52.39282 29.11035 46.638481
#> Track_13       NA 55.15500 22.66349 38.849466
#> Track_15       NA       NA 22.70665 30.877034
#> Track_16       NA       NA       NA  7.694175
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.886075 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 111.0486 47.06678 55.67551 15.04648 72.23412 17.331641
#> Track_02       NA       NA 45.71063 92.92195 67.83854 11.27907 84.468218
#> Track_03       NA       NA       NA 20.02838 29.67905 23.18037 39.035687
#> Track_04       NA       NA       NA       NA 47.09986 57.63995 56.936726
#> Track_07       NA       NA       NA       NA       NA 40.31457  3.924873
#> Track_08       NA       NA       NA       NA       NA       NA 51.353686
#> Track_09       NA       NA       NA       NA       NA       NA        NA
#> Track_13       NA       NA       NA       NA       NA       NA        NA
#> Track_15       NA       NA       NA       NA       NA       NA        NA
#> Track_16       NA       NA       NA       NA       NA       NA        NA
#> Track_18       NA       NA       NA       NA       NA       NA        NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 30.66015 73.721315 31.78491 42.872356
#> Track_02 68.68578 10.420537 44.98438 66.163206
#> Track_03 16.33194 24.713969 21.48553 38.337485
#> Track_04 15.06560 60.648779 47.40184 69.395494
#> Track_07 24.20749 42.064458 11.08610 17.784763
#> Track_08 39.20715  3.017993 23.21403 40.266998
#> Track_09 31.37366 52.684070 15.24706 21.900795
#> Track_13       NA 41.206240 24.09047 39.573440
#> Track_15       NA        NA 24.52980 41.323615
#> Track_16       NA        NA       NA  7.012941
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.932918 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 29.82583 76.98702 134.47100 68.36885 21.736768 88.64141
#> Track_02       NA       NA 58.29584 119.52348 45.52279  7.563011 83.20804
#> Track_03       NA       NA       NA  42.06357 14.69168 34.051970 44.04086
#> Track_04       NA       NA       NA        NA 64.01465 81.441142 42.37553
#> Track_07       NA       NA       NA        NA       NA 26.901570 52.48370
#> Track_08       NA       NA       NA        NA       NA        NA 51.50552
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 92.25796 80.339281 40.44897 34.959433
#> Track_02 75.87597 66.043488 37.64560 14.512100
#> Track_03 10.69428  9.104663 22.72932 30.911660
#> Track_04 32.93714 30.951593 46.08669 80.476322
#> Track_07 27.23764 23.185977 24.37902 21.851934
#> Track_08 47.15683 39.416371 18.72602  7.897042
#> Track_09 45.17120 40.340369 18.71467 53.248463
#> Track_13       NA  4.262327 29.20668 44.544511
#> Track_15       NA        NA 23.76367 37.595935
#> Track_16       NA        NA       NA 21.033249
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:52.984879 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 107.8211 33.07008 97.22983 66.26925 63.01656 61.75619
#> Track_02       NA       NA 66.64785 47.26215 62.45045 22.54610 42.54269
#> Track_03       NA       NA       NA 79.15648 22.70205 39.82614 47.95269
#> Track_04       NA       NA       NA       NA 82.71513 14.86199 15.59438
#> Track_07       NA       NA       NA       NA       NA 41.68976 53.09558
#> Track_08       NA       NA       NA       NA       NA       NA 13.14025
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 84.95291 47.768317 52.97349 29.345564
#> Track_02 63.29453 46.214261 19.96747 66.034222
#> Track_03 35.80233 38.011623 23.12401  2.565378
#> Track_04 93.16630 25.532753 39.99441 76.197131
#> Track_07 12.50835 47.792773 20.11360 23.732034
#> Track_08 48.25468 16.424466 13.80262 38.424316
#> Track_09 62.78188  9.147042 25.70417 45.753592
#> Track_13       NA 57.109310 22.82867 36.585914
#> Track_15       NA        NA 21.66140 35.752492
#> Track_16       NA        NA       NA 22.680216
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:53.031383 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 117.2626 51.85489 101.01041 61.29208 57.618953 72.070441
#> Track_02       NA       NA 54.91335  84.74980 49.12284 51.888616 62.964384
#> Track_03       NA       NA       NA  18.47935 41.68044 41.299244  8.933198
#> Track_04       NA       NA       NA        NA 85.46707 85.335866 10.120320
#> Track_07       NA       NA       NA        NA       NA  3.006035 55.975015
#> Track_08       NA       NA       NA        NA       NA        NA 55.860586
#> Track_09       NA       NA       NA        NA       NA        NA        NA
#> Track_13       NA       NA       NA        NA       NA        NA        NA
#> Track_15       NA       NA       NA        NA       NA        NA        NA
#> Track_16       NA       NA       NA        NA       NA        NA        NA
#> Track_18       NA       NA       NA        NA       NA        NA        NA
#>          Track_13 Track_15  Track_16  Track_18
#> Track_01 88.81484 28.93385 39.401990 44.864768
#> Track_02 28.13247 69.64308 37.657579 60.086685
#> Track_03 49.30910 39.52798  6.931271 39.220168
#> Track_04 89.94755 84.04742 28.177643 83.904124
#> Track_07 23.05028 25.58032 23.529341 12.645616
#> Track_08 25.89018 22.94201 23.306964  9.784595
#> Track_09 62.44011 55.29001 13.415334 54.944362
#> Track_13       NA 49.12735 29.780832 34.533647
#> Track_15       NA       NA 23.969561 14.009408
#> Track_16       NA       NA        NA 22.409029
#> Track_18       NA       NA        NA        NA
#> ------------------------------------
#> 2025-05-13 21:08:53.077698 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 110.5276 14.02324  88.50481 49.51397 77.82386  8.357474
#> Track_02       NA       NA 62.34555 116.54855 68.94518 29.34443 77.496324
#> Track_03       NA       NA       NA  63.92138 29.48674 41.95083  6.608122
#> Track_04       NA       NA       NA        NA 15.14804 65.01967 67.459933
#> Track_07       NA       NA       NA        NA       NA 36.15779 33.127856
#> Track_08       NA       NA       NA        NA       NA       NA 50.742475
#> Track_09       NA       NA       NA        NA       NA       NA        NA
#> Track_13       NA       NA       NA        NA       NA       NA        NA
#> Track_15       NA       NA       NA        NA       NA       NA        NA
#> Track_16       NA       NA       NA        NA       NA       NA        NA
#> Track_18       NA       NA       NA        NA       NA       NA        NA
#>          Track_13 Track_15  Track_16  Track_18
#> Track_01 69.98243 52.23824 25.625460 28.871038
#> Track_02 77.08496 41.45517 40.991642 52.361377
#> Track_03 42.99143 23.00143  6.829401  9.988039
#> Track_04 12.71415 82.93520 49.240023 71.553780
#> Track_07 12.49539 44.01702 20.024599 33.793185
#> Track_08 37.13952 37.32351 23.948814 38.408342
#> Track_09 48.73831 33.15364 12.130490 17.649858
#> Track_13       NA 54.76127 28.645055 46.357156
#> Track_15       NA       NA 14.692771 13.273541
#> Track_16       NA       NA        NA  7.858238
#> Track_18       NA       NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 8.365556 8.140001 13.288601 17.963246 14.043056 80.66455
#> Track_02       NA       NA 9.928547 13.870318 15.997694 11.986057 83.72110
#> Track_03       NA       NA       NA  4.242441  9.250629  6.131319 46.49768
#> Track_04       NA       NA       NA        NA  5.401950  3.172849 42.61717
#> Track_07       NA       NA       NA        NA        NA  5.261381 51.73148
#> Track_08       NA       NA       NA        NA        NA        NA 54.42241
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.00788 18.730370 14.932046 22.179224
#> Track_02 47.19796 17.799527  9.836672 20.569492
#> Track_03 20.29566  7.910351 10.158387  9.928217
#> Track_04 19.90114  4.165947  8.969975  6.396727
#> Track_07 29.44950  2.944420  6.952380  5.167036
#> Track_08 27.32867  4.915326  5.349900  6.863688
#> Track_09 22.41087 41.825031 68.973412 42.969072
#> Track_13       NA 20.579060 39.239001 21.761749
#> Track_15       NA        NA  8.572483  2.784293
#> Track_16       NA        NA        NA 10.795457
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0        0      0.1      0.1      0.0      0.5
#> Track_02       NA       NA        0      0.1      0.0      0.2      0.9
#> Track_03       NA       NA       NA      0.0      0.1      0.0      0.7
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.2      0.6
#> Track_08       NA       NA       NA       NA       NA       NA      0.9
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.2      0.1      0.0      0.0
#> Track_02      0.4      0.3      0.0      0.1
#> Track_03      0.3      0.0      0.3      0.2
#> Track_04      0.3      0.0      0.0      0.0
#> Track_07      0.7      0.0      0.0      0.0
#> Track_08      0.2      0.1      0.0      0.0
#> Track_09      0.2      0.5      1.0      0.3
#> Track_13       NA      0.2      1.0      0.0
#> Track_15       NA       NA      0.0      0.0
#> Track_16       NA       NA       NA      0.6
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#>