Similarity metric using Dynamic Time Warping (DTW)

simil_DTW_metric() computes similarity metrics between two or more trajectories using Dynamic Time Warping (DTW). It allows for different superposition methods to align trajectories before calculating the DTW metric. The function also supports testing with simulations to calculate p-values for the DTW distance metrics.

Usage

simil_DTW_metric(data, test = NULL, sim = NULL, superposition = NULL)

Arguments

data

A 'track' R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

test

Logical; if TRUE, the function compares the observed DTW distances against simulated trajectories and calculates p-values. Default is FALSE.

sim

A 'track simulation' R object consisting of a list of simulated trajectories to use for comparison when test = TRUE.

superposition

A character string indicating the method used to align trajectories. Options are "None", "Centroid", or "Origin". Default is "None".

Value

A 'track similarity' R object consisting ofa list containing the following elements:

DTW_distance_metric

A matrix containing the pairwise DTW distances between trajectories.

DTW_distance_metric_p_values

(If test is TRUE) A matrix containing the p-values for the pairwise DTW distances.

DTW_metric_p_values_combined

(If test is TRUE) The overall p-value for the combined DTW distances.

DTW_distance_metric_simulations

(If test is TRUE) A list of DTW distance matrices from each simulated dataset.

Details

The simil_DTW_metric() function calculates the similarity between trajectories using the Dynamic Time Warping (DTW) algorithm from the dtw package. The dtw() function is used with the dist.method argument set to "Euclidean" for computing the local distances between points in the trajectories.

DTW aligns two time series by minimizing the cumulative distance between their points, creating an optimal alignment despite variations in length or temporal distortions. The algorithm constructs a distance matrix where each element represents the cost of aligning points between the two series and finds a warping path through this matrix that minimizes the total distance. The warping path is contiguous and monotonic, starting from the bottom-left corner and ending at the top-right corner (Cleasby et al., 2019).

DTW measures are non-negative and unbounded, with larger values indicating greater dissimilarity between the time series. This method has been used in various contexts, including ecological studies to analyze and cluster trajectory data (Cleasby et al., 2019).

Potential limitations and biases of DTW include sensitivity to noise and outliers, computational complexity, and the need for appropriate distance metrics. Additionally, DTW may not always account for all structural differences between trajectories and can be biased by the chosen alignment constraints. While DTW can handle trajectories of different lengths due to its elastic nature, having trajectories of similar lengths can improve the accuracy and interpretability of the similarity measure. Similar lengths result in a more meaningful alignment and can make the computation more efficient. When trajectories differ significantly in length, preprocessing or normalization might be necessary, and careful analysis is required to understand the alignment path. The function’s flexibility in handling different lengths allows it to be applied in various contexts. However, large differences in trajectory lengths might introduce potential biases that should be considered when interpreting the results.

The function offers three different superposition methods to align the trajectories before DTW() is applied:

"None": No superposition is applied.

"Centroid": Trajectories are shifted to align based on their centroids.

"Origin": Trajectories are shifted to align based on their starting point.

If test = TRUE, the function can compute p-values by comparing the observed DTW distances with those generated from a set of simulated trajectories. The p-values are calculated for both individual trajectory pairs and for the entire set of trajectories.

Logo

References

Cleasby, I. R., Wakefield, E. D., Morrissey, B. J., Bodey, T. W., Votier, S. C., Bearhop, S., & Hamer, K. C. (2019). Using time-series similarity measures to compare animal movement trajectories in ecology. Behavioral Ecology and Sociobiology, 73, 1-19.

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Simulating tracks using the "Directed" model and comparing DTW distance
# in the PaluxyRiver dataset
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-03-27 18:44:51.190914 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 14.49982
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.192612 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 20.17821
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.194215 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 49.41036
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.195772 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 35.14164
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.197341 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 58.87895
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.198922 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 34.73086
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.206779 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 32.34093
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.208995 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 21.62544
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.210565 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 17.21768
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.21214 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 28.82577
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.4
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.4
#> 

# Example 2: Simulating tracks using the "Constrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
#> 2025-03-27 18:44:51.240381 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 126.7234
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.241993 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 50.48409
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.243557 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 94.07194
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.245141 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 114.4203
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.24668 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 59.23749
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.248249 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 107.5631
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.249778 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 94.96769
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.251338 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 178.8354
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.252915 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 153.0599
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.2545 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 131.9621
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 3: Simulating tracks using the "Unconstrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
#> 2025-03-27 18:44:51.283191 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 803.9541
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.284824 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 754.3639
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.28645 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 242.9203
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.288037 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 285.2759
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.289596 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 230.2754
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.291176 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 167.2058
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.292732 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 807.9979
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.294304 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 702.3942
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.295872 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 592.6367
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.297405 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 601.8427
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 4: Simulating and comparing DTW distance in the MountTom dataset using the
# "Centroid" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
#> 2025-03-27 18:44:51.521961 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 26.28116 38.34029 65.769133 41.085670 26.87159 48.447503
#> Track_02       NA       NA 26.78936 49.371890 33.519047 11.63239 34.097731
#> Track_03       NA       NA       NA  7.388485  6.817368 11.91380  3.055763
#> Track_04       NA       NA       NA        NA 10.900722 27.67042  6.453183
#> Track_07       NA       NA       NA        NA        NA 16.61460  9.716442
#> Track_08       NA       NA       NA        NA        NA       NA 16.822403
#> Track_09       NA       NA       NA        NA        NA       NA        NA
#> Track_13       NA       NA       NA        NA        NA       NA        NA
#> Track_15       NA       NA       NA        NA        NA       NA        NA
#> Track_16       NA       NA       NA        NA        NA       NA        NA
#> Track_18       NA       NA       NA        NA        NA       NA        NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 45.187872 21.73538 28.374302 39.897819
#> Track_02 32.894124 30.75162 25.362322 33.142588
#> Track_03  5.656572 23.04371  9.670229  7.798413
#> Track_04  8.452216 43.53851 22.213280 13.385828
#> Track_07  8.942786 23.05589  8.504612  2.425161
#> Track_08 16.085216 21.06490 12.548840 16.608880
#> Track_09  3.501353 30.74066 14.755393 10.935861
#> Track_13        NA 28.66284 13.462035 10.373722
#> Track_15        NA       NA 11.879718 21.259887
#> Track_16        NA       NA        NA  7.063874
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.566652 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 45.35691 18.70362 35.00973 18.12650  8.218420 42.23693
#> Track_02       NA       NA 35.31282 19.87756 18.59857 29.519088 10.52655
#> Track_03       NA       NA       NA 35.00189 16.36396  7.159015 27.36303
#> Track_04       NA       NA       NA       NA 10.76420 26.145021 24.03940
#> Track_07       NA       NA       NA       NA       NA 11.181072 18.57237
#> Track_08       NA       NA       NA       NA       NA        NA 24.61636
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01  7.884058 30.144305 23.243120 27.373521
#> Track_02 36.260494  6.822463 12.137990 36.174540
#> Track_03 10.790155 20.160614 12.311625  6.337929
#> Track_04 30.339457 12.550501 17.592169 39.125780
#> Track_07 14.824678  9.635523  8.606986 19.093628
#> Track_08  5.262002 15.816824 10.161410 11.914756
#> Track_09 30.955401  8.638979  8.618817 27.283050
#> Track_13        NA 21.450507 15.618839 16.276932
#> Track_15        NA        NA  5.737652 21.230892
#> Track_16        NA        NA        NA 12.500729
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.611502 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 33.74496  7.324221 67.25493 41.46157  8.874539 49.037197
#> Track_02       NA       NA 19.771819 60.65870 32.46663 16.948413 34.470070
#> Track_03       NA       NA        NA 43.12026 22.16368  2.711702 26.705170
#> Track_04       NA       NA        NA       NA 14.22762 41.618510 22.672536
#> Track_07       NA       NA        NA       NA       NA 20.355287  6.279905
#> Track_08       NA       NA        NA       NA       NA        NA 25.515402
#> Track_09       NA       NA        NA       NA       NA        NA        NA
#> Track_13       NA       NA        NA       NA       NA        NA        NA
#> Track_15       NA       NA        NA       NA       NA        NA        NA
#> Track_16       NA       NA        NA       NA       NA        NA        NA
#> Track_18       NA       NA        NA       NA       NA        NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 42.46836 41.203166 22.897810 18.984120
#> Track_02 20.42886 32.754128 12.012841  8.870136
#> Track_03 22.99684 21.815357  9.907022  8.664446
#> Track_04 38.96757 12.156454 29.081165 43.077611
#> Track_07 16.22462  1.348179 11.303578 20.233507
#> Track_08 20.44786 20.094480  8.673369  6.490844
#> Track_09 14.79090  7.373563 13.579900 23.756196
#> Track_13       NA 16.997059  9.091632 16.370056
#> Track_15       NA        NA 11.485972 20.203163
#> Track_16       NA        NA        NA  6.944082
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.656919 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 16.52396 26.67118 65.65110 13.475095 30.86808 43.669499
#> Track_02       NA       NA 29.51299 62.76807  5.820875 21.14502 33.082452
#> Track_03       NA       NA       NA 31.53488 17.462746 20.62629 27.541366
#> Track_04       NA       NA       NA       NA 43.403853 34.80425 33.347472
#> Track_07       NA       NA       NA       NA        NA 11.95901 21.578405
#> Track_08       NA       NA       NA       NA        NA       NA  8.353757
#> Track_09       NA       NA       NA       NA        NA       NA        NA
#> Track_13       NA       NA       NA       NA        NA       NA        NA
#> Track_15       NA       NA       NA       NA        NA       NA        NA
#> Track_16       NA       NA       NA       NA        NA       NA        NA
#> Track_18       NA       NA       NA       NA        NA       NA        NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 41.854916 11.75088 15.565287 33.528310
#> Track_02 32.136809 18.75519 15.853490 24.180290
#> Track_03 26.633650 11.10562  7.724677 20.262279
#> Track_04 33.985791 42.91390 28.505747 31.099707
#> Track_07 20.133099 11.55416  8.157335 13.576617
#> Track_08  8.537406 19.54594 10.967031  3.005118
#> Track_09  3.839681 28.67362 17.957207  5.057851
#> Track_13        NA 27.62413 17.264825  6.090790
#> Track_15        NA       NA  6.637579 20.271324
#> Track_16        NA       NA        NA 11.174339
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.702536 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 58.30727 39.596703 22.28070 11.46128 39.624399 16.10464
#> Track_02       NA       NA  7.438035 62.40390 35.59831  8.774076 41.03007
#> Track_03       NA       NA        NA 44.88627 22.46456  6.133939 25.56156
#> Track_04       NA       NA        NA       NA  9.59952 42.289518 31.50071
#> Track_07       NA       NA        NA       NA       NA 21.215779 15.15992
#> Track_08       NA       NA        NA       NA       NA        NA 26.93931
#> Track_09       NA       NA        NA       NA       NA        NA       NA
#> Track_13       NA       NA        NA       NA       NA        NA       NA
#> Track_15       NA       NA        NA       NA       NA        NA       NA
#> Track_16       NA       NA        NA       NA       NA        NA       NA
#> Track_18       NA       NA        NA       NA       NA        NA       NA
#>          Track_13 Track_15  Track_16 Track_18
#> Track_01 41.65710 31.93694 24.867202 26.62923
#> Track_02 28.55582 24.02867 16.366923 30.21688
#> Track_03 20.83125 12.61180  8.175540 19.12103
#> Track_04 35.93018 42.53422 31.636616 19.76909
#> Track_07 21.44285 19.96816 13.136908 11.10733
#> Track_08 15.38255 16.13243  9.617203 15.61613
#> Track_09 33.70271 15.92987 13.410918 23.19582
#> Track_13       NA 27.84143 18.038914 10.61076
#> Track_15       NA       NA  5.945031 21.22755
#> Track_16       NA       NA        NA 13.01384
#> Track_18       NA       NA        NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:51.748207 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 28.53172 18.561750 55.27709 36.89903 39.77578 16.551579
#> Track_02       NA       NA  6.755932 28.61282 37.09131 31.60530 10.742052
#> Track_03       NA       NA        NA 23.17786 22.43865 18.43257  4.494862
#> Track_04       NA       NA        NA       NA 45.26156 27.82585 31.869905
#> Track_07       NA       NA        NA       NA       NA 14.99902 28.868214
#> Track_08       NA       NA        NA       NA       NA       NA 26.774762
#> Track_09       NA       NA        NA       NA       NA       NA        NA
#> Track_13       NA       NA        NA       NA       NA       NA        NA
#> Track_15       NA       NA        NA       NA       NA       NA        NA
#> Track_16       NA       NA        NA       NA       NA       NA        NA
#> Track_18       NA       NA        NA       NA       NA       NA        NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 46.914531 15.48147 27.973731 38.845153
#> Track_02 40.503398 27.17289 24.695701 24.101696
#> Track_03 26.408274 14.61910 12.650038 14.859142
#> Track_04 41.035642 42.52769 28.152037 14.104725
#> Track_07 14.660677 14.66249  8.492990 21.652403
#> Track_08  8.434646 20.29767  6.859615  9.244302
#> Track_09 34.473974 18.03885 19.188257 22.885291
#> Track_13        NA 24.75483  8.103778 17.477043
#> Track_15        NA       NA 11.587966 22.169945
#> Track_16        NA       NA        NA 10.416027
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.793774 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 59.09715 23.42209 27.59967 39.551258 18.24749 43.993374
#> Track_02       NA       NA 34.31614 64.24886  6.028957 30.92152 14.142162
#> Track_03       NA       NA       NA 12.33965 19.663338 17.39640 26.637339
#> Track_04       NA       NA       NA       NA 44.798933 33.01273 52.438624
#> Track_07       NA       NA       NA       NA        NA 17.54513  9.205302
#> Track_08       NA       NA       NA       NA        NA       NA 18.315658
#> Track_09       NA       NA       NA       NA        NA       NA        NA
#> Track_13       NA       NA       NA       NA        NA       NA        NA
#> Track_15       NA       NA       NA       NA        NA       NA        NA
#> Track_16       NA       NA       NA       NA        NA       NA        NA
#> Track_18       NA       NA       NA       NA        NA       NA        NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 43.87074 40.955586 18.00514 19.022107
#> Track_02 25.43661  9.862133 24.65528 31.802196
#> Track_03 16.99735 19.411371  5.94320 18.399187
#> Track_04 39.63299 44.712002 17.72598 34.155825
#> Track_07 16.16018  4.882782 12.51087 18.042111
#> Track_08 26.30620 20.548743 10.05158  1.451998
#> Track_09 26.38520 14.597155 18.19150 19.147497
#> Track_13       NA 12.963644 14.35769 27.238572
#> Track_15       NA        NA 13.04054 21.130716
#> Track_16       NA        NA       NA 10.625867
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.864977 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 46.81347 36.73303 56.179903 21.67866 19.641528 29.05770
#> Track_02       NA       NA 34.30802  7.382605 16.65493 17.503276 44.36847
#> Track_03       NA       NA       NA 42.481290 22.97813 21.875622 13.77957
#> Track_04       NA       NA       NA        NA 21.98048 23.209379 53.87642
#> Track_07       NA       NA       NA        NA       NA  1.534856 26.45109
#> Track_08       NA       NA       NA        NA       NA        NA 25.18958
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.06547 41.221409 17.902913  6.672802
#> Track_02 19.45432 20.177805 24.186304 28.411412
#> Track_03 19.43115 16.742778  9.661661 18.331948
#> Track_04 22.85388 22.833483 30.721746 36.534055
#> Track_07 20.85226 19.166887 11.608069 11.110959
#> Track_08 20.75272 19.025375 10.799370  9.973711
#> Track_09 32.34834 28.417639  8.510064 16.409722
#> Track_13       NA  2.942939 16.750061 24.722165
#> Track_15       NA        NA 14.007112 21.753132
#> Track_16       NA        NA        NA  7.388940
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:51.911185 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07  Track_08  Track_09
#> Track_01       NA 40.33037 38.36182 37.02595  7.798703  9.148791 30.104521
#> Track_02       NA       NA 20.18665 60.66925 28.246966 27.999181  6.543771
#> Track_03       NA       NA       NA 41.60680 20.569639 19.784746 16.549515
#> Track_04       NA       NA       NA       NA 22.602574 19.745729 48.509568
#> Track_07       NA       NA       NA       NA        NA  2.431532 19.729011
#> Track_08       NA       NA       NA       NA        NA        NA 20.086342
#> Track_09       NA       NA       NA       NA        NA        NA        NA
#> Track_13       NA       NA       NA       NA        NA        NA        NA
#> Track_15       NA       NA       NA       NA        NA        NA        NA
#> Track_16       NA       NA       NA       NA        NA        NA        NA
#> Track_18       NA       NA       NA       NA        NA        NA        NA
#>          Track_13  Track_15  Track_16   Track_18
#> Track_01 41.49347 15.518092 18.101970 38.7435377
#> Track_02 41.12789 32.677643 13.695365 20.9509124
#> Track_03 20.45556 20.847999 10.394437  0.8118664
#> Track_04 24.43256 13.487094 27.263358 41.3956514
#> Track_07 22.47211  5.957785  8.917703 20.5584054
#> Track_08 20.54563  4.635403  8.590192 19.6738698
#> Track_09 33.05650 23.518264  8.688212 17.1017264
#> Track_13       NA 18.256881 17.299340 20.0283986
#> Track_15       NA        NA 10.445058 20.6934004
#> Track_16       NA        NA        NA 10.5702311
#> Track_18       NA        NA        NA         NA
#> ------------------------------------
#> 2025-03-27 18:44:51.958261 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 59.18136 39.870783 40.56687  8.554315 32.04903 44.019900
#> Track_02       NA       NA  9.616514 49.60004 36.008550 25.49513 18.611877
#> Track_03       NA       NA        NA 30.01024 20.967102 16.89328  7.246686
#> Track_04       NA       NA        NA       NA 23.484923 41.68834 25.331033
#> Track_07       NA       NA        NA       NA        NA 18.13271 23.971018
#> Track_08       NA       NA        NA       NA        NA       NA 24.645103
#> Track_09       NA       NA        NA       NA        NA       NA        NA
#> Track_13       NA       NA        NA       NA        NA       NA        NA
#> Track_15       NA       NA        NA       NA        NA       NA        NA
#> Track_16       NA       NA        NA       NA        NA       NA        NA
#> Track_18       NA       NA        NA       NA        NA       NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 43.241814 37.150890 15.193234 17.221885
#> Track_02 21.070063 23.539998 24.476471 34.125946
#> Track_03 17.317509 10.379593 12.750771 20.807155
#> Track_04 48.578547 17.341966 24.942555 36.752084
#> Track_07 25.939516 18.766368  7.432928 11.219796
#> Track_08  9.051317 22.290276  8.903013 10.391508
#> Track_09 26.063134  7.036729 17.291349 26.281126
#> Track_13        NA 25.556256 14.994533 18.634922
#> Track_15        NA        NA 13.651187 22.697356
#> Track_16        NA        NA        NA  5.613073
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 7.155827 6.889059 7.146449 8.550492 9.358030 41.27125
#> Track_02       NA       NA 7.367287 7.095034 8.511176 8.404712 44.46323
#> Track_03       NA       NA       NA 2.752378 5.017897 5.558007 23.61988
#> Track_04       NA       NA       NA       NA 3.748965 5.898481 21.55969
#> Track_07       NA       NA       NA       NA       NA 3.362680 27.87452
#> Track_08       NA       NA       NA       NA       NA       NA 27.11038
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 23.69080 10.006241  8.956456 11.561263
#> Track_02 26.76916  8.652232  8.004512  9.939214
#> Track_03 11.38055  4.011654  6.187040  4.835886
#> Track_04 10.75445  2.107453  7.121107  3.128265
#> Track_07 16.24588  4.076379  5.378815  5.173189
#> Track_08 16.01165  5.869619  3.814577  6.169448
#> Track_09 12.44577 21.459210 36.643358 21.460416
#> Track_13       NA 11.406049 22.371414 11.843307
#> Track_15       NA        NA  7.547234  1.280341
#> Track_16       NA        NA        NA  8.182026
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0      0.0        0      0.1      0.3      0.4
#> Track_02       NA       NA      0.1        0      0.2      0.0      1.0
#> Track_03       NA       NA       NA        0      0.0      0.1      0.5
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.1
#> Track_07       NA       NA       NA       NA       NA      0.2      0.9
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.1      0.0      0.0      0.1
#> Track_02      0.4      0.1      0.0      0.1
#> Track_03      0.2      0.0      0.1      0.1
#> Track_04      0.1      0.0      0.0      0.0
#> Track_07      0.5      0.1      0.0      0.1
#> Track_08      0.5      0.1      0.0      0.2
#> Track_09      0.2      0.6      1.0      0.5
#> Track_13       NA      0.1      1.0      0.3
#> Track_15       NA       NA      0.3      0.0
#> Track_16       NA       NA       NA      0.4
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 5: Simulating and comparing DTW distance in the MountTom dataset using the
# "Origin" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
#> 2025-03-27 18:44:52.212729 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 84.67862 44.02823 128.37961 32.63137 62.62917 28.16155
#> Track_02       NA       NA 20.93363  58.48448 64.23111 67.25571 44.58672
#> Track_03       NA       NA       NA  59.97924 34.54824 41.82973 15.50441
#> Track_04       NA       NA       NA        NA 81.08974 68.83449 86.39044
#> Track_07       NA       NA       NA        NA       NA 21.12570 34.75405
#> Track_08       NA       NA       NA        NA       NA       NA 49.12258
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15 Track_16   Track_18
#> Track_01 90.180245 80.31488 52.89645 32.8473918
#> Track_02 44.395914 55.28440 38.98937 64.8611405
#> Track_03 39.958639 41.69623 22.91202 34.6786566
#> Track_04  8.036855 28.79882 34.46914 81.4893347
#> Track_07 53.261487 44.00544 21.20897  0.7811318
#> Track_08 42.920382 30.08768 14.29605 21.2495794
#> Track_09 59.043350 54.62484 34.10876 34.8241438
#> Track_13        NA 16.25293 18.31718 53.7843321
#> Track_15        NA       NA 11.53932 43.6275587
#> Track_16        NA       NA       NA 21.2637477
#> Track_18        NA       NA       NA         NA
#> ------------------------------------
#> 2025-03-27 18:44:52.2584 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03  Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 108.9672 74.673893 119.29622 73.73681 17.54553  63.07908
#> Track_02       NA       NA  6.523853  20.98902 54.35132 58.54316  84.20863
#> Track_03       NA       NA        NA  17.96204 33.88278 34.60875  54.34229
#> Track_04       NA       NA        NA        NA 75.21609 65.76673 103.97405
#> Track_07       NA       NA        NA        NA       NA 38.24274  26.47583
#> Track_08       NA       NA        NA        NA       NA       NA  40.75702
#> Track_09       NA       NA        NA        NA       NA       NA        NA
#> Track_13       NA       NA        NA        NA       NA       NA        NA
#> Track_15       NA       NA        NA        NA       NA       NA        NA
#> Track_16       NA       NA        NA        NA       NA       NA        NA
#> Track_18       NA       NA        NA        NA       NA       NA        NA
#>           Track_13  Track_15 Track_16  Track_18
#> Track_01 86.535431 74.417989 25.27921 77.123276
#> Track_02  7.858608 11.181675 43.72735 14.445384
#> Track_03  3.407988  5.302130 23.02628  9.138708
#> Track_04 12.957735 11.350020 54.38610 31.363928
#> Track_07 43.465973 38.065355 20.67328 28.173913
#> Track_08 42.495100 34.138017 10.40094 37.729480
#> Track_09 65.814139 56.042250 21.11149 51.313355
#> Track_13        NA  4.631328 30.19811 13.585299
#> Track_15        NA        NA 23.61865 13.950006
#> Track_16        NA        NA       NA 23.409736
#> Track_18        NA        NA       NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:52.303881 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 60.07507 78.23309  8.290084 53.16318 60.74668 39.08265
#> Track_02       NA       NA 51.89848 63.379754 67.76726 66.89442 69.61936
#> Track_03       NA       NA       NA 84.211896 41.64209 34.62862 53.88273
#> Track_04       NA       NA       NA        NA 57.13223 65.26660 42.22333
#> Track_07       NA       NA       NA        NA       NA 10.57371 16.83422
#> Track_08       NA       NA       NA        NA       NA       NA 25.33277
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13 Track_15  Track_16 Track_18
#> Track_01  93.60997 82.02509 46.261027 14.90762
#> Track_02  69.99716 62.07431 45.007385 43.15917
#> Track_03  10.16152  8.50122 20.004738 41.39541
#> Track_04 100.35783 88.20733 49.885458 18.08573
#> Track_07  49.30808 43.86076 12.355948 23.64671
#> Track_08  38.21171 33.43892  7.065489 27.42153
#> Track_09  64.21193 55.55601 21.511966 16.42736
#> Track_13        NA  2.94020 24.509614 52.64022
#> Track_15        NA       NA 20.787626 44.73913
#> Track_16        NA       NA        NA 18.51105
#> Track_18        NA       NA        NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:52.348966 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 14.11167 58.57619 102.62235 46.43713 46.148857 75.47327
#> Track_02       NA       NA 58.77144 103.69894 34.84247 34.706994 73.56871
#> Track_03       NA       NA       NA  13.04019 42.89987 41.548204 11.47947
#> Track_04       NA       NA       NA        NA 84.55501 83.970496 19.11040
#> Track_07       NA       NA       NA        NA       NA  1.473526 48.28239
#> Track_08       NA       NA       NA        NA       NA        NA 47.67570
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>           Track_13 Track_15 Track_16 Track_18
#> Track_01  6.473663 60.76996 52.51589 42.48510
#> Track_02 12.974075 49.49365 46.56018 45.45837
#> Track_03 38.352288 45.71168 17.67302 11.82357
#> Track_04 75.522905 86.71233 41.98420 30.00873
#> Track_07 30.932900 10.46878 21.65787 38.36416
#> Track_08 30.421716 10.25949 21.10332 37.38955
#> Track_09 51.127951 49.38872 18.81117 22.80080
#> Track_13        NA 41.14147 30.85626 27.51286
#> Track_15        NA       NA 21.44064 43.17284
#> Track_16        NA       NA       NA 19.99354
#> Track_18        NA       NA       NA       NA
#> ------------------------------------
#> 2025-03-27 18:44:52.394149 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 107.8238 63.96409 39.30082 72.35213 50.19119 94.25285
#> Track_02       NA       NA 22.37345 91.68545 64.55808 35.03819 14.25190
#> Track_03       NA       NA       NA 48.16952 45.08734 10.60911 24.51430
#> Track_04       NA       NA       NA       NA 81.85488 32.52523 84.56267
#> Track_07       NA       NA       NA       NA       NA 43.99336 45.70293
#> Track_08       NA       NA       NA       NA       NA       NA 33.73257
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 91.20739 26.063247 50.39441 60.265250
#> Track_02 50.18768 57.132092 17.79520 64.802102
#> Track_03 42.79914 26.364481 10.32065 41.358795
#> Track_04 95.29829  8.191719 46.18758 74.468667
#> Track_07 23.17460 46.725734 23.21850  8.374472
#> Track_08 46.99289 16.821835 13.84572 40.381833
#> Track_09 32.07056 48.684451 12.64217 46.985890
#> Track_13       NA 54.409329 22.13016 28.375789
#> Track_15       NA        NA 20.96845 39.228869
#> Track_16       NA        NA       NA 22.042599
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:52.43986 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 89.13579 76.51642 83.13743 38.89514 76.164337 21.03438
#> Track_02       NA       NA 67.65035 21.48351 71.69241 61.479255 75.18761
#> Track_03       NA       NA       NA 81.37328 31.79449  5.423809 48.07401
#> Track_04       NA       NA       NA       NA 77.28092 76.074091 76.63910
#> Track_07       NA       NA       NA       NA       NA 32.572318 16.39947
#> Track_08       NA       NA       NA       NA       NA        NA 48.39522
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 91.24260 78.23050 27.80146 30.634142
#> Track_02 62.39862 43.85020 31.66661 66.873100
#> Track_03 18.51574 23.61193 24.17799 32.747819
#> Track_04 80.59611 60.53386 33.81898 71.045343
#> Track_07 48.24013 43.34618 18.28396  5.261275
#> Track_08 12.91934 18.51383 22.88779 33.138326
#> Track_09 62.17915 53.12123 19.18075 10.406610
#> Track_13       NA 12.13787 30.58839 47.807280
#> Track_15       NA       NA 23.09710 41.684333
#> Track_16       NA       NA       NA 15.732545
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:52.49429 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 31.85189 70.26952  98.54362 72.12444 46.74175 83.194018
#> Track_02       NA       NA 68.53889 109.59075 55.62476 54.40383 81.964414
#> Track_03       NA       NA       NA  26.66281 35.15557 19.76401  4.811409
#> Track_04       NA       NA       NA        NA 78.65412 21.53686 25.588567
#> Track_07       NA       NA       NA        NA       NA 40.30915 45.686974
#> Track_08       NA       NA       NA        NA       NA       NA 23.620223
#> Track_09       NA       NA       NA        NA       NA       NA        NA
#> Track_13       NA       NA       NA        NA       NA       NA        NA
#> Track_15       NA       NA       NA        NA       NA       NA        NA
#> Track_16       NA       NA       NA        NA       NA       NA        NA
#> Track_18       NA       NA       NA        NA       NA       NA        NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 88.98989 78.697062 44.61026 38.591419
#> Track_02 80.56488 72.428276 33.09586 49.211401
#> Track_03 24.70620 17.255072 21.18338 25.065081
#> Track_04 66.87472 53.884746 53.81010 32.642449
#> Track_07 25.42981 28.707906 10.19485 42.026671
#> Track_08 41.20805 33.417426 21.89462  6.022037
#> Track_09 33.28373 24.526860 31.15892 30.168195
#> Track_13       NA  9.950832 22.08602 46.396377
#> Track_15       NA        NA 21.14640 38.567603
#> Track_16       NA        NA       NA 22.484391
#> Track_18       NA        NA       NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:52.540258 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 113.0018 29.14994 59.43945 63.89312 51.28796 71.40484
#> Track_02       NA       NA 56.73605 88.47031 66.28995 40.97953 78.40409
#> Track_03       NA       NA       NA 15.96766 39.21575 14.71804 47.72212
#> Track_04       NA       NA       NA       NA 80.86006 20.63520 94.88639
#> Track_07       NA       NA       NA       NA       NA 42.91823  3.29011
#> Track_08       NA       NA       NA       NA       NA       NA 51.53348
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 11.28142 32.55309 24.66275 52.137576
#> Track_02 77.57129 72.42210 43.63225 39.946291
#> Track_03 13.31481 27.82826  8.86199 14.951314
#> Track_04 37.44984 64.56903 33.05831 21.551751
#> Track_07 43.23195 21.85967 21.29828 43.137505
#> Track_08 28.93252 37.21071 15.56608  1.149891
#> Track_09 51.58650 25.83096 28.88355 51.401964
#> Track_13       NA 24.67094 10.67658 29.416726
#> Track_15       NA       NA 14.33353 37.487274
#> Track_16       NA       NA       NA 15.458463
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:52.586057 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA   105.42 70.70976 46.97632 74.91284 67.542198 95.84026
#> Track_02       NA       NA 60.82923 87.84128 51.00939  9.354451 26.58547
#> Track_03       NA       NA       NA 85.69394 10.67090 36.202399 37.05995
#> Track_04       NA       NA       NA       NA 85.94229 56.111638 90.96024
#> Track_07       NA       NA       NA       NA       NA 31.016212 27.11749
#> Track_08       NA       NA       NA       NA       NA        NA 20.08549
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 77.414735 38.70285 48.424305 66.854319
#> Track_02 76.311019 47.23958 18.368981 63.218500
#> Track_03  7.242391 45.11281 22.137825  4.660955
#> Track_04 98.475175 15.99854 44.633565 83.988065
#> Track_07 19.379787 44.84262 19.873333 14.496707
#> Track_08 47.683610 24.77602  5.742158 37.231592
#> Track_09 49.579003 47.56005 16.553849 40.157682
#> Track_13        NA 54.83104 30.340668  2.577324
#> Track_15        NA       NA 17.639951 43.095579
#> Track_16        NA       NA        NA 22.402291
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-03-27 18:44:52.631058 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 112.5431 70.86235 129.97032 35.86729 74.453867 83.44247
#> Track_02       NA       NA 17.01526  27.97988 67.78241  7.223845 56.03759
#> Track_03       NA       NA       NA  36.57145 41.85584 10.285899 42.66701
#> Track_04       NA       NA       NA        NA 77.44911 20.722156 49.35189
#> Track_07       NA       NA       NA        NA       NA 41.455740 34.47021
#> Track_08       NA       NA       NA        NA       NA        NA 34.75728
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 88.758473 69.506356 52.832185 69.515594
#> Track_02 10.670773 30.302275 17.601326 24.074096
#> Track_03 17.694691  9.488867  9.775945  5.043313
#> Track_04 15.166828 52.528165 25.674063 44.634703
#> Track_07 52.208032 45.266806 23.708832 42.596121
#> Track_08  6.077123 18.779818  6.512978 15.025736
#> Track_09 38.760933 50.670947 23.079275 46.685815
#> Track_13        NA 28.245818 11.759402 23.248743
#> Track_15        NA        NA 14.989599  4.910604
#> Track_16        NA        NA        NA 12.217277
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 8.365556 8.140001 13.288601 17.963246 14.043056 80.66455
#> Track_02       NA       NA 9.928547 13.870318 15.997694 11.986057 83.72110
#> Track_03       NA       NA       NA  4.242441  9.250629  6.131319 46.49768
#> Track_04       NA       NA       NA        NA  5.401950  3.172849 42.61717
#> Track_07       NA       NA       NA        NA        NA  5.261381 51.73148
#> Track_08       NA       NA       NA        NA        NA        NA 54.42241
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.00788 18.730370 14.932046 22.179224
#> Track_02 47.19796 17.799527  9.836672 20.569492
#> Track_03 20.29566  7.910351 10.158387  9.928217
#> Track_04 19.90114  4.165947  8.969975  6.396727
#> Track_07 29.44950  2.944420  6.952380  5.167036
#> Track_08 27.32867  4.915326  5.349900  6.863688
#> Track_09 22.41087 41.825031 68.973412 42.969072
#> Track_13       NA 20.579060 39.239001 21.761749
#> Track_15       NA        NA  8.572483  2.784293
#> Track_16       NA        NA        NA 10.795457
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0      0.0      0.1        0      0.0      0.6
#> Track_02       NA       NA      0.1      0.0        0      0.2      0.9
#> Track_03       NA       NA       NA      0.0        0      0.1      0.6
#> Track_04       NA       NA       NA       NA        0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.1      1.0
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.2      0.0      0.0      0.1
#> Track_02      0.4      0.1      0.0      0.1
#> Track_03      0.6      0.1      0.2      0.3
#> Track_04      0.3      0.0      0.0      0.0
#> Track_07      0.3      0.0      0.0      0.1
#> Track_08      0.2      0.0      0.0      0.2
#> Track_09      0.0      0.2      1.0      0.6
#> Track_13       NA      0.5      1.0      0.2
#> Track_15       NA       NA      0.0      0.0
#> Track_16       NA       NA       NA      0.0
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#>