Similarity metric using Dynamic Time Warping (DTW)

simil_DTW_metric() computes similarity metrics between two or more trajectories using Dynamic Time Warping (DTW). It allows for different superposition methods to align trajectories before calculating the DTW metric. The function also supports testing with simulations to calculate p-values for the DTW distance metrics.

Usage

simil_DTW_metric(data, test = NULL, sim = NULL, superposition = NULL)

Arguments

data

A track R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

test

Logical; if TRUE, the function compares the observed DTW distances against simulated trajectories and calculates p-values. Default is FALSE.

sim

A track simulation R object consisting of a list of simulated trajectories to use for comparison when test = TRUE.

superposition

A character string indicating the method used to align trajectories. Options are "None", "Centroid", or "Origin". Default is "None".

Value

A track similarity R object consisting ofa list containing the following elements:

DTW_distance_metric

A matrix containing the pairwise DTW distances between trajectories.

DTW_distance_metric_p_values

(If test is TRUE) A matrix containing the p-values for the pairwise DTW distances.

DTW_metric_p_values_combined

(If test is TRUE) The overall p-value for the combined DTW distances.

DTW_distance_metric_simulations

(If test is TRUE) A list of DTW distance matrices from each simulated dataset.

Details

The simil_DTW_metric() function calculates the similarity between trajectories using the Dynamic Time Warping (DTW) algorithm from the dtw package. The dtw() function is used with the dist.method argument set to "Euclidean" for computing the local distances between points in the trajectories.

DTW aligns two time series by minimizing the cumulative distance between their points, creating an optimal alignment despite variations in length or temporal distortions. The algorithm constructs a distance matrix where each element represents the cost of aligning points between the two series and finds a warping path through this matrix that minimizes the total distance. The warping path is contiguous and monotonic, starting from the bottom-left corner and ending at the top-right corner (Cleasby et al., 2019).

DTW measures are non-negative and unbounded, with larger values indicating greater dissimilarity between the time series. This method has been used in various contexts, including ecological studies to analyze and cluster trajectory data (Cleasby et al., 2019).

Potential limitations and biases of DTW include sensitivity to noise and outliers, computational complexity, and the need for appropriate distance metrics. Additionally, DTW may not always account for all structural differences between trajectories and can be biased by the chosen alignment constraints. While DTW can handle trajectories of different lengths due to its elastic nature, having trajectories of similar lengths can improve the accuracy and interpretability of the similarity measure. Similar lengths result in a more meaningful alignment and can make the computation more efficient. When trajectories differ significantly in length, preprocessing or normalization might be necessary, and careful analysis is required to understand the alignment path. The function’s flexibility in handling different lengths allows it to be applied in various contexts. However, large differences in trajectory lengths might introduce potential biases that should be considered when interpreting the results.

The function offers three different superposition methods to align the trajectories before DTW() is applied:

"None": No superposition is applied.

"Centroid": Trajectories are shifted to align based on their centroids.

"Origin": Trajectories are shifted to align based on their starting point.

If test = TRUE, the function can compute p-values by comparing the observed DTW distances with those generated from a set of simulated trajectories. The p-values are calculated for both individual trajectory pairs and for the entire set of trajectories.

Logo

References

Cleasby, I. R., Wakefield, E. D., Morrissey, B. J., Bodey, T. W., Votier, S. C., Bearhop, S., & Hamer, K. C. (2019). Using time-series similarity measures to compare animal movement trajectories in ecology. Behavioral Ecology and Sociobiology, 73, 1-19.

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Simulating tracks using the "Directed" model and comparing DTW distance
# in the PaluxyRiver dataset
s1 <- simulate_track(PaluxyRiver, nsim = 3, model = "Directed")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-05-19 13:21:14.583509 Iteration 1
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.585735 Iteration 2
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.587885 Iteration 3
#>  
#> DTW metric
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1   Track_2
#> Track_1      NA 0.6666667
#> Track_2      NA        NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.6666667
#> 
#> $DTW_distance_metric_simulations
#> $DTW_distance_metric_simulations[[1]]
#>         Track_1  Track_2
#> Track_1      NA 30.72955
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_simulations[[2]]
#>         Track_1  Track_2
#> Track_1      NA 23.34381
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_simulations[[3]]
#>         Track_1  Track_2
#> Track_1      NA 16.36784
#> Track_2      NA       NA
#> 
#> 

# Example 2: Simulating tracks using the "Constrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s2 <- simulate_track(PaluxyRiver, nsim = 3, model = "Constrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
#> 2025-05-19 13:21:14.601973 Iteration 1
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.604133 Iteration 2
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.606268 Iteration 3
#>  
#> DTW metric
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
#> $DTW_distance_metric_simulations
#> $DTW_distance_metric_simulations[[1]]
#>         Track_1 Track_2
#> Track_1      NA 37.5479
#> Track_2      NA      NA
#> 
#> $DTW_distance_metric_simulations[[2]]
#>         Track_1  Track_2
#> Track_1      NA 96.37455
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_simulations[[3]]
#>         Track_1  Track_2
#> Track_1      NA 172.9115
#> Track_2      NA       NA
#> 
#> 

# Example 3: Simulating tracks using the "Unconstrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s3 <- simulate_track(PaluxyRiver, nsim = 3, model = "Unconstrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
#> 2025-05-19 13:21:14.620435 Iteration 1
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.622627 Iteration 2
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.624752 Iteration 3
#>  
#> DTW metric
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
#> $DTW_distance_metric_simulations
#> $DTW_distance_metric_simulations[[1]]
#>         Track_1  Track_2
#> Track_1      NA 110.3167
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_simulations[[2]]
#>         Track_1  Track_2
#> Track_1      NA 818.0027
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_simulations[[3]]
#>         Track_1 Track_2
#> Track_1      NA 468.311
#> Track_2      NA      NA
#> 
#> 

# Example 4: Simulating and comparing DTW distance in the MountTom dataset using the
# "Centroid" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 3)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
#> 2025-05-19 13:21:14.763379 Iteration 1
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.817275 Iteration 2
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:14.863012 Iteration 3
#>  
#> DTW metric
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 7.155827 6.889059 7.146449 8.550492 9.358030 41.27125
#> Track_02       NA       NA 7.367287 7.095034 8.511176 8.404712 44.46323
#> Track_03       NA       NA       NA 2.752378 5.017897 5.558007 23.61988
#> Track_04       NA       NA       NA       NA 3.748965 5.898481 21.55969
#> Track_07       NA       NA       NA       NA       NA 3.362680 27.87452
#> Track_08       NA       NA       NA       NA       NA       NA 27.11038
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 23.69080 10.006241  8.956456 11.561263
#> Track_02 26.76916  8.652232  8.004512  9.939214
#> Track_03 11.38055  4.011654  6.187040  4.835886
#> Track_04 10.75445  2.107453  7.121107  3.128265
#> Track_07 16.24588  4.076379  5.378815  5.173189
#> Track_08 16.01165  5.869619  3.814577  6.169448
#> Track_09 12.44577 21.459210 36.643358 21.460416
#> Track_13       NA 11.406049 22.371414 11.843307
#> Track_15       NA        NA  7.547234  1.280341
#> Track_16       NA        NA        NA  8.182026
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08  Track_09
#> Track_01       NA        0        0 0.3333333        0 0.0000000 1.0000000
#> Track_02       NA       NA        0 0.0000000        0 0.3333333 1.0000000
#> Track_03       NA       NA       NA 0.0000000        0 0.0000000 0.6666667
#> Track_04       NA       NA       NA        NA        0 0.0000000 0.3333333
#> Track_07       NA       NA       NA        NA       NA 0.0000000 0.3333333
#> Track_08       NA       NA       NA        NA       NA        NA 1.0000000
#> Track_09       NA       NA       NA        NA       NA        NA        NA
#> Track_13       NA       NA       NA        NA       NA        NA        NA
#> Track_15       NA       NA       NA        NA       NA        NA        NA
#> Track_16       NA       NA       NA        NA       NA        NA        NA
#> Track_18       NA       NA       NA        NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 0.0000000 0.0000000 0.0000000 0.0000000
#> Track_02 1.0000000 0.3333333 0.0000000 0.0000000
#> Track_03 0.0000000 0.0000000 0.0000000 0.0000000
#> Track_04 0.3333333 0.0000000 0.0000000 0.0000000
#> Track_07 0.0000000 0.0000000 0.0000000 0.0000000
#> Track_08 0.6666667 0.6666667 0.0000000 0.3333333
#> Track_09 0.3333333 0.0000000 1.0000000 0.6666667
#> Track_13        NA 0.0000000 1.0000000 0.6666667
#> Track_15        NA        NA 0.3333333 0.0000000
#> Track_16        NA        NA        NA 0.0000000
#> Track_18        NA        NA        NA        NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
#> $DTW_distance_metric_simulations
#> $DTW_distance_metric_simulations[[1]]
#>          Track_01 Track_02  Track_03 Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 27.76773 16.606926 39.29551 34.81758 39.27462 24.444980
#> Track_02       NA       NA  7.964637 12.59047 36.42229 32.47174  5.471068
#> Track_03       NA       NA        NA 14.06870 21.77121 19.41190  5.336298
#> Track_04       NA       NA        NA       NA 43.52327 35.66142  9.254562
#> Track_07       NA       NA        NA       NA       NA 14.87605 28.546630
#> Track_08       NA       NA        NA       NA       NA       NA 24.794765
#> Track_09       NA       NA        NA       NA       NA       NA        NA
#> Track_13       NA       NA        NA       NA       NA       NA        NA
#> Track_15       NA       NA        NA       NA       NA       NA        NA
#> Track_16       NA       NA        NA       NA       NA       NA        NA
#> Track_18       NA       NA        NA       NA       NA       NA        NA
#>           Track_13  Track_15  Track_16 Track_18
#> Track_01 32.878500 40.968002 27.360404 36.53011
#> Track_02 12.079593 38.061227 24.425339 22.78325
#> Track_03 11.402170 24.027246 12.712601 14.57835
#> Track_04  8.698299 44.009864 28.935749 21.78974
#> Track_07 28.508115  9.749711  8.591807 20.83021
#> Track_08 20.984687  9.973360  6.725304 10.39459
#> Track_09  9.555908 29.938502 18.188392 17.07310
#> Track_13        NA 28.319585 16.793482 11.96461
#> Track_15        NA        NA  7.971705 18.58816
#> Track_16        NA        NA        NA 10.28874
#> Track_18        NA        NA        NA       NA
#> 
#> $DTW_distance_metric_simulations[[2]]
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08  Track_09
#> Track_01       NA 58.69358 33.03430  6.856834 27.835432 38.510715 22.280089
#> Track_02       NA       NA 23.63636 64.764154 29.616260  7.431294 40.451192
#> Track_03       NA       NA       NA 38.095991  6.072348 14.399450 14.364730
#> Track_04       NA       NA       NA        NA 32.413664 44.411137 26.242018
#> Track_07       NA       NA       NA        NA        NA 17.184381  8.938576
#> Track_08       NA       NA       NA        NA        NA        NA 25.813877
#> Track_09       NA       NA       NA        NA        NA        NA        NA
#> Track_13       NA       NA       NA        NA        NA        NA        NA
#> Track_15       NA       NA       NA        NA        NA        NA        NA
#> Track_16       NA       NA       NA        NA        NA        NA        NA
#> Track_18       NA       NA       NA        NA        NA        NA        NA
#>          Track_13  Track_15 Track_16  Track_18
#> Track_01 43.67650 40.179740 13.04731 38.410844
#> Track_02 21.00538  5.610299 25.11322 12.168317
#> Track_03 25.14100 15.393681 11.88703 17.201362
#> Track_04 49.57503 46.934297 17.69210 44.753436
#> Track_07 27.69336 19.179222 10.76841 19.888639
#> Track_08 12.15406  2.638546 12.74667  5.088981
#> Track_09 33.87745 27.020710 12.13173 27.138809
#> Track_13       NA 13.319691 17.11342  8.511120
#> Track_15       NA        NA 14.01889  5.358591
#> Track_16       NA        NA       NA 12.579536
#> Track_18       NA        NA       NA        NA
#> 
#> $DTW_distance_metric_simulations[[3]]
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08 Track_09
#> Track_01       NA  59.8885 21.54234 42.93670 30.136750 39.56759 37.10054
#> Track_02       NA       NA 33.66332 54.78223 29.710395 14.42452 29.01322
#> Track_03       NA       NA       NA 10.57934  6.471104 16.24989 27.19814
#> Track_04       NA       NA       NA       NA  8.979812 29.70083 52.52196
#> Track_07       NA       NA       NA       NA        NA 12.21869 28.03700
#> Track_08       NA       NA       NA       NA        NA       NA 24.55548
#> Track_09       NA       NA       NA       NA        NA       NA       NA
#> Track_13       NA       NA       NA       NA        NA       NA       NA
#> Track_15       NA       NA       NA       NA        NA       NA       NA
#> Track_16       NA       NA       NA       NA        NA       NA       NA
#> Track_18       NA       NA       NA       NA        NA       NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 46.819395 40.713488 28.706714 34.225126
#> Track_02  9.945081 13.081015 13.680003 20.576881
#> Track_03 27.423297 17.713220 11.042673 21.341319
#> Track_04 50.107065 32.091865 24.549037 43.775684
#> Track_07 25.610937 13.777570  9.271004 21.369884
#> Track_08 15.400130  1.788868  4.739688 16.357925
#> Track_09 18.446607 24.575370 16.822872  5.372652
#> Track_13        NA 14.864443 12.459303 11.686186
#> Track_15        NA        NA  5.602034 16.339850
#> Track_16        NA        NA        NA 10.825538
#> Track_18        NA        NA        NA        NA
#> 
#> 

# Example 5: Simulating and comparing DTW distance in the MountTom dataset using the
# "Origin" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 3)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
#> 2025-05-19 13:21:15.011021 Iteration 1
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:15.056144 Iteration 2
#>  
#> DTW metric
#> ------------------------------------
#> 2025-05-19 13:21:15.108811 Iteration 3
#>  
#> DTW metric
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 8.365556 8.140001 13.288601 17.963246 14.043056 80.66455
#> Track_02       NA       NA 9.928547 13.870318 15.997694 11.986057 83.72110
#> Track_03       NA       NA       NA  4.242441  9.250629  6.131319 46.49768
#> Track_04       NA       NA       NA        NA  5.401950  3.172849 42.61717
#> Track_07       NA       NA       NA        NA        NA  5.261381 51.73148
#> Track_08       NA       NA       NA        NA        NA        NA 54.42241
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.00788 18.730370 14.932046 22.179224
#> Track_02 47.19796 17.799527  9.836672 20.569492
#> Track_03 20.29566  7.910351 10.158387  9.928217
#> Track_04 19.90114  4.165947  8.969975  6.396727
#> Track_07 29.44950  2.944420  6.952380  5.167036
#> Track_08 27.32867  4.915326  5.349900  6.863688
#> Track_09 22.41087 41.825031 68.973412 42.969072
#> Track_13       NA 20.579060 39.239001 21.761749
#> Track_15       NA        NA  8.572483  2.784293
#> Track_16       NA        NA        NA 10.795457
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01  Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 0.3333333        0        0        0 0.3333333 0.3333333
#> Track_02       NA        NA        0        0        0 0.0000000 0.6666667
#> Track_03       NA        NA       NA        0        0 0.0000000 0.6666667
#> Track_04       NA        NA       NA       NA        0 0.0000000 0.0000000
#> Track_07       NA        NA       NA       NA       NA 0.3333333 0.6666667
#> Track_08       NA        NA       NA       NA       NA        NA 1.0000000
#> Track_09       NA        NA       NA       NA       NA        NA        NA
#> Track_13       NA        NA       NA       NA       NA        NA        NA
#> Track_15       NA        NA       NA       NA       NA        NA        NA
#> Track_16       NA        NA       NA       NA       NA        NA        NA
#> Track_18       NA        NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15 Track_16  Track_18
#> Track_01 0.3333333 0.0000000        0 0.3333333
#> Track_02 0.3333333 0.0000000        0 0.0000000
#> Track_03 0.3333333 0.0000000        0 0.0000000
#> Track_04 0.0000000 0.0000000        0 0.0000000
#> Track_07 0.3333333 0.0000000        0 0.0000000
#> Track_08 0.6666667 0.0000000        0 0.0000000
#> Track_09 0.0000000 0.3333333        1 0.6666667
#> Track_13        NA 0.3333333        1 0.3333333
#> Track_15        NA        NA        0 0.0000000
#> Track_16        NA        NA       NA 0.3333333
#> Track_18        NA        NA       NA        NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 
#> $DTW_distance_metric_simulations
#> $DTW_distance_metric_simulations[[1]]
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 30.34654 61.17894 57.39041 63.86482 22.24351 89.76501
#> Track_02       NA       NA 64.40437 77.85452 44.09158 35.30704 76.16587
#> Track_03       NA       NA       NA 37.31680 43.05850 23.70804 40.52057
#> Track_04       NA       NA       NA       NA 82.31857 19.40577 89.86573
#> Track_07       NA       NA       NA       NA       NA 38.41663 30.84763
#> Track_08       NA       NA       NA       NA       NA       NA 47.40589
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>           Track_13  Track_15 Track_16  Track_18
#> Track_01 66.901590 37.297644 50.65292 11.450625
#> Track_02 72.650000 48.780689 45.52621 25.009096
#> Track_03  5.154424 19.075187 12.75656 29.247809
#> Track_04 36.147189  8.839846 45.67071 31.136891
#> Track_07 53.822234 45.067243 22.05078 35.290739
#> Track_08 26.581234  9.282636 19.83560  7.458807
#> Track_09 49.145878 48.420505 18.48658 47.853277
#> Track_13        NA 19.947516 17.61872 33.645858
#> Track_15        NA        NA 19.95537 16.053343
#> Track_16        NA        NA       NA 21.325683
#> Track_18        NA        NA       NA        NA
#> 
#> $DTW_distance_metric_simulations[[2]]
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 4.860765 74.94096 88.57201 21.33641 18.423094 45.49924
#> Track_02       NA       NA 67.06140 80.33960 17.99008 14.760807 39.76980
#> Track_03       NA       NA       NA 78.60816 37.99303 37.000399 36.39697
#> Track_04       NA       NA       NA       NA 70.60049 64.933286 88.46726
#> Track_07       NA       NA       NA       NA       NA  3.633356 15.52430
#> Track_08       NA       NA       NA       NA       NA        NA 16.82532
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13 Track_15 Track_16 Track_18
#> Track_01 14.082114 22.40024 47.04993 41.27621
#> Track_02 12.544307 19.78054 40.14648 36.35104
#> Track_03 52.042002 44.91634 19.62763 41.14927
#> Track_04 77.785730 43.18852 31.47510 21.05263
#> Track_07  7.760319 21.53307 22.86114 29.30752
#> Track_08  7.166999 18.53730 21.37563 26.30923
#> Track_09 26.034415 36.38214 31.35557 42.33758
#> Track_13        NA 22.09544 30.69268 33.60780
#> Track_15        NA       NA 19.79778 13.51471
#> Track_16        NA       NA       NA 14.00335
#> Track_18        NA       NA       NA       NA
#> 
#> $DTW_distance_metric_simulations[[3]]
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 35.99635 14.56888 42.93174 54.70509 13.49592 84.78659
#> Track_02       NA       NA 16.58853 71.48639 29.76541 30.44216 84.23506
#> Track_03       NA       NA       NA 35.72677 25.26877 10.48571 50.56719
#> Track_04       NA       NA       NA       NA 74.97716 19.54464 71.75940
#> Track_07       NA       NA       NA       NA       NA 32.17560 53.94281
#> Track_08       NA       NA       NA       NA       NA       NA 44.06968
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 89.72557 69.47827 53.28253 76.105919
#> Track_02 76.99464 47.19972 42.48455 64.268346
#> Track_03 50.92449 34.59122 22.98817 40.485779
#> Track_04 95.86413 85.83591 56.94821 82.511621
#> Track_07 41.55532 14.28681 19.16697 32.372112
#> Track_08 51.29674 39.63463 23.23001 40.633628
#> Track_09 36.08344 51.92638 24.45454 31.659919
#> Track_13       NA 31.33465 11.09801  3.171102
#> Track_15       NA       NA 15.45749 23.781094
#> Track_16       NA       NA       NA  6.569464
#> Track_18       NA       NA       NA        NA
#> 
#>