Similarity metric using Dynamic Time Warping (DTW)

simil_DTW_metric() computes similarity metrics between two or more trajectories using Dynamic Time Warping (DTW). It allows for different superposition methods to align trajectories before calculating the DTW metric. The function also supports testing with simulations to calculate p-values for the DTW distance metrics.

Usage

simil_DTW_metric(data, test = NULL, sim = NULL, superposition = NULL)

Arguments

data

A 'track' R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

test

Logical; if TRUE, the function compares the observed DTW distances against simulated trajectories and calculates p-values. Default is FALSE.

sim

A 'track simulation' R object consisting of a list of simulated trajectories to use for comparison when test = TRUE.

superposition

A character string indicating the method used to align trajectories. Options are "None", "Centroid", or "Origin". Default is "None".

Value

A 'track similarity' R object consisting ofa list containing the following elements:

DTW_distance_metric

A matrix containing the pairwise DTW distances between trajectories.

DTW_distance_metric_p_values

(If test is TRUE) A matrix containing the p-values for the pairwise DTW distances.

DTW_metric_p_values_combined

(If test is TRUE) The overall p-value for the combined DTW distances.

DTW_distance_metric_simulations

(If test is TRUE) A list of DTW distance matrices from each simulated dataset.

Details

The simil_DTW_metric() function calculates the similarity between trajectories using the Dynamic Time Warping (DTW) algorithm from the dtw package. The dtw() function is used with the dist.method argument set to "Euclidean" for computing the local distances between points in the trajectories.

DTW aligns two time series by minimizing the cumulative distance between their points, creating an optimal alignment despite variations in length or temporal distortions. The algorithm constructs a distance matrix where each element represents the cost of aligning points between the two series and finds a warping path through this matrix that minimizes the total distance. The warping path is contiguous and monotonic, starting from the bottom-left corner and ending at the top-right corner (Cleasby et al., 2019).

DTW measures are non-negative and unbounded, with larger values indicating greater dissimilarity between the time series. This method has been used in various contexts, including ecological studies to analyze and cluster trajectory data (Cleasby et al., 2019).

Potential limitations and biases of DTW include sensitivity to noise and outliers, computational complexity, and the need for appropriate distance metrics. Additionally, DTW may not always account for all structural differences between trajectories and can be biased by the chosen alignment constraints. While DTW can handle trajectories of different lengths due to its elastic nature, having trajectories of similar lengths can improve the accuracy and interpretability of the similarity measure. Similar lengths result in a more meaningful alignment and can make the computation more efficient. When trajectories differ significantly in length, preprocessing or normalization might be necessary, and careful analysis is required to understand the alignment path. The function’s flexibility in handling different lengths allows it to be applied in various contexts. However, large differences in trajectory lengths might introduce potential biases that should be considered when interpreting the results.

The function offers three different superposition methods to align the trajectories before DTW() is applied:

"None": No superposition is applied.

"Centroid": Trajectories are shifted to align based on their centroids.

"Origin": Trajectories are shifted to align based on their starting point.

If test = TRUE, the function can compute p-values by comparing the observed DTW distances with those generated from a set of simulated trajectories. The p-values are calculated for both individual trajectory pairs and for the entire set of trajectories.

Logo

References

Cleasby, I. R., Wakefield, E. D., Morrissey, B. J., Bodey, T. W., Votier, S. C., Bearhop, S., & Hamer, K. C. (2019). Using time-series similarity measures to compare animal movement trajectories in ecology. Behavioral Ecology and Sociobiology, 73, 1-19.

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Simulating tracks using the "Directed" model and comparing DTW distance
# in the PaluxyRiver dataset
s1 <- simulate_track(PaluxyRiver, nsim = 10, model = "Directed")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
#> 2025-03-29 23:36:28.878252 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 21.68999
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.879976 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 41.66089
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.881572 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 28.53861
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.883237 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 27.68667
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.884847 Iteration 5
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 14.21587
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.886447 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 20.47582
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.888055 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 17.18041
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.889702 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 23.32304
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.891297 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 26.82374
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.892896 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 17.52372
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA     0.6
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0.6
#> 

# Example 2: Simulating tracks using the "Constrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s2 <- simulate_track(PaluxyRiver, nsim = 10, model = "Constrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
#> 2025-03-29 23:36:28.921259 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 33.76018
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.922877 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 139.8911
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.924453 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 35.00938
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.926047 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 42.03352
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.927624 Iteration 5
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 38.1155
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-03-29 23:36:28.929229 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 30.00584
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.930816 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 35.90138
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.932387 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 102.9326
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.933976 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 101.6141
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.93554 Iteration 10
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 168.3487
#> Track_2      NA       NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 3: Simulating tracks using the "Unconstrained" model and comparing DTW distance
# in the PaluxyRiver dataset
s3 <- simulate_track(PaluxyRiver, nsim = 10, model = "Unconstrained")
simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
#> 2025-03-29 23:36:28.964287 Iteration 1
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 719.4052
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.965922 Iteration 2
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 531.7201
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.967519 Iteration 3
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 258.5845
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.969145 Iteration 4
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 803.0809
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.97074 Iteration 5
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 771.076
#> Track_2      NA      NA
#> ------------------------------------
#> 2025-03-29 23:36:28.972321 Iteration 6
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 758.7927
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.973921 Iteration 7
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 630.9672
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.975489 Iteration 8
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 473.3425
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.977081 Iteration 9
#>  
#> DTW metric
#>         Track_1  Track_2
#> Track_1      NA 735.5614
#> Track_2      NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:28.978654 Iteration 10
#>  
#> DTW metric
#>         Track_1 Track_2
#> Track_1      NA 703.828
#> Track_2      NA      NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>         Track_1  Track_2
#> Track_1      NA 23.47722
#> Track_2      NA       NA
#> 
#> $DTW_distance_metric_p_values
#>         Track_1 Track_2
#> Track_1      NA       0
#> Track_2      NA      NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 4: Simulating and comparing DTW distance in the MountTom dataset using the
# "Centroid" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
#> 2025-03-29 23:36:29.20738 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 53.89461 22.79239 39.889616 38.42230  8.105630 21.88556
#> Track_02       NA       NA 36.68412 63.011135 31.66323 33.887707 29.61769
#> Track_03       NA       NA       NA  7.541115 13.28223  9.739126 24.16217
#> Track_04       NA       NA       NA        NA 24.21617 23.032437 45.57068
#> Track_07       NA       NA       NA        NA       NA 18.564977 28.74558
#> Track_08       NA       NA       NA        NA       NA        NA 14.64289
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 33.76659 36.145587 14.445326 26.680602
#> Track_02 17.15969  6.941005 24.922323 18.414356
#> Track_03 26.47822 21.891823  7.748162 20.225181
#> Track_04 49.04264 42.956556 20.419687 41.484481
#> Track_07 27.56542 19.697244 11.920658 22.163770
#> Track_08 20.39748 18.996633  4.463644 14.526539
#> Track_09 12.58119 17.416398 12.097935  7.842610
#> Track_13       NA 10.196827 15.351590  6.486999
#> Track_15       NA        NA 12.489615  8.877057
#> Track_16       NA        NA        NA 10.222981
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.25292 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08  Track_09
#> Track_01       NA 57.19075 33.32475 65.832172  7.906537 10.407530 50.161775
#> Track_02       NA       NA 29.51110  8.864571 33.229113 30.991994  5.776708
#> Track_03       NA       NA       NA 38.642317 18.602481 17.808232 24.855621
#> Track_04       NA       NA       NA        NA 40.785878 37.350439  7.802513
#> Track_07       NA       NA       NA        NA        NA  2.845559 27.897196
#> Track_08       NA       NA       NA        NA        NA        NA 26.288554
#> Track_09       NA       NA       NA        NA        NA        NA        NA
#> Track_13       NA       NA       NA        NA        NA        NA        NA
#> Track_15       NA       NA       NA        NA        NA        NA        NA
#> Track_16       NA       NA       NA        NA        NA        NA        NA
#> Track_18       NA       NA       NA        NA        NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01  6.400469 41.281548 23.834494 24.288636
#> Track_02 39.107760  7.789175 22.946930 32.414226
#> Track_03 22.133453 16.672432  5.183005  6.748191
#> Track_04 47.385804 12.799729 29.371940 41.150444
#> Track_07  5.512065 22.566296 11.627887 13.992605
#> Track_08  6.884917 20.099664 11.106649 13.917997
#> Track_09 34.109862  5.942740 18.279686 27.112768
#> Track_13        NA 27.656244 14.897011 16.618990
#> Track_15        NA        NA 12.279332 19.272943
#> Track_16        NA        NA        NA  5.288930
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.298481 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 14.79142 37.47459 67.71236 34.27106 34.504249  7.092883
#> Track_02       NA       NA 34.48869 58.77505 25.17461 32.836527  8.261116
#> Track_03       NA       NA       NA 27.41036 19.88342  3.796294 25.154918
#> Track_04       NA       NA       NA       NA 24.57679 31.949986 49.934216
#> Track_07       NA       NA       NA       NA       NA 20.453906 20.875511
#> Track_08       NA       NA       NA       NA       NA        NA 23.425078
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 44.566424 38.537083 15.767101 36.482038
#> Track_02 35.249622 29.594682 11.466374 33.855153
#> Track_03 21.484634 20.439988 12.644066  1.093599
#> Track_04 17.043035 21.193978 29.210392 27.918719
#> Track_07  9.395437  3.927502  9.920386 19.772394
#> Track_08 23.406012 21.388039 12.013275  3.223258
#> Track_09 29.624420 24.183009  7.429217 24.703303
#> Track_13        NA  6.052129 15.644546 21.420646
#> Track_15        NA        NA 11.761595 20.431411
#> Track_16        NA        NA        NA 12.408269
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.34396 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07 Track_08  Track_09
#> Track_01       NA 12.74012 32.64542 60.41498 38.283454 15.21338  7.542223
#> Track_02       NA       NA 24.07313 50.10639 31.038085 18.87740  7.457079
#> Track_03       NA       NA       NA 13.78321  7.457616 19.47676 20.939273
#> Track_04       NA       NA       NA       NA  7.540752 41.60188 45.041086
#> Track_07       NA       NA       NA       NA        NA 20.95982 26.001811
#> Track_08       NA       NA       NA       NA        NA       NA 13.400708
#> Track_09       NA       NA       NA       NA        NA       NA        NA
#> Track_13       NA       NA       NA       NA        NA       NA        NA
#> Track_15       NA       NA       NA       NA        NA       NA        NA
#> Track_16       NA       NA       NA       NA        NA       NA        NA
#> Track_18       NA       NA       NA       NA        NA       NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 46.98130 27.310848 27.006263 33.759729
#> Track_02 42.11350 29.101722 24.613612 25.049154
#> Track_03 20.71391 23.514924 12.531678  1.644188
#> Track_04 27.52852 43.115423 26.024504 12.293993
#> Track_07 14.70559 23.246482 11.168190  6.162119
#> Track_08 25.11464  8.243477 10.252813 19.871287
#> Track_09 35.22166 22.481134 18.944152 21.931795
#> Track_13       NA 23.179082 10.494091 19.438359
#> Track_15       NA        NA  9.520839 23.415172
#> Track_16       NA        NA        NA 12.242886
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.395981 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 40.93515 35.76092 41.24645 19.35427 38.009631 38.045239
#> Track_02       NA       NA 34.63842 61.73574 34.38028 32.354295 43.866632
#> Track_03       NA       NA       NA 20.61313 13.25005  4.506918  6.495339
#> Track_04       NA       NA       NA       NA 13.19992 27.116945 15.065223
#> Track_07       NA       NA       NA       NA       NA 15.777618 14.032028
#> Track_08       NA       NA       NA       NA       NA        NA 11.299496
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 17.687266 34.981382 15.077748 10.605126
#> Track_02 38.567038  8.233282 22.369835 29.762465
#> Track_03 18.466360 21.463269 10.478019 15.884669
#> Track_04 18.783928 45.216431 20.977043 21.283958
#> Track_07  6.060526 23.716060  6.784543  5.782485
#> Track_08 21.381893 19.035846 11.249215 17.560775
#> Track_09 18.889786 28.915353 13.622045 18.300155
#> Track_13        NA 27.530377  7.920249  6.920822
#> Track_15        NA        NA 13.555960 20.389292
#> Track_16        NA        NA        NA  4.729260
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.440977 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 54.97195 39.78566 49.31052 40.50736 34.542330 46.777731
#> Track_02       NA       NA 25.46469 24.43570  6.15172 32.226982 16.042412
#> Track_03       NA       NA       NA 40.01334 14.02398  7.918408 11.348500
#> Track_04       NA       NA       NA       NA 22.07970 43.500023 33.757478
#> Track_07       NA       NA       NA       NA       NA 18.955162  8.877319
#> Track_08       NA       NA       NA       NA       NA        NA 18.388610
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>           Track_13 Track_15 Track_16  Track_18
#> Track_01 45.965677 34.45342 16.17077 33.413511
#> Track_02 33.209656 35.12489 20.85850 32.539046
#> Track_03  6.012915 10.96250 12.71650  8.641691
#> Track_04 48.503305 45.74355 21.15701 43.047808
#> Track_07 20.451747 22.10493 11.70070 19.193266
#> Track_08  9.677748  4.05680 12.23453  1.006297
#> Track_09 16.395796 21.56040 16.21958 18.965348
#> Track_13        NA 12.63860 17.91683 10.371307
#> Track_15        NA       NA 13.35552  3.545485
#> Track_16        NA       NA       NA 12.078487
#> Track_18        NA       NA       NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.486076 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 55.60917  6.933181 65.83361 29.00723 39.230927 43.388885
#> Track_02       NA       NA 33.913099 35.60051 20.96349 22.315925 32.153765
#> Track_03       NA       NA        NA 41.37297 15.19631 20.623968 24.518375
#> Track_04       NA       NA        NA       NA 40.49483  6.618948 13.880129
#> Track_07       NA       NA        NA       NA       NA 20.785363 27.199466
#> Track_08       NA       NA        NA       NA       NA        NA  6.886829
#> Track_09       NA       NA        NA       NA       NA        NA        NA
#> Track_13       NA       NA        NA       NA       NA        NA        NA
#> Track_15       NA       NA        NA       NA       NA        NA        NA
#> Track_16       NA       NA        NA       NA       NA        NA        NA
#> Track_18       NA       NA        NA       NA       NA        NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 19.77360 12.133564 27.324926 24.347144
#> Track_02 41.63752 36.277193 10.762507 33.580358
#> Track_03 12.11354  5.817678 12.340682 12.329444
#> Track_04 42.30016 40.763024 23.035511 29.283217
#> Track_07 25.34752 19.516457  8.667122 20.741051
#> Track_08 22.12178 20.463946  9.938312 13.609021
#> Track_09 23.11239 23.181865 15.599108 13.810159
#> Track_13       NA  7.451959 17.977892  8.378440
#> Track_15       NA        NA 14.080755  9.722432
#> Track_16       NA        NA        NA 12.534522
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.531536 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02  Track_03 Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 20.65911  9.472037 56.06962  7.378506 36.69544 29.54744
#> Track_02       NA       NA 18.724519 39.72345 16.836388 36.12609 36.76545
#> Track_03       NA       NA        NA 39.20510  2.373064 16.95769 13.26149
#> Track_04       NA       NA        NA       NA 38.748214 37.36683 50.71644
#> Track_07       NA       NA        NA       NA        NA 18.52898 15.60667
#> Track_08       NA       NA        NA       NA        NA       NA 14.03865
#> Track_09       NA       NA        NA       NA        NA       NA       NA
#> Track_13       NA       NA        NA       NA        NA       NA       NA
#> Track_15       NA       NA        NA       NA        NA       NA       NA
#> Track_16       NA       NA        NA       NA        NA       NA       NA
#> Track_18       NA       NA        NA       NA        NA       NA       NA
#>           Track_13 Track_15  Track_16  Track_18
#> Track_01 29.324491 40.71036 28.568890 36.112537
#> Track_02 36.371697 34.00688 26.326779 26.011854
#> Track_03 13.455699 22.05489 12.328837 19.624570
#> Track_04 49.823042 17.94824 25.992196  7.716605
#> Track_07 15.527043 23.29126 13.237492 19.916285
#> Track_08 15.113554 14.94588  5.216030 17.498288
#> Track_09  3.285175 27.02345 13.508340 27.041724
#> Track_13        NA 27.52087 14.349745 27.483143
#> Track_15        NA       NA  9.851883  7.484125
#> Track_16        NA       NA        NA 10.877805
#> Track_18        NA       NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.576756 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08  Track_09
#> Track_01       NA 20.22445 27.33551 49.94590 11.99799 36.037811 43.952416
#> Track_02       NA       NA 15.44787 31.65420 21.17058 25.770448 31.824774
#> Track_03       NA       NA       NA 10.20225 17.65369  8.803502 13.094491
#> Track_04       NA       NA       NA       NA 38.30274 12.780177 15.547565
#> Track_07       NA       NA       NA       NA       NA 20.338421 26.745029
#> Track_08       NA       NA       NA       NA       NA        NA  3.965548
#> Track_09       NA       NA       NA       NA       NA        NA        NA
#> Track_13       NA       NA       NA       NA       NA        NA        NA
#> Track_15       NA       NA       NA       NA       NA        NA        NA
#> Track_16       NA       NA       NA       NA       NA        NA        NA
#> Track_18       NA       NA       NA       NA       NA        NA        NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 45.75747 27.68409 28.17853 38.151395
#> Track_02 42.21348 32.65448 23.46805 35.182864
#> Track_03 24.62430 21.88215 10.24440 19.332410
#> Track_04 41.05635 41.91298 21.51560 36.219472
#> Track_07 25.58563 11.28942 13.20415 18.914546
#> Track_08 18.21953 21.57862  7.18433 14.721528
#> Track_09 22.41092 27.99895 10.15661 19.357491
#> Track_13       NA 19.37625  9.46962  6.174775
#> Track_15       NA       NA 12.46528 13.989926
#> Track_16       NA       NA       NA  7.409926
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.621985 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04  Track_07  Track_08  Track_09
#> Track_01       NA 60.15162 40.47642 52.48005 40.816385 39.585614 49.875240
#> Track_02       NA       NA  8.26963 35.97798 14.489219 10.408725  5.973954
#> Track_03       NA       NA       NA 29.93268  5.295428  7.212609  4.623188
#> Track_04       NA       NA       NA       NA 36.421535 19.302801 30.777952
#> Track_07       NA       NA       NA       NA        NA 11.789573 10.848588
#> Track_08       NA       NA       NA       NA        NA        NA  6.899931
#> Track_09       NA       NA       NA       NA        NA        NA        NA
#> Track_13       NA       NA       NA       NA        NA        NA        NA
#> Track_15       NA       NA       NA       NA        NA        NA        NA
#> Track_16       NA       NA       NA       NA        NA        NA        NA
#> Track_18       NA       NA       NA       NA        NA        NA        NA
#>          Track_13 Track_15  Track_16  Track_18
#> Track_01 14.50927 29.73276 14.266883 40.476903
#> Track_02 40.68544 25.54184 24.865484  7.060387
#> Track_03 26.56731 16.20156 12.818454  4.570738
#> Track_04 31.53318 10.21441 24.498158 23.814043
#> Track_07 27.83283 19.40708 13.329997  9.569601
#> Track_08 23.62341 11.51544 12.189127  2.877569
#> Track_09 32.85975 19.89441 18.615004  4.242384
#> Track_13       NA 14.93545  8.047702 25.287278
#> Track_15       NA       NA  9.469126 13.526618
#> Track_16       NA       NA        NA 12.661449
#> Track_18       NA       NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 7.155827 6.889059 7.146449 8.550492 9.358030 41.27125
#> Track_02       NA       NA 7.367287 7.095034 8.511176 8.404712 44.46323
#> Track_03       NA       NA       NA 2.752378 5.017897 5.558007 23.61988
#> Track_04       NA       NA       NA       NA 3.748965 5.898481 21.55969
#> Track_07       NA       NA       NA       NA       NA 3.362680 27.87452
#> Track_08       NA       NA       NA       NA       NA       NA 27.11038
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 23.69080 10.006241  8.956456 11.561263
#> Track_02 26.76916  8.652232  8.004512  9.939214
#> Track_03 11.38055  4.011654  6.187040  4.835886
#> Track_04 10.75445  2.107453  7.121107  3.128265
#> Track_07 16.24588  4.076379  5.378815  5.173189
#> Track_08 16.01165  5.869619  3.814577  6.169448
#> Track_09 12.44577 21.459210 36.643358 21.460416
#> Track_13       NA 11.406049 22.371414 11.843307
#> Track_15       NA        NA  7.547234  1.280341
#> Track_16       NA        NA        NA  8.182026
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0        0        0      0.2      0.1      0.5
#> Track_02       NA       NA        0        0      0.1      0.0      1.0
#> Track_03       NA       NA       NA        0      0.1      0.2      0.6
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.4
#> Track_07       NA       NA       NA       NA       NA      0.1      0.8
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.4      0.0      0.0      0.1
#> Track_02      0.1      0.3      0.0      0.1
#> Track_03      0.1      0.0      0.1      0.3
#> Track_04      0.0      0.0      0.0      0.0
#> Track_07      0.5      0.1      0.0      0.0
#> Track_08      0.3      0.1      0.0      0.3
#> Track_09      0.1      0.3      1.0      0.6
#> Track_13       NA      0.3      1.0      0.5
#> Track_15       NA       NA      0.0      0.0
#> Track_16       NA       NA       NA      0.3
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#> 

# Example 5: Simulating and comparing DTW distance in the MountTom dataset using the
# "Origin" superposition method
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 10)
#> Warning: `model` is NULL. Defaulting to 'Unconstrained'.
simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
#> 2025-03-29 23:36:29.920409 Iteration 1
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03 Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 27.37988 47.63223 95.39700 74.45628 20.300574 98.02401
#> Track_02       NA       NA 54.96790 68.16387 70.01518  7.085528 86.99290
#> Track_03       NA       NA       NA 83.56796 24.51316 31.168787 47.27065
#> Track_04       NA       NA       NA       NA 80.93916 48.328951 76.24743
#> Track_07       NA       NA       NA       NA       NA 41.336790 24.18228
#> Track_08       NA       NA       NA       NA       NA        NA 52.88650
#> Track_09       NA       NA       NA       NA       NA        NA       NA
#> Track_13       NA       NA       NA       NA       NA        NA       NA
#> Track_15       NA       NA       NA       NA       NA        NA       NA
#> Track_16       NA       NA       NA       NA       NA        NA       NA
#> Track_18       NA       NA       NA       NA       NA        NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01 46.196838 58.298202 29.477216 29.584135
#> Track_02 57.333443 62.301133 19.444722 40.180411
#> Track_03  6.580922  8.379618 21.282746 11.555072
#> Track_04 95.826864 84.280257 35.650589 76.525229
#> Track_07 34.087623 18.226260 24.842563 30.817966
#> Track_08 34.163300 35.596261  7.189742 22.223217
#> Track_09 59.506317 41.663848 34.466821 51.202741
#> Track_13        NA 15.905523 25.921509  8.671263
#> Track_15        NA        NA 23.065804 18.134121
#> Track_16        NA        NA        NA 16.724371
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:29.966814 Iteration 2
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08  Track_09
#> Track_01       NA 100.3943 11.19845 123.75467 75.58708 48.74715 89.079144
#> Track_02       NA       NA 59.96376 103.44170 54.87559 27.57771 70.432549
#> Track_03       NA       NA       NA  78.23493 40.48342 23.83362 50.311156
#> Track_04       NA       NA       NA        NA 16.86625 79.41063 10.629352
#> Track_07       NA       NA       NA        NA       NA 37.23479  6.838218
#> Track_08       NA       NA       NA        NA       NA       NA 50.433251
#> Track_09       NA       NA       NA        NA       NA       NA        NA
#> Track_13       NA       NA       NA        NA       NA       NA        NA
#> Track_15       NA       NA       NA        NA       NA       NA        NA
#> Track_16       NA       NA       NA        NA       NA       NA        NA
#> Track_18       NA       NA       NA        NA       NA       NA        NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 47.67372 73.789644 48.959171 15.869027
#> Track_02 82.01217  8.183239 17.671302 55.495119
#> Track_03 27.99431 39.947789 21.143680  4.946666
#> Track_04 61.11002 74.540840 50.096517 80.763987
#> Track_07 38.40507 34.445227 19.535309 41.086930
#> Track_08 45.58853 19.787820 11.407853 20.127592
#> Track_09 43.09383 47.043867 29.201281 51.058719
#> Track_13       NA 56.364401 31.733189 31.876916
#> Track_15       NA        NA  9.648989 37.766691
#> Track_16       NA        NA        NA 19.903371
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:30.013277 Iteration 3
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 96.15991 76.28743  74.58084 73.92298 15.55919 38.14762
#> Track_02       NA       NA 52.47117 117.91770 12.90711 52.16258 85.28796
#> Track_03       NA       NA       NA  57.84824 28.75987 40.43603 46.40529
#> Track_04       NA       NA       NA        NA 80.94726 54.73420 23.12145
#> Track_07       NA       NA       NA        NA       NA 35.92887 55.46628
#> Track_08       NA       NA       NA        NA       NA       NA 28.29172
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>          Track_13 Track_15  Track_16  Track_18
#> Track_01 29.64462 70.22656 51.429243 71.869809
#> Track_02 52.50317 66.11937 35.856144 57.641749
#> Track_03 52.01347 18.06193  7.480974  8.075085
#> Track_04 75.77321 34.18895 42.846730 45.979422
#> Track_07 40.38785 41.86257 17.788511 32.507198
#> Track_08 10.53090 39.24979 22.311713 38.147936
#> Track_09 43.98704 34.19280 29.862068 39.178233
#> Track_13       NA 52.25866 30.368607 50.574502
#> Track_15       NA       NA 14.095722 10.544392
#> Track_16       NA       NA        NA  8.023299
#> Track_18       NA       NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:30.068768 Iteration 4
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 61.35588 65.99121 65.350498 27.66650 75.03157 11.19522
#> Track_02       NA       NA 30.43769  5.963162 57.17244 50.39489 39.40473
#> Track_03       NA       NA       NA 39.500646 42.91042 16.33483 42.02789
#> Track_04       NA       NA       NA        NA 65.86181 62.12679 43.16466
#> Track_07       NA       NA       NA        NA       NA 42.77048 22.69216
#> Track_08       NA       NA       NA        NA       NA       NA 49.72377
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>           Track_13  Track_15  Track_16  Track_18
#> Track_01  9.897785 76.776773 40.160144 18.112296
#> Track_02 40.393122 54.276396 21.717659 27.629280
#> Track_03 42.460103 18.389760  9.875426 30.192097
#> Track_04 43.872873 66.680479 29.560888 31.136000
#> Track_07 21.895726 46.788385 21.914765 21.279905
#> Track_08 50.048403  3.504491 15.355494 37.559717
#> Track_09  4.742525 51.518329 22.386819  6.501991
#> Track_13        NA 53.424173 23.003352  7.197257
#> Track_15        NA        NA 16.894600 39.994488
#> Track_16        NA        NA        NA 14.444456
#> Track_18        NA        NA        NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:30.115153 Iteration 5
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 78.77519 55.09909  33.27422 61.229441 58.166581 81.57102
#> Track_02       NA       NA 68.44849 101.54767  8.028976 66.965635 84.39633
#> Track_03       NA       NA       NA  42.18306 43.997698  5.129432 16.85537
#> Track_04       NA       NA       NA        NA 76.852862 47.660607 69.87151
#> Track_07       NA       NA       NA        NA        NA 40.641607 53.01151
#> Track_08       NA       NA       NA        NA        NA        NA 11.36660
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01 90.83419 34.08657 55.81593 73.99259
#> Track_02 75.50740 64.82573 37.84057 29.68589
#> Track_03 31.54191 15.30491 20.65306 41.28179
#> Track_04 89.84850 17.78913 58.01432 82.83836
#> Track_07 48.41993 42.74050 20.08266 15.95438
#> Track_08 25.59399 17.79112 18.45415 37.48911
#> Track_09 23.07438 33.18724 23.06090 48.00720
#> Track_13       NA 44.63670 15.03906 35.07878
#> Track_15       NA       NA 24.32143 43.18561
#> Track_16       NA       NA       NA 13.22771
#> Track_18       NA       NA       NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:30.160677 Iteration 6
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA  91.9356 23.25400 130.96503 78.11898 17.865068 59.87845
#> Track_02       NA       NA 62.97124 103.33145 26.31650 59.251072 80.98505
#> Track_03       NA       NA       NA  73.53425 42.93146  4.090012 23.42870
#> Track_04       NA       NA       NA        NA 54.97756 75.147375 57.47115
#> Track_07       NA       NA       NA        NA       NA 41.081255 48.94854
#> Track_08       NA       NA       NA        NA       NA        NA 26.76654
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>          Track_13 Track_15 Track_16  Track_18
#> Track_01 78.39695 79.49710 26.99594 64.367934
#> Track_02 79.35321 65.67693 32.37552  7.710383
#> Track_03 35.13984 38.78751 13.29676 37.423703
#> Track_04 33.96697 10.53994 56.42951 69.994717
#> Track_07 45.04189 32.01009 21.49045 14.719193
#> Track_08 37.28857 39.29538 11.10412 35.100915
#> Track_09 19.61202 31.16953 27.43420 49.118105
#> Track_13       NA 16.32495 29.47940 49.627822
#> Track_15       NA       NA 26.00887 39.082814
#> Track_16       NA       NA       NA 16.610571
#> Track_18       NA       NA       NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:30.205905 Iteration 7
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08  Track_09
#> Track_01       NA 100.5459 75.69365  62.93413 75.860949 76.042393 91.304293
#> Track_02       NA       NA 44.18701 122.85203 55.236141 39.795458 70.177931
#> Track_03       NA       NA       NA  71.42980  9.165987  4.151803 13.130747
#> Track_04       NA       NA       NA        NA 65.122518 74.625359 70.894067
#> Track_07       NA       NA       NA        NA        NA 12.806415  5.750041
#> Track_08       NA       NA       NA        NA        NA        NA 18.021382
#> Track_09       NA       NA       NA        NA        NA        NA        NA
#> Track_13       NA       NA       NA        NA        NA        NA        NA
#> Track_15       NA       NA       NA        NA        NA        NA        NA
#> Track_16       NA       NA       NA        NA        NA        NA        NA
#> Track_18       NA       NA       NA        NA        NA        NA        NA
#>          Track_13  Track_15 Track_16 Track_18
#> Track_01 42.66984 47.344914 26.40374 63.96868
#> Track_02 47.52873 35.160161 36.30839 64.09755
#> Track_03 48.62968 39.736957 22.92271 20.86633
#> Track_04 81.07071 78.687797 44.33081 42.39377
#> Track_07 52.94214 45.074207 24.41498 14.76849
#> Track_08 47.90064 38.470334 22.86122 23.98195
#> Track_09 65.86075 55.819793 35.66869 17.69509
#> Track_13       NA  8.076112 13.63851 50.53821
#> Track_15       NA        NA 12.82039 43.41564
#> Track_16       NA        NA       NA 21.78808
#> Track_18       NA        NA       NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:30.251162 Iteration 8
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07  Track_08 Track_09
#> Track_01       NA 112.8605 76.01677  78.50251 57.21344 74.797347 48.27883
#> Track_02       NA       NA 40.57657 116.86697 65.19672  8.103923 82.52857
#> Track_03       NA       NA       NA  55.24365 23.62911 25.485431 38.44701
#> Track_04       NA       NA       NA        NA 17.46825 80.555328 13.63594
#> Track_07       NA       NA       NA        NA       NA 39.761901 14.27902
#> Track_08       NA       NA       NA        NA       NA        NA 50.75693
#> Track_09       NA       NA       NA        NA       NA        NA       NA
#> Track_13       NA       NA       NA        NA       NA        NA       NA
#> Track_15       NA       NA       NA        NA       NA        NA       NA
#> Track_16       NA       NA       NA        NA       NA        NA       NA
#> Track_18       NA       NA       NA        NA       NA        NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01  6.70567 20.42358 27.76001 63.92825
#> Track_02 78.67572 64.56879 46.14541 30.78782
#> Track_03 49.15614 42.97129 21.24272 36.84696
#> Track_04 57.16585 64.44059 31.26781 83.10131
#> Track_07 36.29972 36.74535 14.28338 44.09440
#> Track_08 47.97234 37.36854 23.66257 16.65294
#> Track_09 32.45840 37.44413 13.28661 51.47144
#> Track_13       NA 11.72871 14.02093 40.28959
#> Track_15       NA       NA 14.98702 28.45697
#> Track_16       NA       NA       NA 22.84815
#> Track_18       NA       NA       NA       NA
#> ------------------------------------
#> 2025-03-29 23:36:30.296046 Iteration 9
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07 Track_08 Track_09
#> Track_01       NA 48.59229 47.15755 119.82268 51.380729 78.73596 73.27043
#> Track_02       NA       NA 61.12719  75.43311 62.890623 69.26142 80.99909
#> Track_03       NA       NA       NA  85.67595  3.594356 30.31662 16.78742
#> Track_04       NA       NA       NA        NA 85.404508 64.07753 96.28604
#> Track_07       NA       NA       NA        NA        NA 27.85379 13.11942
#> Track_08       NA       NA       NA        NA        NA       NA 25.77278
#> Track_09       NA       NA       NA        NA        NA       NA       NA
#> Track_13       NA       NA       NA        NA        NA       NA       NA
#> Track_15       NA       NA       NA        NA        NA       NA       NA
#> Track_16       NA       NA       NA        NA        NA       NA       NA
#> Track_18       NA       NA       NA        NA        NA       NA       NA
#>           Track_13 Track_15 Track_16  Track_18
#> Track_01 93.198434 58.21841 55.07556 66.141528
#> Track_02 81.407170 25.07348 41.99912 35.574485
#> Track_03 41.689600 43.29862 22.97373 42.971610
#> Track_04 66.237261 30.78930 35.68085 18.118724
#> Track_07 39.157242 43.95916 22.26297 42.913197
#> Track_08  7.247338 39.58071 12.05373 34.652884
#> Track_09 37.267899 52.70244 27.82162 51.310327
#> Track_13        NA 46.28944 14.03880 39.441816
#> Track_15        NA       NA 18.99564  8.736572
#> Track_16        NA       NA       NA 15.351832
#> Track_18        NA       NA       NA        NA
#> ------------------------------------
#> 2025-03-29 23:36:30.341159 Iteration 10
#>  
#> DTW metric
#>          Track_01 Track_02 Track_03  Track_04 Track_07 Track_08 Track_09
#> Track_01       NA 110.3409 12.10093 128.37337 71.86623 74.34987 61.70169
#> Track_02       NA       NA 64.02722  45.51336 45.55738 29.59785 74.44693
#> Track_03       NA       NA       NA  77.77853 36.92946 38.32670 33.21734
#> Track_04       NA       NA       NA        NA 26.84341 11.18799 65.25438
#> Track_07       NA       NA       NA        NA       NA 11.89074 23.67096
#> Track_08       NA       NA       NA        NA       NA       NA 33.65652
#> Track_09       NA       NA       NA        NA       NA       NA       NA
#> Track_13       NA       NA       NA        NA       NA       NA       NA
#> Track_15       NA       NA       NA        NA       NA       NA       NA
#> Track_16       NA       NA       NA        NA       NA       NA       NA
#> Track_18       NA       NA       NA        NA       NA       NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 48.58873 74.932262 21.674103 74.272585
#> Track_02 62.10160  8.870791 44.162969 35.344175
#> Track_03 28.83157 39.454034  6.211708 38.796870
#> Track_04 91.41914 42.207551 55.827065 15.554764
#> Track_07 55.01272 31.944861 23.669901  8.275409
#> Track_08 49.57569 21.917427 23.317472  4.317380
#> Track_09 60.04098 50.464721 26.545696 30.962872
#> Track_13       NA 37.382687 19.692267 52.006350
#> Track_15       NA        NA 23.774613 25.589444
#> Track_16       NA        NA        NA 23.788674
#> Track_18       NA        NA        NA        NA
#> ------------------------------------
#> ANALYSIS COMPLETED
#> ------------------------------------
#>  
#> $DTW_distance_metric
#>          Track_01 Track_02 Track_03  Track_04  Track_07  Track_08 Track_09
#> Track_01       NA 8.365556 8.140001 13.288601 17.963246 14.043056 80.66455
#> Track_02       NA       NA 9.928547 13.870318 15.997694 11.986057 83.72110
#> Track_03       NA       NA       NA  4.242441  9.250629  6.131319 46.49768
#> Track_04       NA       NA       NA        NA  5.401950  3.172849 42.61717
#> Track_07       NA       NA       NA        NA        NA  5.261381 51.73148
#> Track_08       NA       NA       NA        NA        NA        NA 54.42241
#> Track_09       NA       NA       NA        NA        NA        NA       NA
#> Track_13       NA       NA       NA        NA        NA        NA       NA
#> Track_15       NA       NA       NA        NA        NA        NA       NA
#> Track_16       NA       NA       NA        NA        NA        NA       NA
#> Track_18       NA       NA       NA        NA        NA        NA       NA
#>          Track_13  Track_15  Track_16  Track_18
#> Track_01 45.00788 18.730370 14.932046 22.179224
#> Track_02 47.19796 17.799527  9.836672 20.569492
#> Track_03 20.29566  7.910351 10.158387  9.928217
#> Track_04 19.90114  4.165947  8.969975  6.396727
#> Track_07 29.44950  2.944420  6.952380  5.167036
#> Track_08 27.32867  4.915326  5.349900  6.863688
#> Track_09 22.41087 41.825031 68.973412 42.969072
#> Track_13       NA 20.579060 39.239001 21.761749
#> Track_15       NA        NA  8.572483  2.784293
#> Track_16       NA        NA        NA 10.795457
#> Track_18       NA        NA        NA        NA
#> 
#> $DTW_distance_metric_p_values
#>          Track_01 Track_02 Track_03 Track_04 Track_07 Track_08 Track_09
#> Track_01       NA        0        0      0.0      0.0      0.0      0.6
#> Track_02       NA       NA        0      0.1      0.2      0.2      0.7
#> Track_03       NA       NA       NA      0.0      0.2      0.3      0.8
#> Track_04       NA       NA       NA       NA      0.0      0.0      0.3
#> Track_07       NA       NA       NA       NA       NA      0.0      0.8
#> Track_08       NA       NA       NA       NA       NA       NA      1.0
#> Track_09       NA       NA       NA       NA       NA       NA       NA
#> Track_13       NA       NA       NA       NA       NA       NA       NA
#> Track_15       NA       NA       NA       NA       NA       NA       NA
#> Track_16       NA       NA       NA       NA       NA       NA       NA
#> Track_18       NA       NA       NA       NA       NA       NA       NA
#>          Track_13 Track_15 Track_16 Track_18
#> Track_01      0.4      0.0      0.0      0.2
#> Track_02      0.1      0.2      0.0      0.1
#> Track_03      0.1      0.0      0.3      0.2
#> Track_04      0.0      0.0      0.0      0.0
#> Track_07      0.1      0.0      0.0      0.0
#> Track_08      0.3      0.1      0.0      0.1
#> Track_09      0.2      0.5      1.0      0.4
#> Track_13       NA      0.4      1.0      0.2
#> Track_15       NA       NA      0.0      0.0
#> Track_16       NA       NA       NA      0.1
#> Track_18       NA       NA       NA       NA
#> 
#> $DTW_metric_p_values_combined
#> [1] 0
#>