Test for differences in velocity means with pairwise comparisons

test_velocity() evaluates differences in velocity means across different tracks using a specified statistical test. It includes options for ANOVA, Kruskal-Wallis test, and Generalized Linear Models (GLM), and checks for assumptions such as normality and homogeneity of variances. For datasets with more than two tracks, it performs pairwise comparisons to identify specific differences between tracks.

Usage

test_velocity(data, trackvel, plot = FALSE, analysis = NULL)

Arguments

data

A 'track' R object, which is a list consisting of two elements:

Trajectories: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.

Footprints: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.

trackvel

A 'track velocity' R object consisting of a list where each element corresponds to a track and contains velocity or relative stride length data.

plot

A logical value indicating whether to plot a boxplot of velocities by track (default is FALSE).

analysis

A character string specifying the type of analysis: "ANOVA", "Kruskal-Wallis", or "GLM". Default is "ANOVA".

Value

A list with the results of the statistical analysis and diagnostic tests:

normality_results: A matrix of test statistics and p-values from the Shapiro-Wilk test for each track, with rows for the test statistic and p-value, and columns for each track.

homogeneity_test: The result of Levene's test, including the p-value for homogeneity of variances.

ANOVA (If analysis is "ANOVA"): A list containing the ANOVA table and Tukey HSD post-hoc test results.

Kruskal_Wallis (If analysis is "Kruskal-Wallis"): A list containing the Kruskal-Wallis test result and Dunn's test post-hoc results.

GLM (If analysis is "GLM"): A summary of the GLM fit and pairwise comparisons.

plot (If plot is TRUE): A boxplot of velocities by track is generated and displayed.

Details

The test_velocity function performs the following operations:

Condition Testing:

Normality: Shapiro-Wilk test for normality on velocity data within each track.

Homogeneity of Variances: Levene's test for equal variances across tracks.

Statistical Analysis:

ANOVA: Compares mean velocities across tracks, assuming normality and homogeneity of variances. Includes Tukey's HSD post-hoc test for pairwise comparisons.

Kruskal-Wallis Test: Non-parametric alternative to ANOVA for comparing median velocities across tracks when assumptions are violated. Includes Dunn's test for pairwise comparisons.

GLM: Generalized Linear Model with a Gaussian family for comparing means if ANOVA assumptions are not met. Pairwise comparisons in the GLM are conducted using estimated marginal means (least-squares means) with the emmeans package, which computes differences between group means while adjusting for multiple comparisons using Tukey’s method.

Plotting:

If plot is TRUE, a boxplot of velocities by track is generated.

Logo

See also

Author

Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477

Examples

# Example 1: Test for Differences in Velocity Means with Pairwise Comparisons in Trajectories
# in MountTom dataset.

# Hip heights for each track in the MountTom dataset
H_mounttom <- c(
  1.380, 1.404, 1.320, 1.736, 1.364, 1.432, 1.508, 1.768, 1.600, 1.848,
  1.532, 1.532, 0.760, 1.532, 1.688, 1.620, 0.636, 1.784, 1.676, 1.872,
  1.648, 1.760, 1.612
)

# Calculate velocities using the default Method "A"
V_mounttom <- velocity_track(MountTom, H = H_mounttom)
#> $Track_01
#> $Track_01$Step_velocities
#> [1] 1.917671 1.699122 1.724574 1.690528 1.646310 1.521394 1.326607 1.634729
#> 
#> $Track_01$Mean_velocity
#> [1] 1.645117
#> 
#> $Track_01$Standard_deviation_velocity
#> [1] 0.17006
#> 
#> $Track_01$Maximum_velocity
#> [1] 1.917671
#> 
#> $Track_01$Minimum_velocity
#> [1] 1.326607
#> 
#> $Track_01$Step_relative_stride
#> [1] 1.553054 1.444507 1.457425 1.440128 1.417452 1.352034 1.245544 1.411473
#> 
#> $Track_01$Mean_relative_stride
#> [1] 1.415202
#> 
#> $Track_01$Standard_deviation_relative_stride
#> [1] 0.08868808
#> 
#> $Track_01$Maximum_relative_stride
#> [1] 1.553054
#> 
#> $Track_01$Minimum_relative_stride
#> [1] 1.245544
#> 
#> 
#> $Track_02
#> $Track_02$Step_velocities
#> [1] 1.446657 1.678316 1.833994 1.901630 2.320863 1.891043 1.662219 1.903494
#> 
#> $Track_02$Mean_velocity
#> [1] 1.829777
#> 
#> $Track_02$Standard_deviation_velocity
#> [1] 0.2544829
#> 
#> $Track_02$Maximum_velocity
#> [1] 2.320863
#> 
#> $Track_02$Minimum_velocity
#> [1] 1.446657
#> 
#> $Track_02$Step_relative_stride
#> [1] 1.305107 1.426506 1.504327 1.537305 1.732091 1.532175 1.418298 1.538208
#> 
#> $Track_02$Mean_relative_stride
#> [1] 1.499252
#> 
#> $Track_02$Standard_deviation_relative_stride
#> [1] 0.1241122
#> 
#> $Track_02$Maximum_relative_stride
#> [1] 1.732091
#> 
#> $Track_02$Minimum_relative_stride
#> [1] 1.305107
#> 
#> 
#> $Track_03
#> $Track_03$Step_velocities
#> [1] 2.434079 2.860861 3.171859 2.291902
#> 
#> $Track_03$Mean_velocity
#> [1] 2.689676
#> 
#> $Track_03$Standard_deviation_velocity
#> [1] 0.4022265
#> 
#> $Track_03$Maximum_velocity
#> [1] 3.171859
#> 
#> $Track_03$Minimum_velocity
#> [1] 2.291902
#> 
#> $Track_03$Step_relative_stride
#> [1] 1.815427 1.999825 2.127300 1.751165
#> 
#> $Track_03$Mean_relative_stride
#> [1] 1.923429
#> 
#> $Track_03$Standard_deviation_relative_stride
#> [1] 0.1719874
#> 
#> $Track_03$Maximum_relative_stride
#> [1] 2.1273
#> 
#> $Track_03$Minimum_relative_stride
#> [1] 1.751165
#> 
#> 
#> $Track_04
#> $Track_04$Step_velocities
#> [1] 1.830533 1.623483 1.409200 1.532089
#> 
#> $Track_04$Mean_velocity
#> [1] 1.598826
#> 
#> $Track_04$Standard_deviation_velocity
#> [1] 0.1776776
#> 
#> $Track_04$Maximum_velocity
#> [1] 1.830533
#> 
#> $Track_04$Minimum_velocity
#> [1] 1.4092
#> 
#> $Track_04$Step_relative_stride
#> [1] 1.410104 1.312308 1.205659 1.267558
#> 
#> $Track_04$Mean_relative_stride
#> [1] 1.298907
#> 
#> $Track_04$Standard_deviation_relative_stride
#> [1] 0.08606683
#> 
#> $Track_04$Maximum_relative_stride
#> [1] 1.410104
#> 
#> $Track_04$Minimum_relative_stride
#> [1] 1.205659
#> 
#> 
#> $Track_05
#> $Track_05$Step_velocities
#> [1] 2.175188
#> 
#> $Track_05$Mean_velocity
#> [1] 2.175188
#> 
#> $Track_05$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_05$Maximum_velocity
#> [1] 2.175188
#> 
#> $Track_05$Minimum_velocity
#> [1] 2.175188
#> 
#> $Track_05$Step_relative_stride
#> [1] 1.680626
#> 
#> $Track_05$Mean_relative_stride
#> [1] 1.680626
#> 
#> $Track_05$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_05$Maximum_relative_stride
#> [1] 1.680626
#> 
#> $Track_05$Minimum_relative_stride
#> [1] 1.680626
#> 
#> 
#> $Track_06
#> $Track_06$Step_velocities
#> [1] 1.868553 1.995720
#> 
#> $Track_06$Mean_velocity
#> [1] 1.932136
#> 
#> $Track_06$Standard_deviation_velocity
#> [1] 0.08992089
#> 
#> $Track_06$Maximum_velocity
#> [1] 1.99572
#> 
#> $Track_06$Minimum_velocity
#> [1] 1.868553
#> 
#> $Track_06$Step_relative_stride
#> [1] 1.512270 1.573083
#> 
#> $Track_06$Mean_relative_stride
#> [1] 1.542676
#> 
#> $Track_06$Standard_deviation_relative_stride
#> [1] 0.04300131
#> 
#> $Track_06$Maximum_relative_stride
#> [1] 1.573083
#> 
#> $Track_06$Minimum_relative_stride
#> [1] 1.51227
#> 
#> 
#> $Track_07
#> $Track_07$Step_velocities
#> [1] 1.1580528 0.8799351 1.1565610 1.5947244 1.6218920 1.3654251
#> 
#> $Track_07$Mean_velocity
#> [1] 1.296098
#> 
#> $Track_07$Standard_deviation_velocity
#> [1] 0.2869998
#> 
#> $Track_07$Maximum_velocity
#> [1] 1.621892
#> 
#> $Track_07$Minimum_velocity
#> [1] 0.8799351
#> 
#> $Track_07$Step_relative_stride
#> [1] 1.1181185 0.9485583 1.1172557 1.3542404 1.3680084 1.2340304
#> 
#> $Track_07$Mean_relative_stride
#> [1] 1.190035
#> 
#> $Track_07$Standard_deviation_relative_stride
#> [1] 0.1608436
#> 
#> $Track_07$Maximum_relative_stride
#> [1] 1.368008
#> 
#> $Track_07$Minimum_relative_stride
#> [1] 0.9485583
#> 
#> 
#> $Track_08
#> $Track_08$Step_velocities
#> [1] 1.840475 1.822117 1.615424 1.429871 1.893798
#> 
#> $Track_08$Mean_velocity
#> [1] 1.720337
#> 
#> $Track_08$Standard_deviation_velocity
#> [1] 0.1938159
#> 
#> $Track_08$Maximum_velocity
#> [1] 1.893798
#> 
#> $Track_08$Minimum_velocity
#> [1] 1.429871
#> 
#> $Track_08$Step_relative_stride
#> [1] 1.406970 1.398550 1.301268 1.209585 1.431239
#> 
#> $Track_08$Mean_relative_stride
#> [1] 1.349522
#> 
#> $Track_08$Standard_deviation_relative_stride
#> [1] 0.09259133
#> 
#> $Track_08$Maximum_relative_stride
#> [1] 1.431239
#> 
#> $Track_08$Minimum_relative_stride
#> [1] 1.209585
#> 
#> 
#> $Track_09
#> $Track_09$Step_velocities
#> [1] 2.836024 2.628135 2.426088 1.824772
#> 
#> $Track_09$Mean_velocity
#> [1] 2.428755
#> 
#> $Track_09$Standard_deviation_velocity
#> [1] 0.4360513
#> 
#> $Track_09$Maximum_velocity
#> [1] 2.836024
#> 
#> $Track_09$Minimum_velocity
#> [1] 1.824772
#> 
#> $Track_09$Step_relative_stride
#> [1] 1.878065 1.794374 1.710448 1.442246
#> 
#> $Track_09$Mean_relative_stride
#> [1] 1.706283
#> 
#> $Track_09$Standard_deviation_relative_stride
#> [1] 0.188858
#> 
#> $Track_09$Maximum_relative_stride
#> [1] 1.878065
#> 
#> $Track_09$Minimum_relative_stride
#> [1] 1.442246
#> 
#> 
#> $Track_10
#> $Track_10$Step_velocities
#> [1] 1.531647
#> 
#> $Track_10$Mean_velocity
#> [1] 1.531647
#> 
#> $Track_10$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_10$Maximum_velocity
#> [1] 1.531647
#> 
#> $Track_10$Minimum_velocity
#> [1] 1.531647
#> 
#> $Track_10$Step_relative_stride
#> [1] 1.243837
#> 
#> $Track_10$Mean_relative_stride
#> [1] 1.243837
#> 
#> $Track_10$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_10$Maximum_relative_stride
#> [1] 1.243837
#> 
#> $Track_10$Minimum_relative_stride
#> [1] 1.243837
#> 
#> 
#> $Track_11
#> $Track_11$Step_velocities
#> [1] 2.076045
#> 
#> $Track_11$Mean_velocity
#> [1] 2.076045
#> 
#> $Track_11$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_11$Maximum_velocity
#> [1] 2.076045
#> 
#> $Track_11$Minimum_velocity
#> [1] 2.076045
#> 
#> $Track_11$Step_relative_stride
#> [1] 1.578469
#> 
#> $Track_11$Mean_relative_stride
#> [1] 1.578469
#> 
#> $Track_11$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_11$Maximum_relative_stride
#> [1] 1.578469
#> 
#> $Track_11$Minimum_relative_stride
#> [1] 1.578469
#> 
#> 
#> $Track_12
#> $Track_12$Step_velocities
#> [1] 2.055415
#> 
#> $Track_12$Mean_velocity
#> [1] 2.055415
#> 
#> $Track_12$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_12$Maximum_velocity
#> [1] 2.055415
#> 
#> $Track_12$Minimum_velocity
#> [1] 2.055415
#> 
#> $Track_12$Step_relative_stride
#> [1] 1.569058
#> 
#> $Track_12$Mean_relative_stride
#> [1] 1.569058
#> 
#> $Track_12$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_12$Maximum_relative_stride
#> [1] 1.569058
#> 
#> $Track_12$Minimum_relative_stride
#> [1] 1.569058
#> 
#> 
#> $Track_13
#> $Track_13$Step_velocities
#> [1] 1.0479855 0.9131596 0.7996623 0.9715874
#> 
#> $Track_13$Mean_velocity
#> [1] 0.9330987
#> 
#> $Track_13$Standard_deviation_velocity
#> [1] 0.1046951
#> 
#> $Track_13$Maximum_velocity
#> [1] 1.047985
#> 
#> $Track_13$Minimum_velocity
#> [1] 0.7996623
#> 
#> $Track_13$Step_relative_stride
#> [1] 1.293046 1.190695 1.099729 1.235746
#> 
#> $Track_13$Mean_relative_stride
#> [1] 1.204804
#> 
#> $Track_13$Standard_deviation_relative_stride
#> [1] 0.08161704
#> 
#> $Track_13$Maximum_relative_stride
#> [1] 1.293046
#> 
#> $Track_13$Minimum_relative_stride
#> [1] 1.099729
#> 
#> 
#> $Track_14
#> $Track_14$Step_velocities
#> [1] 1.642322 1.384645
#> 
#> $Track_14$Mean_velocity
#> [1] 1.513483
#> 
#> $Track_14$Standard_deviation_velocity
#> [1] 0.1822052
#> 
#> $Track_14$Maximum_velocity
#> [1] 1.642322
#> 
#> $Track_14$Minimum_velocity
#> [1] 1.384645
#> 
#> $Track_14$Step_relative_stride
#> [1] 1.371800 1.238534
#> 
#> $Track_14$Mean_relative_stride
#> [1] 1.305167
#> 
#> $Track_14$Standard_deviation_relative_stride
#> [1] 0.09423395
#> 
#> $Track_14$Maximum_relative_stride
#> [1] 1.3718
#> 
#> $Track_14$Minimum_relative_stride
#> [1] 1.238534
#> 
#> 
#> $Track_15
#> $Track_15$Step_velocities
#> [1] 1.481571 1.819262 1.714949 1.727258
#> 
#> $Track_15$Mean_velocity
#> [1] 1.68576
#> 
#> $Track_15$Standard_deviation_velocity
#> [1] 0.1438632
#> 
#> $Track_15$Maximum_velocity
#> [1] 1.819262
#> 
#> $Track_15$Minimum_velocity
#> [1] 1.481571
#> 
#> $Track_15$Step_relative_stride
#> [1] 1.252836 1.416743 1.367525 1.373394
#> 
#> $Track_15$Mean_relative_stride
#> [1] 1.352624
#> 
#> $Track_15$Standard_deviation_relative_stride
#> [1] 0.0700531
#> 
#> $Track_15$Maximum_relative_stride
#> [1] 1.416743
#> 
#> $Track_15$Minimum_relative_stride
#> [1] 1.252836
#> 
#> 
#> $Track_16
#> $Track_16$Step_velocities
#> [1] 1.204506 1.319338 1.530753 1.656302 1.486676 1.558870 1.438280
#> 
#> $Track_16$Mean_velocity
#> [1] 1.456389
#> 
#> $Track_16$Standard_deviation_velocity
#> [1] 0.1524638
#> 
#> $Track_16$Maximum_velocity
#> [1] 1.656302
#> 
#> $Track_16$Minimum_velocity
#> [1] 1.204506
#> 
#> $Track_16$Step_relative_stride
#> [1] 1.120470 1.183262 1.293401 1.355917 1.270970 1.307575 1.246030
#> 
#> $Track_16$Mean_relative_stride
#> [1] 1.253947
#> 
#> $Track_16$Standard_deviation_relative_stride
#> [1] 0.07957735
#> 
#> $Track_16$Maximum_relative_stride
#> [1] 1.355917
#> 
#> $Track_16$Minimum_relative_stride
#> [1] 1.12047
#> 
#> 
#> $Track_17
#> $Track_17$Step_velocities
#> [1] 0.5915414
#> 
#> $Track_17$Mean_velocity
#> [1] 0.5915414
#> 
#> $Track_17$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_17$Maximum_velocity
#> [1] 0.5915414
#> 
#> $Track_17$Minimum_velocity
#> [1] 0.5915414
#> 
#> $Track_17$Step_relative_stride
#> [1] 0.9683894
#> 
#> $Track_17$Mean_relative_stride
#> [1] 0.9683894
#> 
#> $Track_17$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_17$Maximum_relative_stride
#> [1] 0.9683894
#> 
#> $Track_17$Minimum_relative_stride
#> [1] 0.9683894
#> 
#> 
#> $Track_18
#> $Track_18$Step_velocities
#> [1] 1.855606 1.676314 1.873988 1.404891
#> 
#> $Track_18$Mean_velocity
#> [1] 1.7027
#> 
#> $Track_18$Standard_deviation_velocity
#> [1] 0.2176438
#> 
#> $Track_18$Maximum_velocity
#> [1] 1.873988
#> 
#> $Track_18$Minimum_velocity
#> [1] 1.404891
#> 
#> $Track_18$Step_relative_stride
#> [1] 1.410077 1.326836 1.418425 1.193662
#> 
#> $Track_18$Mean_relative_stride
#> [1] 1.33725
#> 
#> $Track_18$Standard_deviation_relative_stride
#> [1] 0.1042735
#> 
#> $Track_18$Maximum_relative_stride
#> [1] 1.418425
#> 
#> $Track_18$Minimum_relative_stride
#> [1] 1.193662
#> 
#> 
#> $Track_19
#> $Track_19$Step_velocities
#> [1] 2.650903
#> 
#> $Track_19$Mean_velocity
#> [1] 2.650903
#> 
#> $Track_19$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_19$Maximum_velocity
#> [1] 2.650903
#> 
#> $Track_19$Minimum_velocity
#> [1] 2.650903
#> 
#> $Track_19$Step_relative_stride
#> [1] 1.778779
#> 
#> $Track_19$Mean_relative_stride
#> [1] 1.778779
#> 
#> $Track_19$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_19$Maximum_relative_stride
#> [1] 1.778779
#> 
#> $Track_19$Minimum_relative_stride
#> [1] 1.778779
#> 
#> 
#> $Track_20
#> $Track_20$Step_velocities
#> [1] 1.631206 1.396409
#> 
#> $Track_20$Mean_velocity
#> [1] 1.513807
#> 
#> $Track_20$Standard_deviation_velocity
#> [1] 0.1660266
#> 
#> $Track_20$Maximum_velocity
#> [1] 1.631206
#> 
#> $Track_20$Minimum_velocity
#> [1] 1.396409
#> 
#> $Track_20$Step_relative_stride
#> [1] 1.286657 1.172319
#> 
#> $Track_20$Mean_relative_stride
#> [1] 1.229488
#> 
#> $Track_20$Standard_deviation_relative_stride
#> [1] 0.08084897
#> 
#> $Track_20$Maximum_relative_stride
#> [1] 1.286657
#> 
#> $Track_20$Minimum_relative_stride
#> [1] 1.172319
#> 
#> 
#> $Track_21
#> $Track_21$Step_velocities
#> [1] 1.719009 1.860508
#> 
#> $Track_21$Mean_velocity
#> [1] 1.789759
#> 
#> $Track_21$Standard_deviation_velocity
#> [1] 0.1000548
#> 
#> $Track_21$Maximum_velocity
#> [1] 1.860508
#> 
#> $Track_21$Minimum_velocity
#> [1] 1.719009
#> 
#> $Track_21$Step_relative_stride
#> [1] 1.379331 1.446237
#> 
#> $Track_21$Mean_relative_stride
#> [1] 1.412784
#> 
#> $Track_21$Standard_deviation_relative_stride
#> [1] 0.04730944
#> 
#> $Track_21$Maximum_relative_stride
#> [1] 1.446237
#> 
#> $Track_21$Minimum_relative_stride
#> [1] 1.379331
#> 
#> 
#> $Track_22
#> $Track_22$Step_velocities
#> [1] 0.917559 1.250737
#> 
#> $Track_22$Mean_velocity
#> [1] 1.084148
#> 
#> $Track_22$Standard_deviation_velocity
#> [1] 0.2355923
#> 
#> $Track_22$Maximum_velocity
#> [1] 1.250737
#> 
#> $Track_22$Minimum_velocity
#> [1] 0.917559
#> 
#> $Track_22$Step_relative_stride
#> [1] 0.9286649 1.1179364
#> 
#> $Track_22$Mean_relative_stride
#> [1] 1.023301
#> 
#> $Track_22$Standard_deviation_relative_stride
#> [1] 0.1338351
#> 
#> $Track_22$Maximum_relative_stride
#> [1] 1.117936
#> 
#> $Track_22$Minimum_relative_stride
#> [1] 0.9286649
#> 
#> 
#> $Track_23
#> $Track_23$Step_velocities
#> [1] 3.753784
#> 
#> $Track_23$Mean_velocity
#> [1] 3.753784
#> 
#> $Track_23$Standard_deviation_velocity
#> [1] NA
#> 
#> $Track_23$Maximum_velocity
#> [1] 3.753784
#> 
#> $Track_23$Minimum_velocity
#> [1] 3.753784
#> 
#> $Track_23$Step_relative_stride
#> [1] 2.216404
#> 
#> $Track_23$Mean_relative_stride
#> [1] 2.216404
#> 
#> $Track_23$Standard_deviation_relative_stride
#> [1] NA
#> 
#> $Track_23$Maximum_relative_stride
#> [1] 2.216404
#> 
#> $Track_23$Minimum_relative_stride
#> [1] 2.216404
#> 
#> 

# Test for Differences in Velocity Means with Pairwise Comparisons
test_velocity(MountTom, V_mounttom)
#> Warning: The following tracks were removed from the analysis due to having 3 or fewer footprints: Track 05, Track 06, Track 10, Track 11, Track 12, Track 14, Track 17, Track 19, Track 20, Track 21, Track 22, Track 23.
#> $normality_results
#>              Track 01  Track 02  Track 03  Track 04  Track 07   Track 08
#> statistic.W 0.9209702 0.9049639 0.8522781 0.9286605 0.9371515 0.81516913
#> p_value     0.4003329 0.2821994 0.2018271 0.5872649 0.6132123 0.08011557
#>              Track 09  Track 13  Track 15  Track 16  Track 18
#> statistic.W 0.8697921 0.9420435 0.8148552 0.9692820 0.8301712
#> p_value     0.2655830 0.6804196 0.1064974 0.8923149 0.1395325
#> 
#> $homogeneity_test
#> Levene's Test for Homogeneity of Variance (center = median)
#>       Df F value  Pr(>F)  
#> group 10  1.8073 0.07947 .
#>       58                  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> $ANOVA
#> $ANOVA$ANOVA
#>             Df Sum Sq Mean Sq F value   Pr(>F)    
#> track       10 10.729  1.0729      19 7.35e-15 ***
#> Residuals   58  3.275  0.0565                     
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> $ANOVA$Tukey
#>   Tukey multiple comparisons of means
#>     95% family-wise confidence level
#> 
#> Fit: aov(formula = vel ~ track, data = M_analysis)
#> 
#> $track
#>                           diff           lwr          upr     p adj
#> Track 02-Track 01  0.194005116 -0.1812699957  0.569280228 0.8128384
#> Track 03-Track 01  0.966158204  0.5221267034  1.410189704 0.0000001
#> Track 04-Track 01 -0.058483846 -0.5025153465  0.385547654 0.9999966
#> Track 07-Track 01 -0.337960523 -0.7391464928  0.063225447 0.1756512
#> Track 08-Track 01  0.105284346 -0.3142859841  0.524854676 0.9987899
#> Track 09-Track 01  0.663995566  0.2199640654  1.108027066 0.0002666
#> Track 13-Track 01 -0.703166313 -1.1471978130 -0.259134812 0.0000926
#> Track 15-Track 01  0.050096645 -0.3939348551  0.494128146 0.9999992
#> Track 16-Track 01 -0.189837124 -0.5766618555  0.196987607 0.8563548
#> Track 18-Track 01 -0.000824843 -0.4448563435  0.443206657 1.0000000
#> Track 03-Track 02  0.772153088  0.3281215872  1.216184588 0.0000136
#> Track 04-Track 02 -0.252488962 -0.6965204627  0.191542538 0.7114764
#> Track 07-Track 02 -0.531965639 -0.9331516090 -0.130779669 0.0018707
#> Track 08-Track 02 -0.088720770 -0.5082911003  0.330849560 0.9997273
#> Track 09-Track 02  0.469990450  0.0259589492  0.914021950 0.0294239
#> Track 13-Track 02 -0.897171429 -1.3412029292 -0.453139928 0.0000004
#> Track 15-Track 02 -0.143908471 -0.5879399713  0.300123030 0.9904498
#> Track 16-Track 02 -0.383842241 -0.7706669717  0.002982491 0.0534952
#> Track 18-Track 02 -0.194829959 -0.6388614597  0.249201541 0.9236910
#> Track 04-Track 03 -1.024642050 -1.5281264461 -0.521157654 0.0000003
#> Track 07-Track 03 -1.304118727 -1.7702547008 -0.837982753 0.0000000
#> Track 08-Track 03 -0.860873858 -1.3429234671 -0.378824249 0.0000076
#> Track 09-Track 03 -0.302162638 -0.8056470343  0.201321758 0.6425087
#> Track 13-Track 03 -1.669324516 -2.1728089127 -1.165840120 0.0000000
#> Track 15-Track 03 -0.916061559 -1.4195459547 -0.412577162 0.0000050
#> Track 16-Track 03 -1.155995328 -1.6098300300 -0.702160626 0.0000000
#> Track 18-Track 03 -0.966983047 -1.4704674431 -0.463498651 0.0000014
#> Track 07-Track 04 -0.279476677 -0.7456126509  0.186659297 0.6438142
#> Track 08-Track 04  0.163768192 -0.3182814172  0.645817801 0.9864022
#> Track 09-Track 04  0.722479412  0.2189950157  1.225963808 0.0005416
#> Track 13-Track 04 -0.644682467 -1.1481668627 -0.141198070 0.0030869
#> Track 15-Track 04  0.108580491 -0.3949039048  0.612064888 0.9996753
#> Track 16-Track 04 -0.131353278 -0.5851879800  0.322481423 0.9960595
#> Track 18-Track 04  0.057659003 -0.4458253932  0.561143399 0.9999991
#> Track 08-Track 07  0.443244869  0.0003475505  0.886142188 0.0496557
#> Track 09-Track 07  1.001956089  0.5358201149  1.468092063 0.0000001
#> Track 13-Track 07 -0.365205790 -0.8313417635  0.100930184 0.2604157
#> Track 15-Track 07  0.388057168 -0.0780788056  0.854193142 0.1882672
#> Track 16-Track 07  0.148123399 -0.2638864865  0.560133284 0.9794893
#> Track 18-Track 07  0.337135680 -0.1290002940  0.803271654 0.3706410
#> Track 09-Track 08  0.558711220  0.0766616105  1.040760829 0.0110562
#> Track 13-Track 08 -0.808450659 -1.2905002679 -0.326401049 0.0000294
#> Track 15-Track 08 -0.055187701 -0.5372373100  0.426861909 0.9999991
#> Track 16-Track 08 -0.295121470 -0.7250531220  0.134810181 0.4479344
#> Track 18-Track 08 -0.106109189 -0.5881587984  0.375940420 0.9996108
#> Track 13-Track 09 -1.367161878 -1.8706462746 -0.863677482 0.0000000
#> Track 15-Track 09 -0.613898920 -1.1173833167 -0.110414524 0.0059373
#> Track 16-Track 09 -0.853832690 -1.3076673919 -0.399997988 0.0000023
#> Track 18-Track 09 -0.664820409 -1.1683048051 -0.161336013 0.0019895
#> Track 15-Track 13  0.753262958  0.2497785617  1.256747354 0.0002643
#> Track 16-Track 13  0.513329188  0.0594944865  0.967163890 0.0146246
#> Track 18-Track 13  0.702341470  0.1988570733  1.205825866 0.0008589
#> Track 16-Track 15 -0.239933770 -0.6937684714  0.213900932 0.7915111
#> Track 18-Track 15 -0.050921488 -0.5544058846  0.452562908 0.9999997
#> Track 18-Track 16  0.189012281 -0.2648224205  0.642846983 0.9447031
#> 
#> 
#> 

# Example 2: Test for Differences in Velocity Means with Pairwise Comparisons in Trajectories
# in PaluxyRiver dataset.

# Hip heights for each track in the PaluxyRiver dataset
H_paluxyriver <- c(3.472, 2.200)

# Specify different methods for different tracks
Method_paluxyriver <- c("A", "B")

# Calculate velocities using specified methods
V_paluxyriver <- velocity_track(PaluxyRiver, H = H_paluxyriver, method = Method_paluxyriver)
#> $Track_1
#> $Track_1$Step_velocities
#>  [1] 0.2884087 0.2667234 0.2246319 0.2038125 0.2297207 0.2673225 0.2495612
#>  [8] 0.2303131 0.2352243 0.1496251 0.1616271 0.2372131 0.2426553 0.2254857
#> [15] 0.2041018 0.2211659 0.2609607 0.2515620 0.3030204 0.2519311 0.2015630
#> [22] 0.1627231 0.2052861 0.2788176 0.2525451 0.2455988 0.2475302 0.3025814
#> 
#> $Track_1$Mean_velocity
#> [1] 0.2357754
#> 
#> $Track_1$Standard_deviation_velocity
#> [1] 0.03865919
#> 
#> $Track_1$Maximum_velocity
#> [1] 0.3030204
#> 
#> $Track_1$Minimum_velocity
#> [1] 0.1496251
#> 
#> $Track_1$Step_relative_stride
#>  [1] 0.3789154 0.3615884 0.3262496 0.3077913 0.3306554 0.3620745 0.3474711
#>  [8] 0.3311657 0.3353764 0.2557886 0.2678840 0.3370715 0.3416810 0.3269916
#> [15] 0.3080528 0.3232259 0.3568899 0.3491365 0.3902964 0.3494432 0.3057525
#> [22] 0.2689704 0.3091219 0.3713188 0.3499529 0.3441569 0.3457750 0.3899578
#> 
#> $Track_1$Mean_relative_stride
#> [1] 0.3347413
#> 
#> $Track_1$Standard_deviation_relative_stride
#> [1] 0.03373255
#> 
#> $Track_1$Maximum_relative_stride
#> [1] 0.3902964
#> 
#> $Track_1$Minimum_relative_stride
#> [1] 0.2557886
#> 
#> 
#> $Track_2
#> $Track_2$Step_velocities
#>  [1] 0.3812738 0.3999907 0.4803527 0.4653572 0.4868985 0.4856273 0.4333103
#>  [8] 0.3584286 0.3452030 0.3788715 0.3803733 0.3428550 0.5076662 0.7056220
#> [15] 0.4931734 0.3356635 0.3580868 0.3716003 0.4622697 0.5033969 0.5608097
#> [22] 0.6078206 0.6424547
#> 
#> $Track_2$Mean_velocity
#> [1] 0.4559611
#> 
#> $Track_2$Standard_deviation_velocity
#> [1] 0.1015011
#> 
#> $Track_2$Maximum_velocity
#> [1] 0.705622
#> 
#> $Track_2$Minimum_velocity
#> [1] 0.3356635
#> 
#> $Track_2$Step_relative_stride
#>  [1] 0.5453924 0.5612701 0.6263007 0.6145187 0.6313974 0.6304098 0.5888162
#>  [8] 0.5255821 0.5138819 0.5433321 0.5446207 0.5117860 0.6473884 0.7884788
#> [15] 0.6362574 0.5053306 0.5252820 0.5370638 0.6120740 0.6441228 0.6871563
#> [22] 0.7210902 0.7454200
#> 
#> $Track_2$Mean_relative_stride
#> [1] 0.6037814
#> 
#> $Track_2$Standard_deviation_relative_stride
#> [1] 0.0789691
#> 
#> $Track_2$Maximum_relative_stride
#> [1] 0.7884788
#> 
#> $Track_2$Minimum_relative_stride
#> [1] 0.5053306
#> 
#> 

# Test for Differences in Velocity Means with Pairwise Comparisons
test_velocity(PaluxyRiver, V_paluxyriver)
#> Warning: One or more tracks do not follow a normal distribution (p-value <= 0.05). Assumptions for ANOVA are not met. Consider using 'Kruskal-Wallis' or 'GLM'.
#> Warning: Homogeneity of variances assumption is violated (Levene's test p-value <= 0.05). Assumptions for ANOVA are not met. Consider using 'Kruskal-Wallis' or 'GLM'.
#> $normality_results
#>               Track 1    Track 2
#> statistic.W 0.9586486 0.91134871
#> p_value     0.3043396 0.03773134
#> 
#> $homogeneity_test
#> Levene's Test for Homogeneity of Variance (center = median)
#>       Df F value    Pr(>F)    
#> group  1  18.698 7.121e-05 ***
#>       51                      
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> $ANOVA
#> $ANOVA$ANOVA
#>             Df Sum Sq Mean Sq F value   Pr(>F)    
#> track        1 0.6687  0.6687   111.9 1.82e-14 ***
#> Residuals   51 0.3046  0.0060                     
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> $ANOVA$Tukey
#>   Tukey multiple comparisons of means
#>     95% family-wise confidence level
#> 
#> Fit: aov(formula = vel ~ track, data = M_analysis)
#> 
#> $track
#>                      diff       lwr     upr p adj
#> Track 2-Track 1 0.2256526 0.1828352 0.26847     0
#> 
#> 
#>